### 2024-03-21 #6522 @editorialbot
[PRE REVIEW]: ASGarD: Adaptive Sparse Grid Discretization
**Submitting author:** @quantumsteve (<a href="http://orcid.org/0000-0002-2018-7904">Steven Hahn</a>)
**Repository:** https://github.com/project-asgard/asgard
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.0
