### 2024-03-21 #6522 @danielskatz
👋 @quantumsteve - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. (In this case, if a person doesn't have an ORCID, just remove that line from the paper header.) Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
