### 2024-03-22 #6522 @danielskatz
@quantumsteve - your footnote 1 is a bit odd, as it has nothing to do with the text where it appears.  This would be more appropriate as part of the acknowledgements, in my opinion.
