### 2024-04-16 590488d
remove old debug prints with no context
