### 2024-04-16 ba4d4e2
added some sanity to the current kronmult reporting
