+++ src/asgard_kronmult_matrix.cpp
@@ -1396,8 +1396,8 @@ void set_specific_mode(PDE<precision> const &pde,
       gflops += mat.gvals_[mat.patterns_per_dim * (t * num_dimensions + d)].size();
   gflops *= 2; // matrix vector product uses multiply-add 2 flops per entry
 
-  std::cout << "Kronmult using global algorithm:\n";
-  std::cout << "    work: " << static_cast<double>(gflops) * 1.E-9 << " Gflops\n";
+  std::cout << "  kronmult using global algorithm\n";
+  std::cout << "  -- work: " << static_cast<double>(gflops) * 1.E-9 << " Gflops\n";
   mat.flops_[imex_indx] = std::max(gflops, int64_t{1}); // cannot be zero
 
   int64_t num_ints = 0;
@@ -1416,12 +1416,15 @@ void set_specific_mode(PDE<precision> const &pde,
       num_fps += mat.gvals_[mat.patterns_per_dim * t * num_dimensions + d].size();
 
   num_fps += 2 * mat.num_active_;
-  std::cout << "  -- memory usage:\n";
+  std::cout << "  -- memory usage:";
   int64_t total = get_MB<precision>(num_fps) + get_MB<int>(num_ints);
   if (total > 1024)
-    std::cout << "    CPU: " << (1 + total / 1024) << "GB\n";
+    std::cout << "  CPU: " << (1 + total / 1024) << "GB";
   else
-    std::cout << "    CPU: " << total << "MB\n";
+    std::cout << "  CPU: " << total << "MB";
+#ifndef ASGARD_USE_CUDA
+  std::cout << '\n';
+#endif
 }
 
 #ifdef ASGARD_USE_CUDA
@@ -1446,9 +1449,9 @@ void global_kron_matrix<precision>::
 
   int64_t total = gpu_global[imex_indx].memory() / (1024 * 1024);
   if (total > 1024)
-    std::cout << "    GPU: " << (1 + total / 1024) << "GB\n";
+    std::cout << "  GPU: " << (1 + total / 1024) << "GB\n";
   else
-    std::cout << "    GPU: " << total << "MB\n";
+    std::cout << "  GPU: " << total << "MB\n";
 }
 #endif
 
