### 2024-04-18 0296bda
added a set union method
