### 2024-05-14 #6522 @jedbrown
Sorry about my extended delay and thanks for your patience. This looks very interesting. Some quick items:
1. In the repo, you have `paper.md` and `paper/paper.md`. I think the former should be deleted to avoid confusion.
2. There is a `[^1]`  (markdown footnote anchor) appearing in the first sentence. I think it was meant to be deleted in the commit that deleted the footnote content.
3. I'm about to invite a few reviewers, but you're also welcome to suggest. If you do so, please don't tag them here yet.
