### 2024-05-20 #6522 @joglekara
> 👋 @dannys4 @joglekara @MikeHeiber 👋 Hi, would any of you be available to review this submission for JOSS?

yes happy to review if you still need
