### 2024-06-04 #6766 @gnorman7
I was able to set up and run everything without any hiccups, although I did not verify the claims of the MATLAB share engine or plotting. The examples nicely showcase both basic examples and more complicated core functionality (adaptivity). I did not look at any cases beyond slight changes to the examples.

My only negative notes are about referring to other packages.

I'm unsure of the state of the field, or if there are other packages that do Sparse Grid with DG (even non-adaptive). That being said, there are few mentions to other commonly-used packages (although the authors mention other techniques, such as Monte Carlo and the "combination technique"). I do think there could be more referring to similar packages, even if they are not in the same niche.

I think it might be worth mentioning the "separate MATLAB implementation https://github.com/project-asgard/DG-SparseGrid" more prominently. This package seems quite closely related.
