+++ src/asgard_kronmult_matrix.cpp
@@ -1172,10 +1172,20 @@ void split_pattern(std::vector<int> const &pntr, std::vector<int> const &indx,
 template<typename precision>
 bool check_identity_term(PDE<precision> const &pde, int term_id, int dim)
 {
+  // Check that the volumne jacobian in this dimension is not identity,
+  if (pde.get_dimensions()[dim].volume_jacobian_dV != nullptr)
+    return false;
+  // Now check g_func, the lhs_func, and local surface jacobian are identity.
+  // TODO: There is an edge case where the mass matrices with the same volume
+  // jacobians can cancel out, but this requires us to know that both the
+  // dimensions' and the local terms' volume jacobian are equal.
+  // In the edge case, identity will be multiplied instead of ignored
+  // resulting in extra work but correct output.
   for (auto const &pt : pde.get_terms()[term_id][dim].get_partial_terms())
     if (pt.coeff_type != coefficient_type::mass or
         pt.g_func != nullptr or
-        pt.lhs_mass_func != nullptr)
+        pt.lhs_mass_func != nullptr or
+        pt.dv_func != nullptr)
       return false;
   return true;
 }
