### 2024-06-10 03c0717
adding block global algorithm
