+++ src/time_advance.cpp
@@ -439,9 +439,12 @@ implicit_advance(PDE<P> &pde, matrix_list<P> &operator_matrices,
     fk::vector<P> fx(x);
     // TODO: do something better to save gmres output to pde
 #ifdef KRON_MODE_GLOBAL
+#ifdef KRON_MODE_GLOBAL_BLOCK
+#else
     pde.gmres_outputs[0] = solver::simple_gmres_euler<P, resource::host>(
         pde.get_dt(), matrix_entry::regular, operator_matrices.kglobal,
         fx, x, restart, max_iter, tolerance);
+#endif
 #else
     pde.gmres_outputs[0] = solver::simple_gmres_euler(
         pde.get_dt(), operator_matrices[matrix_entry::regular],
@@ -879,9 +882,9 @@ imex_advance(PDE<P> &pde, matrix_list<P> &operator_matrices,
     if (solver == solve_opts::gmres)
     {
 #ifdef KRON_MODE_GLOBAL
-      pde.gmres_outputs[0] = solver::simple_gmres_euler(
-          pde.get_dt(), matrix_entry::imex_implicit, operator_matrices.kglobal,
-          f_1, f, restart, max_iter, tolerance);
+    pde.gmres_outputs[0] = solver::simple_gmres_euler(
+        pde.get_dt(), matrix_entry::imex_implicit, operator_matrices.kglobal,
+        f_1, f, restart, max_iter, tolerance);
 #else
       pde.gmres_outputs[0] = solver::simple_gmres_euler(
           pde.get_dt(), operator_matrices[matrix_entry::imex_implicit],
@@ -997,9 +1000,9 @@ imex_advance(PDE<P> &pde, matrix_list<P> &operator_matrices,
     if (solver == solve_opts::gmres)
     {
 #ifdef KRON_MODE_GLOBAL
-      pde.gmres_outputs[1] = solver::simple_gmres_euler(
-          P{0.5} * pde.get_dt(), matrix_entry::imex_implicit, operator_matrices.kglobal,
-          f_2, f, restart, max_iter, tolerance);
+    pde.gmres_outputs[1] = solver::simple_gmres_euler(
+        P{0.5} * pde.get_dt(), matrix_entry::imex_implicit, operator_matrices.kglobal,
+        f_2, f, restart, max_iter, tolerance);
 #else
       pde.gmres_outputs[1] = solver::simple_gmres_euler(
           P{0.5} * pde.get_dt(), operator_matrices[matrix_entry::imex_implicit],
