+++ src/build_info.hpp.in
@@ -23,3 +23,4 @@
 #define ASGARD_NUM_QUADRATURE @ASGARD_NUM_QUADRATURE@
 
 #cmakedefine KRON_MODE_GLOBAL
+#cmakedefine KRON_MODE_GLOBAL_BLOCK

[…]
