### 2024-06-10 6aa00b7
handle case when non-identity and no flux term
