### 2024-06-11 0f1efee
fix the missing penalty term
