+++ src/asgard_kronmult_matrix.hpp
@@ -1290,21 +1290,60 @@ void set_specific_mode(
       options const &program_options, imex_flag const imex,
       block_global_kron_matrix<precision> &mat);
 
+
 template<typename precision>
 class block_global_kron_matrix {
 public:
-  block_global_kron_matrix() : conn_volumes_(1), conn_full_(1) {}
+  block_global_kron_matrix() : num_dimensions_(0), conn_volumes_(1), conn_full_(1), workspace_(nullptr) {}
 
-  block_global_kron_matrix(int num_dimensions, int blockn, int64_t block_size,
+  block_global_kron_matrix(int64_t num_active, int64_t num_padded,
+                           int num_dimensions, int blockn, int64_t block_size,
                            vector2d<int> &&ilist, dimension_sort &&dsort,
-                           std::vector<kronmult::permutes> &&perms, std::vector<int> &&flux_dir_,
+                           std::vector<kronmult::permutes> &&perms, std::vector<int> &&flux_dir,
                            connect_1d &&conn_volumes, connect_1d &&conn_full,
-                           kronmult::block_global_workspace<precision> &workspace)
-  : num_dimensions_(num_dimensions), blockn_(blockn), block_size_(block_size),
+                           kronmult::block_global_workspace<precision> *workspace)
+  : num_active_(num_active), num_padded_(num_padded),
+    num_dimensions_(num_dimensions), blockn_(blockn), block_size_(block_size),
     ilist_(std::move(ilist)), dsort_(std::move(dsort)), perms_(std::move(perms)),
-    flux_dir_(std::move(flux_dir_)), conn_volumes_(std::move(conn_volumes)),
-    conn_full_(std::move(conn_full)), workspace_(workspace)
-  {}
+    flux_dir_(std::move(flux_dir)), conn_volumes_(std::move(conn_volumes)),
+    conn_full_(std::move(conn_full)), gvals_(flux_dir_.size() * num_dimensions_),
+    workspace_(workspace)
+  {
+    std::cout << " in contructor gvals_.size() = " << gvals_.size() << "\n";
+  }
+
+  template<resource rec>
+  void apply(matrix_entry etype, precision alpha, precision const *x, precision beta, precision *y) const;
+
+  operator bool () const { return (num_dimensions_ > 0); }
+
+  bool specific_is_set(matrix_entry etype)
+  {
+    std::vector<int> const &terms = term_groups_[flag2int(etype)];
+    if (terms.empty())
+      return true; // nothing to set, so we're OK
+
+    for (int d = 0; d < num_dimensions_; d++)
+      if (not gvals_[terms.front() * num_dimensions_ + d].empty())
+        return true;
+    return false;
+  }
+
+    //! \brief Allows overwriting of the loaded coefficients.
+  template<resource rec>
+  auto const &get_diagonal_preconditioner() const
+  {
+    static_assert(rec == resource::host, "GPU not enabled");
+    return pre_con_;
+  }
+
+  //! \brief Return the number of flops for the current matrix type
+  int64_t flops(matrix_entry etype) const
+  {
+    // return flops_[flag2int(etype)];
+    ignore(etype);
+    return 1000000;
+  }
 
   // made friends for two reasons
   // 1. Keeps the matrix API free from references to pde, which will allow an easier
@@ -1315,13 +1354,6 @@ public:
       adapt::distributed_grid<precision> const &dis_grid,
       options const &program_options, imex_flag const imex,
       block_global_kron_matrix<precision> &mat);
-
-//   friend void update_matrix_coefficients<precision>(
[…]
