+++ src/asgard_kronmult_matrix.cpp
@@ -1600,12 +1600,64 @@ void global_kron_matrix<precision>::apply(
 }
 #endif
 
+#ifdef KRON_MODE_GLOBAL_BLOCK
+
+template<typename precision>
+template<resource rec>
+void block_global_kron_matrix<precision>::apply(
+    matrix_entry etype, precision alpha, precision const *x,
+    precision beta, precision *y) const
+{
+  int const imex = flag2int(etype);
+
+  std::vector<int> const &used_terms = term_groups_[imex];
+
+  if (beta == 0)
+    kronmult::set_buffer_to_zero<rec>(num_active_, y);
+  else
+    lib_dispatch::scal<resource::host>(num_active_, beta, y, 1);
+
+  if (used_terms.size() == 0)
+    return;
+
+  std::copy_n(x, num_active_, workspace_->x.begin());
+  std::fill_n(workspace_->y.begin(), num_active_, precision{0});
+
+//   kronmult::global_cpu(num_dimensions_, perms_, gpntr_, gindx_, gdiag_, gvals_,
+//                        used_terms, get_buffer<workspace::pad_x>(),
+//                        get_buffer<workspace::pad_y>(),
+//                        get_buffer<workspace::stage1>(),
+//                        get_buffer<workspace::stage2>());
+
+  kronmult::global_cpu(num_dimensions_, blockn_, block_size_, ilist_, dsort_,
+                       perms_, flux_dir_, conn_volumes_, conn_full_,
+                       gvals_, used_terms, workspace_->x.data(), workspace_->y.data(),
+                       *workspace_);
+
+// template<typename precision>
+// void global_cpu(int num_dimensions, int n, int64_t block_size,
+//                 vector2d<int> const &ilist, dimension_sort const &dsort,
+//                 std::vector<permutes> const &perms,
+//                 std::vector<bool> const &has_flux,
+//                 connect_1d const &conn_volumes, connect_1d const &conn_full,
+//                 std::vector<std::vector<precision>> const &gvals,
+//                 std::vector<int> const &terms,
+//                 precision const x[], precision y[],
+//                 block_global_workspace<precision> &workspace);
+
+  precision *py = workspace_->y.data();
+#pragma omp parallel for
+  for (int64_t i = 0; i < num_active_; i++)
+    y[i] += alpha * py[i];
+
+}
+
 template<typename precision>
 block_global_kron_matrix<precision>
 make_block_global_kron_matrix(PDE<precision> const &pde,
                               adapt::distributed_grid<precision> const &dis_grid,
                               options const &program_options,
-                              kronmult::block_global_workspace<precision> &workspace)
+                              kronmult::block_global_workspace<precision> *workspace)
 {
   auto const &grid         = dis_grid.get_subgrid(get_rank());
   int const *const asg_idx = dis_grid.get_table().get_active_table().data();
@@ -1660,7 +1712,20 @@ make_block_global_kron_matrix(PDE<precision> const &pde,
     permutations.back().remap_directions(active_dirs);
   }
 
+  int64_t num_active = cells.num_strips() * block_size;
+  workspace->x.resize(num_active);
+  workspace->y.resize(num_active);
+  workspace->w1.resize(num_active);
+  workspace->w2.resize(num_active);
+
+  std::cout << " num_terms = " << num_terms
+            << " flux_dir.size() = " << flux_dir.size()
+            << " num_active = " << num_active
+            << " num_cells = " << num_cells
+            << "\n";
+
   return block_global_kron_matrix<precision>(
+      num_cells * block_size, cells.num_strips() * block_size,
[…]
