### 2024-06-12 #6766 @quantumsteve
@gnorman7 Thank you for taking the time to review ASGarD. We are glad to hear you successfully set up and ran everything without hiccups.

In response to your notes, we added a state of the field section to the paper and reference both GalerkinSparseGrids.jl and AdaM-DG. Both combine the discontinuous Galerkin method and sparse grids and AdaM-DG supports adaptivity and shared-memory parallelism. They currently are missing ASGarD's support for GPU accelerators and distributed memory parallelism.
