### 2024-06-12 ecb2b8b
Update hdf5 patch version to support Apple silicon
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
