### 2024-06-12 #6766 @quantumsteve
The MATLAB code DG-Sparse-Grid hasn't been updated in several years and is no longer maintained or kept up to date with ASGarD. I added a note with this information at the top of the README.
