+++ src/device/asgard_glkronmult_bcpu.cpp
@@ -12,6 +12,13 @@ namespace asgard::kronmult
 template<typename precision, int num_dimensions, int dim, int n>
 void gbkron_mult_add(precision const A[], precision const x[], precision y[])
 {
+  if constexpr (n == 1) // dimension does not matter here
+  {
+    y[0] += A[0] * x[0];
+    return;
+  }
+
+  static_assert(num_dimensions >= 1 and num_dimensions <= 6);
   if constexpr (num_dimensions == 1)
   {
     ASGARD_PRAGMA_OMP_SIMD(collapse(2))
@@ -25,21 +32,133 @@ void gbkron_mult_add(precision const A[], precision const x[], precision y[])
     {
       ASGARD_PRAGMA_OMP_SIMD(collapse(3))
       for (int j = 0; j < n; j++)
-        for (int s = 0; s < n; s++)
-          for (int k = 0; k < n; k++)
-            y[s * n + k] += A[j * n + s] * x[n * j + k];
-//       std::cout << "y = " << y[0] << "  " << y[1] << "  " << y[2] << "  " << y[3] << "\n";
-//       std::cout << "x = " << x[0] << "  " << x[1] << "  " << x[2] << "  " << x[3] << "\n";
-//       std::cout << "A = " << A[0] << "  " << A[1] << "  " << A[2] << "  " << A[3] << "\n";
+        for (int k = 0; k < n; k++)
+          for (int s = 0; s < n; s++)
+            y[s + k * n] += A[j * n + s] * x[j + k * n];
     }
     else
     {
       ASGARD_PRAGMA_OMP_SIMD(collapse(3))
-      for (int k = 0; k < n; k++)
+      for (int j = 0; j < n; j++)
+        for (int s = 0; s < n; s++)
+          for (int k = 0; k < n; k++)
+            y[k + s * n] += A[j * n + s] * x[k + j * n];
+    }
+  }
+  else if constexpr (num_dimensions == 3)
+  {
+    if constexpr (dim == 2)
+    {
+      ASGARD_PRAGMA_OMP_SIMD(collapse(4))
+      for (int l = 0; l < n; l++)
+        for (int k = 0; k < n; k++)
+          for (int j = 0; j < n; j++)
+            for (int s = 0; s < n; s++)
+              y[l * n * n + k * n + s] += A[j * n + s] * x[l * n * n + n * k + j];
+    }
+    else if constexpr (dim == 1)
+    {
+      ASGARD_PRAGMA_OMP_SIMD(collapse(4))
+      for (int l = 0; l < n; l++)
         for (int j = 0; j < n; j++)
           for (int s = 0; s < n; s++)
-            y[k * n + s] += A[j * n + s] * x[k * n + j];
+            for (int k = 0; k < n; k++)
+              y[l * n * n + s * n + k] += x[l * n * n + j * n + k] * A[j * n + s];
     }
+    else // dim == 0
+    {
+      ASGARD_PRAGMA_OMP_SIMD(collapse(4))
+      for (int j = 0; j < n; j++)
+        for (int s = 0; s < n; s++)
+          for (int l = 0; l < n; l++)
+            for (int k = 0; k < n; k++)
+              y[s * n * n + l * n + k] += x[n * n * j + n * l + k] * A[j * n + s];
+    }
+  }
+  else if constexpr (num_dimensions == 4)
+  {
+    if constexpr (dim == 3)
+    {
+      ASGARD_PRAGMA_OMP_SIMD(collapse(5))
+      for (int p = 0; p < n; p++)
+        for (int l = 0; l < n; l++)
+          for (int k = 0; k < n; k++)
+            for (int j = 0; j < n; j++)
[…]
