+++ src/time_advance.cpp
@@ -238,6 +238,10 @@ explicit_advance(PDE<P> const &pde, matrix_list<P> &operator_matrices,
   fm::axpy(rk_1, x, rk_scale_1);
 
   // -- RK step 2
+//   std::cout << " ---------------------------------- \n";
+//   for(auto s : x)
+//     std::cout << s << "\n";
+//   std::cout << " ---------------------------------- \n";
   {
     tools::time_event performance(
         "kronmult", operator_matrices.flops(matrix_entry::regular));
