+++ src/device/asgard_glkronmult_bcpu.cpp
@@ -85,7 +85,7 @@ void gbkron_mult_add(precision const A[], precision const x[], precision y[])
           for (int k = 0; k < n; k++)
             for (int j = 0; j < n; j++)
               for (int s = 0; s < n; s++)
-                y[p * n * n *n + l * n * n + k * n + s] += A[j * n + s] * x[p * n * n * n + l * n * n + k * n + j];
+                y[p * ipow<n, 3>() + l * n * n + k * n + s] += A[j * n + s] * x[p * ipow<n, 3>() + l * n * n + k * n + j];
     }
     else if constexpr (dim == 2)
     {
@@ -95,7 +95,7 @@ void gbkron_mult_add(precision const A[], precision const x[], precision y[])
           for (int j = 0; j < n; j++)
             for (int s = 0; s < n; s++)
               for (int k = 0; k < n; k++)
-                y[p * n * n * n + l * n * n + s * n + k] += x[p * n * n * n + l * n * n + j * n + k] * A[j * n + s];
+                y[p * ipow<n, 3>() + l * n * n + s * n + k] += x[p * ipow<n, 3>() + l * n * n + j * n + k] * A[j * n + s];
     }
     else if constexpr (dim == 1)
     {
@@ -105,7 +105,7 @@ void gbkron_mult_add(precision const A[], precision const x[], precision y[])
           for (int s = 0; s < n; s++)
             for (int l = 0; l < n; l++)
               for (int k = 0; k < n; k++)
-                y[p * n * n * n + s * n * n + l * n + k] += x[p * n * n * n + j * n * n + l * n + k] * A[j * n + s];
+                y[p * ipow<n, 3>() + s * n * n + l * n + k] += x[p * ipow<n, 3>() + j * n * n + l * n + k] * A[j * n + s];
     }
     else // dim == 0
     {
@@ -119,47 +119,154 @@ void gbkron_mult_add(precision const A[], precision const x[], precision y[])
                     x[n * n * n * j + n * n * p + n * l + k] *
                     A[j * n + s];
     }
-          // P W[n][n][n][n] = {{{{{0}}}}}, Y[n][n][n][n] = {{{{{0}}}}};
-          // P const *A = &vA[t][n * n * (ix[0] * num_1d_blocks + iy[0])];
-          // ASGARD_PRAGMA_OMP_SIMD(collapse(5))
+  }
+  else if constexpr (num_dimensions == 5)
+  {
+    if constexpr (dim == 4)
+    {
+      ASGARD_PRAGMA_OMP_SIMD(collapse(6))
+      for (int v = 0; v < n; v++)
+        for (int p = 0; p < n; p++)
+          for (int l = 0; l < n; l++)
+            for (int k = 0; k < n; k++)
+              for (int j = 0; j < n; j++)
+                for (int s = 0; s < n; s++)
+                  y[v * ipow<n, 4>() + p * ipow<n, 3>() + l * n * n + k * n + s]
+                    += A[j * n + s] * x[v * ipow<n, 4>() + p * ipow<n, 3>() + l * n * n + k * n + j];
+    }
+    else if constexpr (dim == 3)
+    {
+      ASGARD_PRAGMA_OMP_SIMD(collapse(6))
+      for (int v = 0; v < n; v++)
+        for (int p = 0; p < n; p++)
+          for (int l = 0; l < n; l++)
+            for (int j = 0; j < n; j++)
+              for (int s = 0; s < n; s++)
+                for (int k = 0; k < n; k++)
+                  y[v * ipow<n, 4>() + p * ipow<n, 3>() + l * n * n + s * n + k]
+                    += x[v * ipow<n, 4>() + p * ipow<n, 3>() + l * n * n + j * n + k] * A[j * n + s];
+    }
+    else if constexpr (dim == 2)
+    {
[…]
