+++ src/asgard_bicgstab.cpp
@@ -56,7 +56,7 @@ public:
 template<typename P>
 bicgstab_info<P>
 bicgstab(fk::matrix<P> const &A, fk::vector<P> &x, fk::vector<P> const &b,
-         fk::matrix<P> const &M, int const restart, int const max_iter,
+         fk::matrix<P> const &M, int const max_iter,
          P const tolerance)
 {
   auto dense_matrix_wrapper =
@@ -66,11 +66,11 @@ bicgstab(fk::matrix<P> const &A, fk::vector<P> &x, fk::vector<P> const &b,
       };
   if (M.size() > 0)
     return bicgstab(dense_matrix_wrapper, fk::vector<P, mem_type::view>(x),
-                    b, dense_preconditioner(M), restart, max_iter,
+                    b, dense_preconditioner(M), max_iter,
                     tolerance);
   else
     return bicgstab(dense_matrix_wrapper, fk::vector<P, mem_type::view>(x),
-                    b, no_op_preconditioner<P>(), restart, max_iter,
+                    b, no_op_preconditioner<P>(), max_iter,
                     tolerance);
 }
 
@@ -98,7 +98,7 @@ bicgstab_euler(const P dt, matrix_entry mentry,
                global_kron_matrix<P> const &mat,
                fk::vector<P, mem_type::owner, resrc> &x,
                fk::vector<P, mem_type::owner, resrc> const &b,
-               int const restart, int const max_iter, P const tolerance)
+               int const max_iter, P const tolerance)
 {
   auto const &pc = mat.template get_diagonal_preconditioner<resrc>();
 
@@ -114,7 +114,7 @@ bicgstab_euler(const P dt, matrix_entry mentry,
         tools::time_event performance("kronmult - preconditioner", pc.size());
         apply_diagonal_precond(pc, dt, x_in);
       },
-      restart, max_iter, tolerance);
+      max_iter, tolerance);
 }
 #else
 template<typename P, resource resrc>
@@ -122,7 +122,7 @@ bicgstab_info<P>
 bicgstab_euler(const P dt, kronmult_matrix<P> const &mat,
                fk::vector<P, mem_type::owner, resrc> &x,
                fk::vector<P, mem_type::owner, resrc> const &b,
-               int const restart, int const max_iter, P const tolerance)
+               int const max_iter, P const tolerance)
 {
   return bicgstab(
       [&](P const alpha, fk::vector<P, mem_type::view, resrc> const x_in,
@@ -132,22 +132,118 @@ bicgstab_euler(const P dt, kronmult_matrix<P> const &mat,
         lib_dispatch::axpy<resrc>(y.size(), alpha, x_in.data(), 1, y.data(), 1);
       },
       fk::vector<P, mem_type::view, resrc>(x), b, no_op_preconditioner<P>(),
-      restart, max_iter, tolerance);
+      max_iter, tolerance);
 }
 #endif
 
-// simple, node-local test version
+//*****************************************************************
+// Iterative template routine -- BiCGSTAB
+//
+// BiCGSTAB solves the unsymmetric linear system Ax = b
+// using the Preconditioned BiConjugate Gradient Stabilized method
+//
+// BiCGSTAB follows the algorithm described on p. 27 of the
+// SIAM Templates book.
+//
+// The return value indicates convergence within max_iter (input)
+// iterations (0), or no convergence within max_iter iterations (1).
+//
+// Upon successful return, output arguments have the following values:
+//
+//        x  --  approximate solution to Ax = b
+// max_iter  --  the number of iterations performed before the
+//               tolerance was reached
[…]
