+++ src/asgard_bicgstab.hpp
@@ -11,7 +11,7 @@ namespace asgard::solver
 template<typename P>
 bicgstab_info<P>
 bicgstab(fk::matrix<P> const &A, fk::vector<P> &x, fk::vector<P> const &b,
-         fk::matrix<P> const &M, int const restart, int const max_iter,
+         fk::matrix<P> const &M, int const max_iter,
          P const tolerance);
 
 #ifdef KRON_MODE_GLOBAL
@@ -22,7 +22,7 @@ bicgstab_euler(const P dt, matrix_entry mentry,
                global_kron_matrix<P> const &mat,
                fk::vector<P, mem_type::owner, resrc> &x,
                fk::vector<P, mem_type::owner, resrc> const &b,
-               int const restart, int const max_iter, P const tolerance);
+               int const max_iter, P const tolerance);
 #else
 // solves ( I - dt * mat ) * x = b
 template<typename P, resource resrc>
@@ -30,7 +30,7 @@ bicgstab_info<P>
 bicgstab_euler(const P dt, kronmult_matrix<P> const &mat,
                fk::vector<P, mem_type::owner, resrc> &x,
                fk::vector<P, mem_type::owner, resrc> const &b,
-               int const restart, int const max_iter, P const tolerance);
+               int const max_iter, P const tolerance);
 #endif
 
 } // namespace asgard::solver
