### 2024-06-14 9c8df49
Add unit tests
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
