+++ src/asgard_bicgstab_tests.cpp
@@ -1,0 +1,374 @@
+#include "asgard_bicgstab.hpp"
+#include "batch.hpp"
+#include "coefficients.cpp"
+#include "quadrature.hpp"
+#include "solver.hpp"
+#include "tests_general.hpp"
+
+using namespace asgard;
+
+struct distribution_test_init
+{
+  distribution_test_init() { initialize_distribution(); }
+  ~distribution_test_init() { finalize_distribution(); }
+};
+
+#ifdef ASGARD_USE_MPI
+static distribution_test_init const distrib_test_info;
+#endif
+
+parser get_parser(PDE_opts const pde_choice,
+                  fk::vector<int> const &starting_levels, int const degree,
+                  int const memory_limit)
+{
+  parser parse(pde_choice, starting_levels, degree);
+  parser_mod::set(parse, parser_mod::memory_limit, memory_limit);
+  return parse;
+}
+template<typename P>
+void test_kronmult(parser const &parse, P const tol_factor)
+{
+  auto pde = make_PDE<P>(parse);
+  options const opts(parse);
+  basis::wavelet_transform<P, resource::host> const transformer(opts, *pde);
+  generate_all_coefficients(*pde, transformer);
+
+  // assume uniform degree across dimensions
+  auto const degree = pde->get_dimensions()[0].get_degree();
+
+  elements::table const table(opts, *pde);
+  element_subgrid const my_subgrid(0, table.size() - 1, 0, table.size() - 1);
+
+  // setup x vector
+  unsigned int seed{666};
+  std::mt19937 mersenne_engine(seed);
+  std::uniform_int_distribution<int> dist(-4, 4);
+  auto const gen = [&dist, &mersenne_engine]() {
+    return dist(mersenne_engine);
+  };
+  auto const elem_size  = static_cast<int>(std::pow(degree, pde->num_dims));
+  fk::vector<P> const b = [&table, gen, elem_size]() {
+    fk::vector<P> output(elem_size * table.size());
+    std::generate(output.begin(), output.end(), gen);
+    return output;
+  }();
+
+  fk::vector<P> const gold = [&pde, &table, &my_subgrid, &b, elem_size]() {
+    auto const system_size = elem_size * table.size();
+    fk::matrix<P> A(system_size, system_size);
+    fk::vector<P> x(b);
+    build_system_matrix(*pde, table, A, my_subgrid);
+    // AA = I - dt*A;
+    fm::scal(P{-1.} * pde->get_dt(), A);
+    for (int i = 0; i < A.nrows(); ++i)
+    {
+      A(i, i) += 1.0;
+    }
+    std::vector<int> ipiv(A.nrows());
+    fm::gesv(A, x, ipiv);
+    return x;
+  }();
+
+  // perform bicgstab with system matrix A
+  fk::vector<P> const bicgstab = [&pde, &table, &my_subgrid, &gold, &b,
+                                  elem_size]() {
+    auto const system_size = elem_size * table.size();
+    fk::matrix<P> A(system_size, system_size);
+    fk::vector<P> x(gold);
+    int const max_iter = parser::DEFAULT_GMRES_OUTER_ITERATIONS;
+    P const tolerance  = std::is_same_v<float, P> ? 1e-6 : 1e-12;
+    build_system_matrix(*pde, table, A, my_subgrid);
+    // AA = I - dt*A;
+    fm::scal(P{-1.} * pde->get_dt(), A);
+    for (int i = 0; i < A.nrows(); ++i)
+    {
+      A(i, i) += 1.0;
+    }
+    solver::bicgstab(A, x, b, fk::matrix<P>(), max_iter,
+                     tolerance);
+    return x;
+  }();
+
[…]
