### 2024-06-14 071ebe9
Add bicgstab solver option
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
