+++ src/asgard_bicgstab.hpp
@@ -7,9 +7,9 @@
 
 namespace asgard::solver
 {
-// simple, node-local test version of gmres
+// simple, node-local test version of bicgstab
 template<typename P>
-bicgstab_info<P>
+gmres_info<P>
 bicgstab(fk::matrix<P> const &A, fk::vector<P> &x, fk::vector<P> const &b,
          fk::matrix<P> const &M, int const max_iter,
          P const tolerance);
@@ -26,7 +26,7 @@ bicgstab_euler(const P dt, matrix_entry mentry,
 #else
 // solves ( I - dt * mat ) * x = b
 template<typename P, resource resrc>
-bicgstab_info<P>
+gmres_info<P>
 bicgstab_euler(const P dt, kronmult_matrix<P> const &mat,
                fk::vector<P, mem_type::owner, resrc> &x,
                fk::vector<P, mem_type::owner, resrc> const &b,
