+++ src/asgard_bicgstab.cpp
@@ -54,7 +54,7 @@ public:
 
 // simple, node-local test version
 template<typename P>
-bicgstab_info<P>
+gmres_info<P>
 bicgstab(fk::matrix<P> const &A, fk::vector<P> &x, fk::vector<P> const &b,
          fk::matrix<P> const &M, int const max_iter,
          P const tolerance)
@@ -93,7 +93,7 @@ void apply_diagonal_precond(gpu::vector<P> const &pc, P dt,
 #endif
 
 template<typename P, resource resrc>
-bicgstab_info<P>
+gmres_info<P>
 bicgstab_euler(const P dt, matrix_entry mentry,
                global_kron_matrix<P> const &mat,
                fk::vector<P, mem_type::owner, resrc> &x,
@@ -118,7 +118,7 @@ bicgstab_euler(const P dt, matrix_entry mentry,
 }
 #else
 template<typename P, resource resrc>
-bicgstab_info<P>
+gmres_info<P>
 bicgstab_euler(const P dt, kronmult_matrix<P> const &mat,
                fk::vector<P, mem_type::owner, resrc> &x,
                fk::vector<P, mem_type::owner, resrc> const &b,
@@ -158,7 +158,7 @@ bicgstab_euler(const P dt, kronmult_matrix<P> const &mat,
 //*****************************************************************
 template<typename P, resource resrc, typename matrix_abstraction,
          typename preconditioner_abstraction>
-bicgstab_info<P>
+gmres_info<P>
 bicgstab(matrix_abstraction mat, fk::vector<P, mem_type::view, resrc> x,
          fk::vector<P, mem_type::owner, resrc> const &b,
          preconditioner_abstraction precondition,
@@ -186,7 +186,7 @@ bicgstab(matrix_abstraction mat, fk::vector<P, mem_type::view, resrc> x,
     //tol = resid;
     //max_iter = 0;
     //return 0;
-    return bicgstab_info<P>{resid, 0};
+    return gmres_info<P>{resid, 0};
   }
 
   for (int i = 1; i <= max_iter; i++)
@@ -196,7 +196,7 @@ bicgstab(matrix_abstraction mat, fk::vector<P, mem_type::view, resrc> x,
     {
       // tol = norm(r) / normb;
       // return 2;
-      return bicgstab_info<P>{resid, i};
+      return gmres_info<P>{resid, i};
     }
     if (i == 1)
       p.resize(r.size()) = r;
@@ -217,7 +217,7 @@ bicgstab(matrix_abstraction mat, fk::vector<P, mem_type::view, resrc> x,
       x = x + alpha(0) * phat;
       //tol = resid;
       //return 0;
-      return bicgstab_info<P>{resid, i};
+      return gmres_info<P>{resid, i};
     }
     shat.resize(s.size()) = s;
     fk::vector<P, mem_type::view> shat_v(shat);
@@ -234,36 +234,36 @@ bicgstab(matrix_abstraction mat, fk::vector<P, mem_type::view, resrc> x,
       //tol = resid;
       //max_iter = i;
       //return 0;
-      return bicgstab_info<P>{resid, i};
+      return gmres_info<P>{resid, i};
     }
     if (omega(0) == 0)
     {
       // tol = norm(r) / normb;
       // return 3;
-      return bicgstab_info<P>{fm::nrm2(r) / normb, i};
+      return gmres_info<P>{fm::nrm2(r) / normb, i};
     }
   }
   // tol = resid;
   // return 1;
-  return bicgstab_info<P>{resid, max_iter};
+  return gmres_info<P>{resid, max_iter};
 }
 
 #ifdef ASGARD_ENABLE_DOUBLE
 
-template bicgstab_info<double>
[…]
