### 2024-06-14 0b0e889
fix tol and max_iter
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
