### 2024-06-14 8810db5
remove debug output
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
