### 2024-06-14 1861596
remove duplicate test
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
