+++ src/asgard_bicgstab_tests.cpp
@@ -306,69 +306,4 @@ TEMPLATE_TEST_CASE("test kronmult w/ decompose", "[kronmult]", test_precs)
     test_kronmult(test_parse, tol_factor);
   }
 }
-
-TEMPLATE_TEST_CASE("poisson setup and solve", "[solver]", test_precs)
-{
-  SECTION("simple test case")
-  {
-    int const N_elements = 128;
-    int const N_nodes    = N_elements + 1;
-    int const degree     = 2;
-
-    TestType const x_min   = -2.0 * M_PI;
-    TestType const x_max   = 2.0 * M_PI;
-    TestType const phi_min = 0.0;
-    TestType const phi_max = 0.0;
-
-    int const N = (degree + 1) * N_elements;
-    fk::vector<TestType> poisson_source(N);
-    fk::vector<TestType> poisson_phi(N);
-    fk::vector<TestType> poisson_E(N);
-    fk::vector<TestType> x(N);
-    fk::vector<TestType> x_e(N_nodes);
-
-    fk::vector<TestType> diag;
-    fk::vector<TestType> off_diag;
-    solver::setup_poisson(N_elements, x_min, x_max, diag, off_diag);
-
-    // Assume Uniform Elements //
-    TestType dx = (x_max - x_min) / static_cast<TestType>(N_elements);
-
-    // Set Finite Element Nodes //
-    for (int i = 0; i < N_nodes; i++)
-    {
-      x_e[i] = x_min + i * dx;
-    }
-
-    // Set Source in DG Elements //
-    auto const lgwt = legendre_weights<TestType>(degree + 1, -1.0, 1.0,
-                                                 quadrature_mode::use_degree);
-    for (int i = 0; i < N_elements; i++)
-    {
-      for (int q = 0; q < degree + 1; q++)
-      {
-        int k             = i * (degree + 1) + q;
-        TestType x_q      = lgwt[0][q];
-        x[k]              = x_e[i] + 0.5 * dx * (1.0 + x_q);
-        poisson_source[k] = 0.5 * (1.0 - 0.5 * std::cos(0.5 * x[k])) - 1.0;
-      }
-    }
-
-    solver::poisson_solver(poisson_source, diag, off_diag, poisson_phi,
-                           poisson_E, degree, N_elements, x_min, x_max, phi_min,
-                           phi_max, solver::poisson_bc::periodic);
-
-    TestType error = 0.0;
-    for (int i = 0; i < N_elements; i++)
-    {
-      for (int q = 0; q < degree + 1; q++)
-      {
-        int k = i * (degree + 1) + q;
-        error += std::pow(poisson_phi[k] + (1.0 + std::cos(0.5 * x[k])), 2);
-      }
-    }
-
-    error = std::sqrt(error) / ((degree + 1) * N_elements);
-    REQUIRE(error < 5.0e-5);
-  }
 }
