### 2024-06-14 9e98e5b
minor cleanup
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
