+++ src/device/asgard_glkronmult_bcpu.cpp
@@ -189,84 +189,83 @@ void gbkron_mult_add(precision const A[], precision const x[], precision y[])
   {
     if constexpr (dim == 5)
     {
-
+      ASGARD_PRAGMA_OMP_SIMD(collapse(7))
+      for (int w = 0; w < n; w++)
+        for (int v = 0; v < n; v++)
+          for (int p = 0; p < n; p++)
+            for (int l = 0; l < n; l++)
+              for (int k = 0; k < n; k++)
+                for (int j = 0; j < n; j++)
+                  for (int s = 0; s < n; s++)
+                    y[ipow<n, 5>() * w + ipow<n, 4>() * v + ipow<n, 3>() * p + n * n * l + n * k + s] +=
+                        A[j * n + s] * x[ipow<n, 5>() * w + ipow<n, 4>() * v + ipow<n, 3>() * p + n * n * l + n * k + j];
+    }
+    else if constexpr (dim == 4)
+    {
+      ASGARD_PRAGMA_OMP_SIMD(collapse(7))
+      for (int w = 0; w < n; w++)
+        for (int v = 0; v < n; v++)
+          for (int p = 0; p < n; p++)
+            for (int l = 0; l < n; l++)
+              for (int j = 0; j < n; j++)
+                for (int s = 0; s < n; s++)
+                  for (int k = 0; k < n; k++)
+                    y[ipow<n, 5>() * w + ipow<n, 4>() * v + ipow<n, 3>() * p + n * n * l + n * s + k] +=
+                        x[ipow<n, 5>() * w + ipow<n, 4>() * v + ipow<n, 3>() * p + n * n * l + n * j + k] * A[j * n + s];
+    }
+    else if constexpr (dim == 3)
+    {
+      ASGARD_PRAGMA_OMP_SIMD(collapse(7))
+      for (int w = 0; w < n; w++)
+        for (int v = 0; v < n; v++)
+          for (int p = 0; p < n; p++)
+            for (int j = 0; j < n; j++)
+              for (int s = 0; s < n; s++)
+                for (int l = 0; l < n; l++)
+                  for (int k = 0; k < n; k++)
+                    y[ipow<n, 5>() * w + ipow<n, 4>() * v + ipow<n, 3>() * p + n * n * s + n * l + k] +=
+                        x[ipow<n, 5>() * w + ipow<n, 4>() * v + ipow<n, 3>() * p + n * n * j + n * l + k] * A[j * n + s];
+    }
+    else if constexpr (dim == 2)
+    {
+      for (int w = 0; w < n; w++)
+        for (int v = 0; v < n; v++)
+          for (int j = 0; j < n; j++)
+            for (int s = 0; s < n; s++)
+              for (int p = 0; p < n; p++)
+                for (int l = 0; l < n; l++)
+                  for (int k = 0; k < n; k++)
+                    y[ipow<n, 5>() * w + ipow<n, 4>() * v + ipow<n, 3>() * s + n * n * p + n * l + k] +=
+                        x[ipow<n, 5>() * w + ipow<n, 4>() * v + ipow<n, 3>() * j + n * n * p + n * l + k] * A[j * n + s];
+    }
+    else if constexpr (dim == 1)
+    {
+      ASGARD_PRAGMA_OMP_SIMD(collapse(7))
+      for (int w = 0; w < n; w++)
+        for (int j = 0; j < n; j++)
+          for (int s = 0; s < n; s++)
+            for (int v = 0; v < n; v++)
+              for (int p = 0; p < n; p++)
+                for (int l = 0; l < n; l++)
+                  for (int k = 0; k < n; k++)
+                    y[ipow<n, 5>() * w + ipow<n, 4>() * s + ipow<n, 3>() * v + n * n * p + n * l + k] +=
[…]
