### 2024-06-17 2b9da12
allow dual initialization and move of the PDE class
