### 2024-06-17 0cef31e
converting public const to private with getter
