+++ src/pde_tests.cpp
@@ -68,12 +68,12 @@ void test_source_vectors(PDE<P> const &pde, std::filesystem::path base_dir,
       auto const full_path = base_dir.replace_filename(
           source_string + "dim" + std::to_string(j) + ".dat");
       auto const gold = read_vector_from_txt_file<P>(full_path);
-      auto const fx   = pde.sources[i].source_funcs[j](x, time);
+      auto const fx   = pde.sources[i].source_funcs()[j](x, time);
       rmse_comparison(fx, gold, tol_factor);
     }
     P const gold = read_scalar_from_txt_file(
         base_dir.replace_filename(source_string + "time.dat"));
-    auto const fx = pde.sources[i].time_func(time);
+    auto const fx = pde.sources[i].time_func()(time);
     relaxed_fp_comparison(fx, gold, pde_eps_multiplier);
   }
 }
