### 2024-06-18 202cff6
Reduce temporaries
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
