+++ src/asgard_bicgstab_tests.cpp
@@ -193,7 +193,7 @@ TEMPLATE_TEST_CASE("simple GMRES", "[solver]", test_precs)
         A_gold, test, b_gold_2, fk::matrix<TestType>(),
         A_gold.ncols(), std::numeric_limits<TestType>::epsilon());
     std::cout.clear();
-    REQUIRE(bicgstab_output.error < std::numeric_limits<TestType>::epsilon());
+    REQUIRE(bicgstab_output.error < TestType{10.} * std::numeric_limits<TestType>::epsilon());
     REQUIRE(test == x_gold_2);
   }
 
