### 2024-06-18 8cdf016
remove extra calls to resize()
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
