+++ src/time_advance.cpp
@@ -880,13 +880,13 @@ imex_advance(PDE<P> &pde, matrix_list<P> &operator_matrices,
     if (solver == solve_opts::gmres)
     {
 #ifdef KRON_MODE_GLOBAL
-    pde.gmres_outputs[0] = solver::simple_gmres_euler(
-        pde.get_dt(), matrix_entry::imex_implicit, operator_matrices.kglobal,
-        f_1, f, restart, max_iter, tolerance);
+      pde.gmres_outputs[0] = solver::simple_gmres_euler(
+          pde.get_dt(), matrix_entry::imex_implicit, operator_matrices.kglobal,
+          f_1, f, restart, max_iter, tolerance);
 #else
-    pde.gmres_outputs[0] = solver::simple_gmres_euler(
-        pde.get_dt(), operator_matrices[matrix_entry::imex_implicit],
-        f_1, f, restart, max_iter, tolerance);
+      pde.gmres_outputs[0] = solver::simple_gmres_euler(
+          pde.get_dt(), operator_matrices[matrix_entry::imex_implicit],
+          f_1, f, restart, max_iter, tolerance);
 #endif
     }
     else if (solver == solve_opts::bicgstab)
@@ -998,9 +998,9 @@ imex_advance(PDE<P> &pde, matrix_list<P> &operator_matrices,
           P{0.5} * pde.get_dt(), matrix_entry::imex_implicit, operator_matrices.kglobal,
           f_2, f, restart, max_iter, tolerance);
 #else
-    pde.gmres_outputs[1] = solver::simple_gmres_euler(
-        P{0.5} * pde.get_dt(), operator_matrices[matrix_entry::imex_implicit],
-        f_2, f, restart, max_iter, tolerance);
+      pde.gmres_outputs[1] = solver::simple_gmres_euler(
+          P{0.5} * pde.get_dt(), operator_matrices[matrix_entry::imex_implicit],
+          f_2, f, restart, max_iter, tolerance);
 #endif
     }
     else if (solver == solve_opts::bicgstab)
