### 2024-06-19 7226053
fix failing test
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
