+++ src/asgard_vector.hpp
@@ -282,14 +282,14 @@ public:
    *  \param right elements on rhs for addition.
    *  \return vector with results of element-wise addition.
    */
-  template<mem_type omem, typename P_ = P, typename = std::enable_if_t<std::is_floating_point_v<P_>>>
-  vector<P, mem_type::owner, resrc> operator+(vector<P, omem, resrc> const &right) const;
+  template<mem_type omem, resource r_ = resrc, typename = enable_for_host<r_>>
+  vector<P> operator+(vector<P, omem> const &right) const;
   /*! element-wise subtraction
    *  \param right elements on rhs for subtraction.
    *  \return vector with results of element-wise subtraction.
    */
-  template<mem_type omem, typename P_ = P, typename = std::enable_if_t<std::is_floating_point_v<P_>>>
-  vector<P, mem_type::owner, resrc> operator-(vector<P, omem, resrc> const &right) const;
+  template<mem_type omem, resource r_ = resrc, typename = enable_for_host<r_>>
+  vector<P> operator-(vector<P, omem> const &right) const;
   /*! dot product of two vectors
    *  \param right elements on rhs for dot product.
    *  \return scalar result of dot product.
@@ -306,7 +306,8 @@ public:
    *  \param value value to multiply each element by
    *  \return result of vector*scalar product.
    */
-  vector<P, mem_type::owner, resrc> operator*(P const) const;
+  template<resource r_ = resrc, typename = enable_for_host<r_>>
+  vector<P> operator*(P const) const;
 
   /*! perform the matrix kronecker product by interpreting vector
    *  operands/return vector as single column matrices.
@@ -500,13 +501,6 @@ inline void fk::copy_vector(fk::vector<P, mem, resrc> &dest,
 // fk::vector class implementation starts here
 //
 //-----------------------------------------------------------------------------
-
-template<typename P, mem_type mem, resource resrc>
-fk::vector<P, mem, resrc> operator*(P a, const fk::vector<P, mem, resrc> &b)
-{
-  return b * a;
-}
-
 template<typename P, mem_type mem, resource resrc>
 fk::vector<P, mem, resrc>::vector() : data_{nullptr}, size_{0}
 {}
@@ -871,13 +865,14 @@ bool fk::vector<P, mem, resrc>::operator<(vector<P, omem> const &other) const
 // vector addition operator
 //
 template<typename P, mem_type mem, resource resrc>
-template<mem_type omem, typename, typename>
-fk::vector<P, mem_type::owner, resrc>
-fk::vector<P, mem, resrc>::operator+(vector<P, omem, resrc> const &right) const
+template<mem_type omem, resource, typename>
+fk::vector<P>
+fk::vector<P, mem, resrc>::operator+(vector<P, omem> const &right) const
 {
   expect(size() == right.size());
-  vector<P, mem_type::owner, resrc> ans(right);
-  lib_dispatch::axpy<resrc>(size(), P{1}, data(), 1, ans.data(), 1);
+  vector<P> ans(size());
+  for (auto i = 0; i < size(); ++i)
+    ans(i) = (*this)(i) + right(i);
   return ans;
 }
 
@@ -885,13 +880,14 @@ fk::vector<P, mem, resrc>::operator+(vector<P, omem, resrc> const &right) const
 // vector subtraction operator
 //
 template<typename P, mem_type mem, resource resrc>
-template<mem_type omem, typename, typename>
-fk::vector<P, mem_type::owner, resrc>
-fk::vector<P, mem, resrc>::operator-(vector<P, omem, resrc> const &right) const
[…]
