### 2024-06-19 c9b43c8
adjust program options
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
