+++ CMakeLists.txt
@@ -115,8 +115,13 @@ if (ASGARD_USE_MPI)
   option (KRON_MODE_GLOBAL "Global or local Kronecker products" OFF)
   option (KRON_MODE_GLOBAL_BLOCK "Use block variant for global kron" OFF)
 else()
-  option (KRON_MODE_GLOBAL "Global or local Kronecker products" ON)
-  option (KRON_MODE_GLOBAL_BLOCK "Use block variant for global kron" ON)
+  if (ASGARD_USE_CUDA)
+    option (KRON_MODE_GLOBAL "Global or local Kronecker products" ON)
+    option (KRON_MODE_GLOBAL_BLOCK "Use block variant for global kron" OFF)
+  else()
+    option (KRON_MODE_GLOBAL "Global or local Kronecker products" ON)
+    option (KRON_MODE_GLOBAL_BLOCK "Use block variant for global kron" ON)
+  endif()
 endif()
 if (KRON_MODE_GLOBAL_BLOCK AND NOT KRON_MODE_GLOBAL)
   message (FATAL_ERROR "KRON_MODE_GLOBAL_BLOCK=ON kronmult requires global KRON_MODE_GLOBAL=ON")
