### 2024-06-19 a3a72ce
fix OpenMP
