### 2024-06-19 a868f10
moved some variables
