### 2024-06-20 b62c67f
fix consts in pde
