+++ src/adapt.hpp
@@ -53,7 +53,7 @@ public:
     return this->get_initial_condition(
         pde.get_dimensions(),
         pde.has_analytic_soln() ? pde.exact_time()(0.0) : static_cast<P>(1.0),
-        pde.num_terms, pde.get_terms(), transformer, cli_opts);
+        pde.num_terms(), pde.get_terms(), transformer, cli_opts);
   }
 
   fk::vector<P> get_initial_condition(
