+++ src/pde/pde_base.hpp
@@ -51,7 +51,7 @@ enum class homogeneity
 // helper - single element size
 auto const element_segment_size = [](auto const &pde) {
   int const degree = pde.get_dimensions()[0].get_degree();
-  return static_cast<int>(std::pow(degree, pde.num_dims));
+  return static_cast<int>(std::pow(degree, pde.num_dims()));
 };
 
 // ---------------------------------------------------------------------------
@@ -550,7 +550,7 @@ template<typename P>
 class PDE
 {
 public:
-  PDE() : num_dims(0), num_sources(0), num_terms(0), max_level(0) {}
+  PDE() : num_dims_(0), num_sources_(0), num_terms_(0), max_level_(0) {}
   PDE(parser const &cli_input, int const num_dims_in, int const num_sources_in,
       int const max_num_terms, std::vector<dimension<P>> const dimensions,
       term_set<P> const terms, std::vector<source<P>> const sources_in,
@@ -596,14 +596,14 @@ public:
       std::vector<moment<P>> const moments_in = {},
       bool const do_collision_operator_in     = true)
   {
-    num_dims    = num_dims_in;
-    num_sources = num_sources_in;
-    num_terms   = get_num_terms(cli_input, max_num_terms);
-    max_level   = get_max_level(cli_input, dimensions);
+    num_dims_    = num_dims_in;
+    num_sources_ = num_sources_in;
+    num_terms_   = get_num_terms(cli_input, max_num_terms);
+    max_level_   = get_max_level(cli_input, dimensions);
 
-    sources            = std::move(sources_in);
-    exact_vector_funcs = std::move(exact_vector_funcs_in);
-    moments            = std::move(moments_in);
+    sources_            = std::move(sources_in);
+    exact_vector_funcs_ = std::move(exact_vector_funcs_in);
+    moments             = std::move(moments_in);
 
     exact_time_ = check_exact_time(exact_time_in);
 
@@ -613,34 +613,34 @@ public:
     dimensions_            = std::move(dimensions);
     terms_                 = std::move(terms);
 
-    expect(num_dims > 0);
-    expect(num_sources >= 0);
-    expect(num_terms > 0);
+    expect(num_dims_ > 0);
+    expect(num_sources_ >= 0);
+    expect(num_terms_ > 0);
 
-    expect(dimensions.size() == static_cast<unsigned>(num_dims));
+    expect(dimensions.size() == static_cast<unsigned>(num_dims_));
     expect(terms.size() == static_cast<unsigned>(max_num_terms));
-    expect(sources.size() == static_cast<unsigned>(num_sources));
+    expect(sources_.size() == static_cast<unsigned>(num_sources_));
 
     // ensure analytic solution functions were provided if this flag is set
     if (has_analytic_soln_)
     {
       // each set of analytical solution functions must have num_dim functions
-      for (const auto &md_func : exact_vector_funcs)
+      for (const auto &md_func : exact_vector_funcs_)
       {
-        expect(md_func.size() == static_cast<size_t>(num_dims) or md_func.size() == static_cast<size_t>(num_dims + 1));
+        expect(md_func.size() == static_cast<size_t>(num_dims_) or md_func.size() == static_cast<size_t>(num_dims_ + 1));
       }
     }
 
     // modify for appropriate level/degree
     // if default lev/degree not used
     auto const user_levels = cli_input.get_starting_levels().size();
-    if (user_levels != 0 && user_levels != num_dims)
+    if (user_levels != 0 && user_levels != num_dims_)
     {
[…]
