### 2024-06-20 1be9446
adjusting the two stream pde
