+++ src/pde/pde_base.hpp
@@ -567,22 +567,22 @@ public:
             moments_in, do_collision_operator_in)
   {}
   PDE(parser const &cli_input, int const num_dims_in, int const num_sources_in,
-      int const max_num_terms, std::vector<dimension<P>> const dimensions,
-      term_set<P> const terms, std::vector<source<P>> const sources_in,
-      std::vector<md_func_type<P>> const exact_vector_funcs_in,
-      scalar_func<P> const exact_time_in, dt_func<P> const get_dt,
-      bool const do_poisson_solve_in          = false,
-      bool const has_analytic_soln_in         = false,
-      std::vector<moment<P>> const moments_in = {},
-      bool const do_collision_operator_in     = true)
+      int const max_num_terms, std::vector<dimension<P>> dimensions,
+      term_set<P> terms, std::vector<source<P>> sources_in,
+      std::vector<md_func_type<P>> exact_vector_funcs_in,
+      scalar_func<P> exact_time_in, dt_func<P> get_dt,
+      bool const do_poisson_solve_in      = false,
+      bool const has_analytic_soln_in     = false,
+      std::vector<moment<P>> moments_in   = {},
+      bool const do_collision_operator_in = true)
   {
     initialize(cli_input, num_dims_in, num_sources_in,
-      max_num_terms, dimensions, terms, sources_in,
-      exact_vector_funcs_in,
-      exact_time_in, get_dt,
+      max_num_terms, std::move(dimensions), std::move(terms), std::move(sources_in),
+      std::move(exact_vector_funcs_in),
+      std::move(exact_time_in), std::move(get_dt),
       do_poisson_solve_in,
       has_analytic_soln_in,
-      moments_in,
+      std::move(moments_in),
       do_collision_operator_in);
   }
 
