+++ src/pde/pde_two_stream.hpp
@@ -16,17 +16,39 @@ class PDE_vlasov_two_stream : public PDE<P>
 public:
   PDE_vlasov_two_stream(parser const &cli_input)
   {
-    std::vector<dimension<P>> dimensions = {
-        dimension<P>(-2.0 * PI, 2.0 * PI, 4, default_degree,
-                     initial_condition_dim_x_0, nullptr, "x"),
-        dimension<P>(-2.0 * PI, 2.0 * PI, 3, default_degree,
-                     initial_condition_dim_v_0, nullptr, "v")};
+    int constexpr num_dims          = 2;
+    int constexpr num_sources       = 0;
+    int constexpr num_terms         = 4;
+    bool constexpr do_poisson_solve = true;
+    // disable implicit steps in IMEX
+    bool constexpr do_collision_operator = false;
+    bool constexpr has_analytic_soln     = false;
+    int constexpr default_degree         = 3;
+
+    term_set<P> terms = {std::vector<term<P>>{term_e1x, term_e1v},
+                         std::vector<term<P>>{term_e2x, term_e2v},
+                         std::vector<term<P>>{E_mass_x_pos, div_v_dn},
+                         std::vector<term<P>>{E_mass_x_neg, div_v_up}};
 
     // using empty instances for exact_vector_funcs and exact_time
-    this->initialize(cli_input, num_dims_, num_sources_, num_terms_, dimensions,
-                     terms_, sources_, std::vector<md_func_type<P>>{}, scalar_func<P>{},
-                     get_dt_, do_poisson_solve_, has_analytic_soln_, moments_,
-                     do_collision_operator_);
+    this->initialize(cli_input, num_dims, num_sources, num_terms,
+                     // defining the dimensions
+                     std::vector<dimension<P>>{
+                         dimension<P>(-2.0 * PI, 2.0 * PI, 4, default_degree,
+                                      initial_condition_dim_x_0, nullptr, "x"),
+                         dimension<P>(-2.0 * PI, 2.0 * PI, 3, default_degree,
+                                      initial_condition_dim_v_0, nullptr, "v")},
+                     // defining the set of terms
+                     term_set<P>{std::vector<term<P>>{term_e1x, term_e1v},
+                                 std::vector<term<P>>{term_e2x, term_e2v},
+                                 std::vector<term<P>>{E_mass_x_pos, div_v_dn},
+                                 std::vector<term<P>>{E_mass_x_neg, div_v_up}},
+                     std::vector<source<P>>{},       // no sources
+                     std::vector<md_func_type<P>>{}, // no exact solution
+                     scalar_func<P>{},               // no exact time
+                     get_dt_, do_poisson_solve, has_analytic_soln,
+                     std::vector<moment<P>>{moment0, moment1, moment2}, // moments
+                     do_collision_operator);
 
     param_manager.add_parameter(parameter<P>{"n", n});
     param_manager.add_parameter(parameter<P>{"u", u});
@@ -37,15 +59,6 @@ public:
   }
 
 private:
-  static int constexpr num_dims_          = 2;
-  static int constexpr num_sources_       = 0;
-  static int constexpr num_terms_         = 4;
-  static bool constexpr do_poisson_solve_ = true;
-  // disable implicit steps in IMEX
-  static bool constexpr do_collision_operator_ = false;
-  static bool constexpr has_analytic_soln_     = false;
-  static int constexpr default_degree          = 3;
-
   static fk::vector<P>
[…]
