### 2024-06-24 ef5f3f3
work around conversion of const char * to bool
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
