### 2024-06-24 6cfb5c6
re-enable int test cases
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
