### 2024-06-24 6b106a8
Try reducing duplicate code
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
