+++ src/solver_tests.cpp
@@ -91,6 +91,28 @@ void test_kronmult(parser const &parse, P const tol_factor)
 
   rmse_comparison(gold, gmres, tol_factor);
 
+  // perform bicgstab with system matrix A
+  fk::vector<P> const bicgstab = [&pde, &table, &my_subgrid, &gold, &b,
+                                  elem_size]() {
+    auto const system_size = elem_size * table.size();
+    fk::matrix<P> A(system_size, system_size);
+    fk::vector<P> x(gold);
+    int const max_iter = parser::DEFAULT_GMRES_OUTER_ITERATIONS;
+    P const tolerance  = std::is_same_v<float, P> ? 1e-6 : 1e-12;
+    build_system_matrix(*pde, table, A, my_subgrid);
+    // AA = I - dt*A;
+    fm::scal(P{-1.} * pde->get_dt(), A);
+    for (int i = 0; i < A.nrows(); ++i)
+    {
+      A(i, i) += 1.0;
+    }
+    solver::bicgstab(A, x, b, fk::matrix<P>(), max_iter,
+                     tolerance);
+    return x;
+  }();
+
+  rmse_comparison(gold, bicgstab, tol_factor);
+
   asgard::matrix_list<P> operator_matrices;
   asgard::adapt::distributed_grid adaptive_grid(*pde, opts);
   operator_matrices.make(matrix_entry::regular, *pde, adaptive_grid, opts);
@@ -114,6 +136,25 @@ void test_kronmult(parser const &parse, P const tol_factor)
   }();
 
   rmse_comparison(gold, matrix_free_gmres, tol_factor);
+
+  // perform matrix-free bicgstab
+  fk::vector<P> const matrix_free_bicgstab = [&operator_matrices, &gold, &b,
+                                              dt]() {
+    fk::vector<P> x(gold);
+    int const max_iter = parser::DEFAULT_GMRES_OUTER_ITERATIONS;
+    P const tolerance  = std::is_same_v<float, P> ? 1e-6 : 1e-12;
+#ifdef KRON_MODE_GLOBAL
+    solver::bicgstab_euler(dt, matrix_entry::regular, operator_matrices.kglobal, x,
+                           b, max_iter, tolerance);
+#else
+    solver::bicgstab_euler(dt, operator_matrices[matrix_entry::regular], x,
+                           b, max_iter, tolerance);
+#endif
+    return x;
+  }();
+
+  rmse_comparison(gold, matrix_free_bicgstab, tol_factor);
+
 #ifdef ASGARD_USE_CUDA
   // perform matrix-free gmres
   fk::vector<P> const mf_gpu_gmres = [&operator_matrices, &gold, &b, dt]() {
@@ -135,6 +176,26 @@ void test_kronmult(parser const &parse, P const tol_factor)
   }();
 
   rmse_comparison(gold, mf_gpu_gmres, tol_factor);
+
+  // perform matrix-free bicgstab
+  fk::vector<P> const mf_gpu_bicgstab = [&operator_matrices, &gold, &b, dt]() {
+    fk::vector<P, mem_type::owner, resource::device> x_d =
+        gold.clone_onto_device();
+    fk::vector<P, mem_type::owner, resource::device> b_d =
+        b.clone_onto_device();
+    int const max_iter = parser::DEFAULT_GMRES_OUTER_ITERATIONS;
+    P const tolerance  = std::is_same_v<float, P> ? 1e-6 : 1e-12;
+#ifdef KRON_MODE_GLOBAL
+    solver::bicgstab_euler(dt, matrix_entry::regular, operator_matrices.kglobal,
+                           x_d, b_d, max_iter, tolerance);
+#else
[…]
