### 2024-06-24 6cf6370
improve input help and throw an exception in case of failure
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
