### 2024-06-24 def3174
Fix macos build
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
