### 2024-06-25 316d0f8
scalapack solver is broken
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
