### 2024-06-25 9c6218b
Update src/asgard_kronmult_matrix.hpp
Co-authored-by: Steven Hahn <hahnse@ornl.gov>
