### 2024-06-25 baa3591
Update src/asgard_kronmult_matrix.cpp
Co-authored-by: Steven Hahn <hahnse@ornl.gov>
