### 2024-06-25 c8280f2
Update src/asgard_kronmult_matrix.hpp
Co-authored-by: Steven Hahn <hahnse@ornl.gov>
