### 2024-06-25 eb9e5a6
Update src/device/asgard_glkronmult_bcpu.cpp
Co-authored-by: Steven Hahn <hahnse@ornl.gov>
