### 2024-06-25 a754bcd
Update src/device/asgard_glkronmult_bcpu.cpp
Co-authored-by: Steven Hahn <hahnse@ornl.gov>
