### 2024-06-25 a3c37ba
Update src/device/asgard_glkronmult_bcpu.cpp
Co-authored-by: Steven Hahn <hahnse@ornl.gov>
