+++ src/solver.cpp
@@ -196,7 +196,6 @@ simple_gmres_euler(adapt::distributed_grid<P> const &adaptive_grid, int const el
 #endif
   return simple_gmres_euler(dt, mat, x, b, restart, max_iter, tolerance);
 }
-#endif
 
 /*! Generates a default number inner iterations when no use input is given
  * \param num_cols Number of columns in the A matrix.
