+++ src/time_advance.cpp
@@ -328,9 +328,9 @@ implicit_advance(PDE<P> &pde, kron_operators<P> &operator_matrices,
   int const degree    = pde.get_dimensions()[0].get_degree();
   int const elem_size = static_cast<int>(std::pow(degree, pde.num_dims()));
   auto const &plan    = adaptive_grid.get_distrib_plan();
+  auto const size = elem_size * adaptive_grid.get_subgrid(get_rank()).nrows();
 
 #ifdef ASGARD_USE_SCALAPACK
-  auto const size = elem_size * adaptive_grid.get_subgrid(get_rank()).nrows();
   fk::vector<P> x(size);
   exchange_results(x_orig, x, size, adaptive_grid.get_distrib_plan(), get_rank());
 #else
@@ -340,7 +340,6 @@ implicit_advance(PDE<P> &pde, kron_operators<P> &operator_matrices,
   {
     auto const sources =
         get_sources(pde, adaptive_grid, transformer, time + dt);
-    auto const size = elem_size * adaptive_grid.get_subgrid(get_rank()).ncols();
     fk::vector<P, mem_type::owner> sources_local(size);
     exchange_results(sources, sources_local, elem_size, plan, get_rank());
     fm::axpy(sources_local, x, dt);
@@ -360,7 +359,6 @@ implicit_advance(PDE<P> &pde, kron_operators<P> &operator_matrices,
   auto const bc = boundary_conditions::generate_scaled_bc(
       unscaled_parts[0], unscaled_parts[1], pde, grid.row_start, grid.row_stop,
       time + dt);
-  auto const size = elem_size * adaptive_grid.get_subgrid(get_rank()).ncols();
   fk::vector<P, mem_type::owner> bc_local(size);
   exchange_results(bc, bc_local, elem_size, plan, get_rank());
   fm::axpy(bc_local, x, dt);
@@ -531,7 +529,7 @@ imex_advance(PDE<P> &pde, kron_operators<P> &operator_matrices,
   int const N_elements = fm::two_raised_to(level);
 
   auto nodes          = gen_realspace_nodes(degree, level, min, max);
-  int const elem_size = static_cast<int>(std::pow(degree, pde.num_dims));
+  int const elem_size = static_cast<int>(std::pow(degree, pde.num_dims()));
 #ifdef ASGARD_USE_CUDA
   fk::vector<P, mem_type::owner, imex_resrc> f = f_0.clone_onto_device();
   fk::vector<P, mem_type::owner, imex_resrc> f_orig_dev =
