### 2024-06-25 df836f0
integer power method and address some pr comments
