+++ src/asgard_kronmult_matrix.hpp
@@ -24,6 +24,10 @@ std::vector<int> get_used_terms(PDE<precision> const &pde, options const &opts,
                                 imex_flag const imex);
 
 #ifndef KRON_MODE_GLOBAL
+// using LOCAL kronmult, can be parallelised using MPI but much more expensive
+// then the global modes below
+// also has the ability to limit memory and do some operations out-of-core
+// hence it is currently the most memory conserving mode
 
 /*!
  * \brief Holds data for the pre-computed memory sizes.
@@ -721,18 +725,205 @@ void update_kronmult_coefficients(PDE<P> const &pde,
                                   kron_sparse_cache &spcache,
                                   kronmult_matrix<P> &mat);
 
-#endif // KRON_MODE_GLOBAL
+/*!
+ * \brief Compute the stats for the memory usage
+ *
+ * Computes how to avoid overflow or the use of more memory than
+ * is available on the GPU device.
+ *
+ * \tparam P is float or double
+ *
+ * \param pde holds the problem data
+ * \param grid is the discretization
+ * \param program_options is the user provided options
+ * \param imex is the flag indicating the IMEX mode
+ * \param spcache holds precomputed data about the sparsity pattern which is
+ *        used between multiple calls to avoid recomputing identical entries
+ * \param memory_limit_MB can override the user specified limit (in MB),
+ *        if set to zero the user selection will be used
+ * \param index_limit should be set to the max value held by the 32-bit index,
+ *        namely 2^31 -2 since 2^31 causes an overflow,
+ *        a different value can be used for testing purposes
+ */
+template<typename P>
+memory_usage
+compute_mem_usage(PDE<P> const &pde, adapt::distributed_grid<P> const &grid,
+                  options const &program_options, imex_flag const imex,
+                  kron_sparse_cache &spcache, int memory_limit_MB = 0,
+                  int64_t index_limit = 2147483646, bool force_sparse = false);
 
-//! \brief Expressive indexing for the matrices
-enum matrix_entry
+
+/*!
+ * \brief Holds a list of matrices used for time-stepping.
+ *
+ * There are multiple types of matrices based on the time-stepping and the
+ * different terms being used. Matrices are grouped in one object so they can go
+ * as a set and reduce the number of matrix making.
+ */
+template<typename precision>
+struct kron_operators
 {
-  //! \brief Regular matrix for implicit or explicit time-stepping
-  regular = 0,
-  //! \brief IMEX explicit matrix
-  imex_explicit = 1,
-  //! \brief IMEX implicit matrix
-  imex_implicit = 2
+  //! \brief Makes a list of uninitialized matrices
+  kron_operators()
+  {
+#ifdef ASGARD_USE_GPU_MEM_LIMIT
+    load_stream = nullptr;
+#endif
+  }
+  //! \brief Frees the matrix list and any cache vectors
+  ~kron_operators()
+  {
+#ifdef ASGARD_USE_GPU_MEM_LIMIT
+    if (load_stream != nullptr)
+    {
+      auto status = cudaStreamDestroy(load_stream);
+      expect(status == cudaSuccess);
+    }
+#endif
+  }
+
+  //! \brief Apply the given matrix entry
+  template<resource rec = resource::host>
+  void apply(imex_flag entry, precision alpha, precision const x[], precision beta, precision y[]) const
+  {
+    matrices[static_cast<int>(entry)].template apply<rec>(alpha, x, beta, y);
+  }
[…]

[…]
