+++ src/asgard_kronmult_matrix.cpp
@@ -1067,6 +1067,42 @@ void build_preconditioner(PDE<precision> const &pde, int64_t const num_active,
   }
 }
 
+//! Returns true if the current term is identity and can be omitted
+template<typename precision>
+bool check_identity_term(PDE<precision> const &pde, int term_id, int dim)
+{
+  // Check that the volumne jacobian in this dimension is not identity,
+  if (pde.get_dimensions()[dim].volume_jacobian_dV != nullptr)
+    return false;
+  // Now check g_func, the lhs_func, and local surface jacobian are identity.
+  // TODO: There is an edge case where the mass matrices with the same volume
+  // jacobians can cancel out, but this requires us to know that both the
+  // dimensions' and the local terms' volume jacobian are equal.
+  // In the edge case, identity will be multiplied instead of ignored
+  // resulting in extra work but correct output.
+  for (auto const &pt : pde.get_terms()[term_id][dim].get_partial_terms())
+    if (pt.coeff_type() != coefficient_type::mass or
+        pt.g_func() != nullptr or
+        pt.lhs_mass_func() != nullptr or
+        pt.dv_func() != nullptr)
+      return false;
+  return true;
+}
+
+//! Return the direction of the flux term or -1 for mass term
+template<typename precision>
+int get_flux_direction(PDE<precision> const &pde, int term_id)
+{
+  for (int d = 0; d < pde.num_dims(); d++)
+    for (auto const &pt : pde.get_terms()[term_id][d].get_partial_terms())
+      if (pt.coeff_type() == coefficient_type::div or
+          pt.coeff_type() == coefficient_type::grad or
+          pt.coeff_type() == coefficient_type::penalty)
+        return d;
+  return -1;
+}
+
+#ifndef KRON_MODE_GLOBAL_BLOCK
 /*!
  * \brief Walks over the sparse matrix pattern and makes call-back for each non-zero
  *
@@ -1173,47 +1209,11 @@ void split_pattern(std::vector<int> const &pntr, std::vector<int> const &indx,
 }
 
 template<typename precision>
-bool check_identity_term(PDE<precision> const &pde, int term_id, int dim)
-{
-  // Check that the volumne jacobian in this dimension is not identity,
-  if (pde.get_dimensions()[dim].volume_jacobian_dV != nullptr)
-    return false;
-  // Now check g_func, the lhs_func, and local surface jacobian are identity.
-  // TODO: There is an edge case where the mass matrices with the same volume
-  // jacobians can cancel out, but this requires us to know that both the
-  // dimensions' and the local terms' volume jacobian are equal.
-  // In the edge case, identity will be multiplied instead of ignored
-  // resulting in extra work but correct output.
-  for (auto const &pt : pde.get_terms()[term_id][dim].get_partial_terms())
-    if (pt.coeff_type() != coefficient_type::mass or
-        pt.g_func() != nullptr or
-        pt.lhs_mass_func() != nullptr or
-        pt.dv_func() != nullptr)
-      return false;
-  return true;
-}
-
-template<typename precision>
-int get_flux_direction(PDE<precision> const &pde, int term_id)
-{
-  for (int d = 0; d < pde.num_dims(); d++)
-    for (auto const &pt : pde.get_terms()[term_id][d].get_partial_terms())
-      if (pt.coeff_type() == coefficient_type::div or
-          pt.coeff_type() == coefficient_type::grad or
-          pt.coeff_type() == coefficient_type::penalty)
-        return d;
-  return -1;
-}
-
-template<typename precision>
 global_kron_matrix<precision>
 make_global_kron_matrix(PDE<precision> const &pde,
                         adapt::distributed_grid<precision> const &dis_grid,
[…]
