+++ src/asgard_kronmult_matrix.hpp
@@ -912,20 +912,15 @@ private:
 #endif
 };
 
-#endif // ifndef KRON_MODE_GLOBAL
-
-// //! \brief Expressive indexing for the matrices
-// enum matrix_entry
-// {
-//   //! \brief Regular matrix for implicit or explicit time-stepping
-//   regular = 0,
-//   //! \brief IMEX explicit matrix
-//   imex_explicit = 1,
-//   //! \brief IMEX implicit matrix
-//   imex_implicit = 2
-// };
+#endif // end the ifndef KRON_MODE_GLOBAL
 
 #ifdef KRON_MODE_GLOBAL
+#ifndef KRON_MODE_GLOBAL_BLOCK
+// using GLOBAL kronmult where the explicitly form the sparse matrices
+// can be used with the cuSparse library (in CUDA mode)
+// will be replaced by the block global mode
+
+// forward declaration that allows set/update methods to be friends
 template<typename precision>
 class global_kron_matrix;
 
@@ -1039,13 +1034,13 @@ public:
 
   //! \brief Apply the operator, including expanding and remapping.
   template<resource rec = resource::host>
-  void apply(matrix_entry etype, precision alpha, precision const *x,
+  void apply(imex_flag etype, precision alpha, precision const *x,
              precision beta, precision *y) const;
 
   //! \brief The matrix evaluates to true if it has been initialized and false otherwise.
   operator bool() const { return (not gvals_.empty()); }
 
-  //! \brief Allows overwriting of the loaded coefficients.
+  //! \brief Returns the preconditioner.
   template<resource rec>
   auto const &get_diagonal_preconditioner() const
   {
@@ -1065,12 +1060,12 @@ public:
   }
 
   //! \brief Return the number of flops for the current matrix type
-  int64_t flops(matrix_entry etype) const
+  int64_t flops(imex_flag etype) const
   {
-    return flops_[flag2int(etype)];
+    return flops_[static_cast<int>(etype)];
   }
   //! \brief Check if the corresponding lock pattern is set.
-  bool local_unset(matrix_entry etype) const
+  bool local_unset(imex_flag etype) const
   {
     return (flops(etype) == 0);
   }
@@ -1107,16 +1102,6 @@ public:
       global_kron_matrix<precision> &mat);
 
 protected:
-  //! \brief Convert the imex flag to an index of the arrays.
-  static int flag2int(imex_flag imex)
-  {
-    return static_cast<int>(imex);
-  }
-  //! \brief Convert the matrix entry to an index of the arrays.
-  static int flag2int(matrix_entry imex)
-  {
-    return static_cast<int>(imex);
-  }
   // the workspace is kept externally to minimize allocations
   mutable workspace_type *work_;
 
@@ -1143,14 +1128,13 @@ protected:
   }
 
 private:
-  static constexpr int num_variants = 3;
   // description of the multi-indexes and the sparsity pattern
   // global case data
   int num_dimensions_;
   int64_t num_active_;
   int64_t num_padded_;
   std::vector<kronmult::permutes> perms_;
-  std::array<int64_t, num_variants> flops_;
+  std::array<int64_t, num_imex_variants> flops_;
   // data for the 1D tensors
   std::vector<std::vector<int>> gpntr_;
   std::vector<std::vector<int>> gindx_;
@@ -1163,7 +1147,7 @@ private:
   // preconditioner
   std::vector<precision> pre_con_;
 #ifdef ASGARD_USE_CUDA
-  std::array<kronmult::global_gpu_operations<precision>, num_variants> gpu_global;
[…]
