### 2024-06-25 a6277cb
block global case
