+++ src/solver.cpp
@@ -96,6 +96,7 @@ bicgstab(fk::matrix<P> const &A, fk::vector<P> &x, fk::vector<P> const &b,
 }
 
 #ifdef KRON_MODE_GLOBAL
+// preconditiner is only available in global mode
 template<typename P>
 void apply_diagonal_precond(std::vector<P> const &pc, P dt,
                             fk::vector<P, mem_type::view, resource::host> &x)
@@ -173,113 +174,6 @@ bicgstab_euler(const P dt, imex_flag imex,
 #endif
       }, max_iter, tolerance);
 }
-
-
-#ifdef KRON_MODE_GLOBAL
-// template<typename P>
-// void apply_diagonal_precond(std::vector<P> const &pc, P dt,
-//                             fk::vector<P, mem_type::view, resource::host> &x)
-// {
-// #pragma omp parallel for
-//   for (size_t i = 0; i < pc.size(); i++)
-//     x[i] /= (1.0 - dt * pc[i]);
-// }
-// #ifdef ASGARD_USE_CUDA
-// template<typename P>
-// void apply_diagonal_precond(gpu::vector<P> const &pc, P dt,
-//                             fk::vector<P, mem_type::view, resource::device> &x)
-// {
-//   kronmult::gpu_precon_jacobi(pc.size(), dt, pc.data(), x.data());
-// }
-// #endif
-
-
-
-// #ifdef KRON_MODE_GLOBAL_BLOCK
-// template<typename P, resource resrc>
-// gmres_info<P>
-// simple_gmres_euler(const P dt, matrix_entry mentry,
-//                    block_global_kron_matrix<P> const &mat,
-//                    fk::vector<P, mem_type::owner, resrc> &x,
-//                    fk::vector<P, mem_type::owner, resrc> const &b,
-//                    int const restart, int const max_iter, P const tolerance)
-// {
-//   auto const &pc = mat.template get_diagonal_preconditioner<resrc>();
-//
-//   return simple_gmres(
-//       [&](P const alpha, fk::vector<P, mem_type::view, resrc> const x_in,
-//           P const beta, fk::vector<P, mem_type::view, resrc> y) -> void {
-//         tools::time_event performance("kronmult - implicit", mat.flops(mentry));
-//         mat.template apply<resrc>(mentry, -dt * alpha, x_in.data(), beta, y.data());
-//         lib_dispatch::axpy<resrc>(y.size(), alpha, x_in.data(), 1, y.data(), 1);
-//       },
-//       fk::vector<P, mem_type::view, resrc>(x), b,
-//       [&](fk::vector<P, mem_type::view, resrc> &x_in) -> void {
-//         tools::time_event performance("kronmult - preconditioner", pc.size());
-//         apply_diagonal_precond(pc, dt, x_in);
-//       }, restart, max_iter, tolerance);
-// }
-// template<typename P, resource resrc>
-// gmres_info<P>
-// bicgstab_euler(const P dt, matrix_entry mentry,
-//                block_global_kron_matrix<P> const &mat,
-//                fk::vector<P, mem_type::owner, resrc> &x,
-//                fk::vector<P, mem_type::owner, resrc> const &b,
-//                int const max_iter, P const tolerance)
-// {
-//   auto const &pc = mat.template get_diagonal_preconditioner<resrc>();
-//
-//   return bicgstab(
-//     [&](P const alpha, fk::vector<P, mem_type::view, resrc> const x_in,
-//           P const beta, fk::vector<P, mem_type::view, resrc> y) -> void {
-//         tools::time_event performance("kronmult - implicit", mat.flops(mentry));
[…]
