+++ src/solver.hpp
@@ -45,46 +45,6 @@ bicgstab_euler(const P dt, imex_flag imex,
                fk::vector<P, mem_type::owner, resrc> const &b,
                int const max_iter, P const tolerance);
 
-
-
-#ifdef KRON_MODE_GLOBAL
-
-
-// #ifdef KRON_MODE_GLOBAL_BLOCK
-// template<typename P, resource resrc>
-// gmres_info<P>
-// simple_gmres_euler(const P dt, matrix_entry mentry,
-//                    block_global_kron_matrix<P> const &mat,
-//                    fk::vector<P, mem_type::owner, resrc> &x,
-//                    fk::vector<P, mem_type::owner, resrc> const &b,
-//                    int const restart, int const max_iter, P const tolerance);
-//
-// template<typename P, resource resrc>
-// gmres_info<P>
-// bicgstab_euler(const P dt, matrix_entry mentry,
-//                block_global_kron_matrix<P> const &mat,
-//                fk::vector<P, mem_type::owner, resrc> &x,
-//                fk::vector<P, mem_type::owner, resrc> const &b,
-//                int const max_iter, P const tolerance);
-// #endif
-
-#else
-// solves ( I - dt * mat ) * x = b
-// template<typename P, resource resrc>
-// gmres_info<P>
-// simple_gmres_euler(const P dt, kronmult_matrix<P> const &mat,
-//                    fk::vector<P, mem_type::owner, resrc> &x,
-//                    fk::vector<P, mem_type::owner, resrc> const &b,
-//                    int const restart, int const max_iter, P const tolerance);
-// // solves ( I - dt * mat ) * x = b
-// template<typename P, resource resrc>
-// gmres_info<P>
-// bicgstab_euler(const P dt, kronmult_matrix<P> const &mat,
-//                fk::vector<P, mem_type::owner, resrc> &x,
-//                fk::vector<P, mem_type::owner, resrc> const &b,
-//                int const max_iter, P const tolerance);
-#endif
-
 template<typename P>
 int default_gmres_restarts(int num_cols);
 
