### 2024-06-25 8b35c13
more cleanup and formatting
