+++ src/time_advance.cpp
@@ -213,9 +213,9 @@ explicit_advance(PDE<P> const &pde, kron_operators<P> &operator_matrices,
   // -- RK step 1
   fk::vector<P> fx(row_size);
   {
-    tools::time_event performance("kronmult");
+    tools::time_event performance(
+        "kronmult", operator_matrices.flops(imex_flag::unspecified));
     operator_matrices.apply(imex_flag::unspecified, 1.0, x.data(), 0.0, fx.data());
-    performance.flops = operator_matrices.flops(imex_flag::unspecified);
   }
   reduce_results(fx, reduced_fx, plan, get_rank());
 
@@ -267,9 +267,9 @@ explicit_advance(PDE<P> const &pde, kron_operators<P> &operator_matrices,
 
   // -- RK step 3
   {
-    tools::time_event performance("kronmult");
+    tools::time_event performance(
+        "kronmult", operator_matrices.flops(imex_flag::unspecified));
     operator_matrices.apply(imex_flag::unspecified, 1.0, x.data(), 0.0, fx.data());
-    performance.flops = operator_matrices.flops(imex_flag::unspecified);
   }
   reduce_results(fx, reduced_fx, plan, get_rank());
 
@@ -439,15 +439,6 @@ implicit_advance(PDE<P> &pde, kron_operators<P> &operator_matrices,
     pde.gmres_outputs[0] = solver::simple_gmres_euler<P, resource::host>(
         pde.get_dt(), imex_flag::unspecified, operator_matrices,
         fx, x, restart, max_iter, tolerance);
-// #ifdef KRON_MODE_GLOBAL
-//     pde.gmres_outputs[0] = solver::simple_gmres_euler<P, resource::host>(
-//         pde.get_dt(), imex_flag::unspecified, operator_matrices.kglobal,
-//         fx, x, restart, max_iter, tolerance);
-// #else
-//     pde.gmres_outputs[0] = solver::simple_gmres_euler(
-//         pde.get_dt(), operator_matrices[imex_flag::unspecified],
-//         fx, x, restart, max_iter, tolerance);
-// #endif
     return fx;
   }
   else if (solver == solve_opts::bicgstab)
@@ -461,15 +452,6 @@ implicit_advance(PDE<P> &pde, kron_operators<P> &operator_matrices,
     pde.gmres_outputs[0] = solver::bicgstab_euler<P, resource::host>(
         pde.get_dt(), imex_flag::unspecified, operator_matrices,
         fx, x, max_iter, tolerance);
-// #ifdef KRON_MODE_GLOBAL
-//     pde.gmres_outputs[0] = solver::bicgstab_euler<P, resource::host>(
-//         pde.get_dt(), matrix_entry::regular, operator_matrices.kglobal,
-//         fx, x, max_iter, tolerance);
-// #else
-//     pde.gmres_outputs[0] = solver::bicgstab_euler(
-//         pde.get_dt(), operator_matrices[imex_flag::unspecified],
-//         fx, x, max_iter, tolerance);
-// #endif
     return fx;
   }
   return x;
@@ -821,26 +803,18 @@ imex_advance(PDE<P> &pde, kron_operators<P> &operator_matrices,
   fk::vector<P, mem_type::owner, imex_resrc> fx(f.size());
 
   {
-//#ifdef KRON_MODE_GLOBAL
-    tools::time_event kronm_("kronmult - explicit", operator_matrices.flops(imex_flag::imex_explicit));
+    tools::time_event kronm_(
+        "kronmult - explicit", operator_matrices.flops(imex_flag::imex_explicit));
[…]
