+++ src/asgard_kronmult_matrix.cpp
@@ -67,7 +67,7 @@ void check_available_memory(int64_t baseline_memory, int64_t available_MB)
 }
 
 template<typename precision>
-kronmult_matrix<precision>
+local_kronmult_matrix<precision>
 make_kronmult_dense(PDE<precision> const &pde,
                     adapt::distributed_grid<precision> const &discretization,
                     options const &program_options, imex_flag const imex)
@@ -88,8 +88,8 @@ make_kronmult_dense(PDE<precision> const &pde,
   int const num_terms               = static_cast<int>(used_terms.size());
 
   if (used_terms.empty())
-    return asgard::kronmult_matrix<precision>(num_dimensions, kron_size,
-                                              num_rows, num_cols);
+    return asgard::local_kronmult_matrix<precision>(num_dimensions, kron_size,
+                                                    num_rows, num_cols);
 
   constexpr resource mode = resource::host;
 
@@ -136,7 +136,7 @@ make_kronmult_dense(PDE<precision> const &pde,
     }
   }
 
-  int64_t flps = kronmult_matrix<precision>::compute_flops(
+  int64_t flps = local_kronmult_matrix<precision>::compute_flops(
       num_dimensions, kron_size, num_terms, int64_t{num_rows} * num_cols);
 
   std::cout << "  kronmult dense matrix size: " << num_rows << " rows/cols\n";
@@ -154,12 +154,12 @@ make_kronmult_dense(PDE<precision> const &pde,
 
   auto gpu_elem = elem.clone_onto_device();
 
-  return asgard::kronmult_matrix<precision>(
+  return asgard::local_kronmult_matrix<precision>(
       num_dimensions, kron_size, num_rows, num_cols, num_terms,
       std::move(gpu_terms), std::move(gpu_elem), grid.row_start, grid.col_start,
       num_1d_blocks);
 #else
-  return asgard::kronmult_matrix<precision>(
+  return asgard::local_kronmult_matrix<precision>(
       num_dimensions, kron_size, num_rows, num_cols, num_terms,
       std::move(terms), std::move(elem), grid.row_start, grid.col_start,
       num_1d_blocks);
@@ -332,7 +332,7 @@ void compute_coefficient_offsets(kron_sparse_cache const &spcache,
 }
 
 template<typename precision>
-kronmult_matrix<precision>
+local_kronmult_matrix<precision>
 make_kronmult_sparse(PDE<precision> const &pde,
                      adapt::distributed_grid<precision> const &discretization,
                      options const &program_options,
@@ -357,8 +357,8 @@ make_kronmult_sparse(PDE<precision> const &pde,
   int const num_terms = static_cast<int>(used_terms.size());
 
   if (used_terms.empty())
-    return asgard::kronmult_matrix<precision>(num_dimensions, kron_size,
-                                              num_rows, num_cols);
+    return asgard::local_kronmult_matrix<precision>(
+        num_dimensions, kron_size, num_rows, num_cols);
 
   // size of the small kron matrices
   int const kron_squared = kron_size * kron_size;
@@ -625,7 +625,7 @@ make_kronmult_sparse(PDE<precision> const &pde,
                    (double(num_rows) * double(num_cols))
             << "%\n";
 
-  int64_t flops = kronmult_matrix<precision>::compute_flops(
+  int64_t flops = local_kronmult_matrix<precision>::compute_flops(
       num_dimensions, kron_size, num_terms, spcache.num_nonz);
   std::cout << "  -- work: " << flops * 1.E-9 << " Gflops\n";
 
@@ -638,7 +638,7 @@ make_kronmult_sparse(PDE<precision> const &pde,
                      get_MB<int>(list_iA[0].size()) +
[…]
