### 2024-06-26 8e65362
superfluous blas search
