+++ CMakeLists.txt
@@ -234,7 +234,6 @@ but then the appropriate CMAKE_CUDA_FLAGS must be set manually.")
         endif()
     endif()
 
-
     enable_language (CUDA)
     find_package (CUDAToolkit REQUIRED)
 
@@ -255,7 +254,6 @@ foreach(_prec ${ASGARD_PRECISIONS})
         message(FATAL_ERROR "invalid precision ${_prec}, supported types are 'float' and 'double'")
     endif()
 endforeach()
-
 
 set (ASGARD_NUM_QUADRATURE "10" CACHE STRING "Number of quadrature points to be used in realspace transformations.")
 
@@ -555,7 +553,7 @@ if (ASGARD_BUILD_TESTS)
 # helps keep track of flags and options and whether things we enabled correctly
   message(STATUS "")
   message(STATUS "ASGarD build options (summary):")
-  foreach(_opt CMAKE_CXX_FLAGS ASGARD_PRECISIONS ASGARD_USE_OPENMP ASGARD_USE_MPI ASGARD_USE_MATLAB KRON_MODE_GLOBAL)
+  foreach(_opt CMAKE_CXX_FLAGS ASGARD_PRECISIONS ASGARD_USE_OPENMP ASGARD_USE_MPI ASGARD_USE_MATLAB KRON_MODE_GLOBAL KRON_MODE_GLOBAL_BLOCK)
     message(STATUS "  ${_opt}=${${_opt}}")
   endforeach()
   if (ASGARD_USE_CUDA)
@@ -567,6 +565,4 @@ if (ASGARD_BUILD_TESTS)
   endif()
   message(STATUS "")
 
-else ()
-  message (WARNING "warning: tests will not be built.")
 endif ()
