### 2024-06-26 2a2e437
added the preconditioner to the local kronmult
