+++ src/pde/pde_fokkerplanck1_4p4.hpp
@@ -29,7 +29,7 @@ class PDE_fokkerplanck_1d_4p4 : public PDE<P>
 public:
   PDE_fokkerplanck_1d_4p4(parser const &cli_input)
       : PDE<P>(cli_input, num_dims_, num_sources_, num_terms_, dimensions_,
-               terms_, sources_, exact_vector_funcs_, exact_scalar_func_,
+               terms_, sources_, exact_vector_funcs_,
                get_dt_, do_poisson_solve_, has_analytic_soln_)
   {}
 
@@ -95,12 +95,6 @@ private:
       f(i)         = A / (2 * std::sinh(A) * std::exp(A * z(i)));
     }
     return f;
-  }
-
-  static P analytic_solution_time(P const time)
-  {
-    ignore(time);
-    return 1.0;
   }
 
   // specify source functions...
@@ -203,8 +197,5 @@ private:
   // define exact soln functions
   inline static std::vector<vector_func<P>> const exact_vector_funcs_ = {
       analytic_solution_dim0};
-
-  inline static scalar_func<P> const exact_scalar_func_ =
-      analytic_solution_time;
 };
 } // namespace asgard
