+++ src/pde/pde_collisional_landau.hpp
@@ -16,7 +16,7 @@ class PDE_collisional_landau : public PDE<P>
 public:
   PDE_collisional_landau(parser const &cli_input)
       : PDE<P>(cli_input, num_dims_, num_sources_, num_terms_, dimensions_,
-               terms_, sources_, exact_vector_funcs_, exact_scalar_func_,
+               terms_, sources_, exact_vector_funcs_,
                get_dt_, do_poisson_solve_, has_analytic_soln_, moments_,
                do_collision_operator_)
   {
@@ -479,7 +479,6 @@ private:
                                             terms_7, terms_8, terms_9};
 
   inline static std::vector<vector_func<P>> const exact_vector_funcs_ = {};
-  inline static scalar_func<P> const exact_scalar_func_               = {};
 
   static P get_dt_(dimension<P> const &dim)
   {
