### 2024-06-27 76588e1
remove unnecessary variable
