### 2024-06-27 e7927b7
allows for single level to be defined in the cli
