### 2024-06-27 0658223
remove some vector copy
