### 2024-06-27 12113b8
cleaned up the initial condition logic
