+++ src/pde/pde_continuity2.hpp
@@ -59,11 +59,7 @@ private:
   initial_condition_dim1(fk::vector<P> const x, P const t = 0)
   {
     ignore(t);
-    fk::vector<P> fx(x.size());
-    std::transform(x.begin(), x.end(), fx.begin(),
-                   [](P const &x_v) { return std::sin(2.0 * PI * x_v); });
-
-    return fx;
+    return fk::vector<P>(x.size());
   }
 
   // specify exact solution vectors/time function...
