+++ src/adapt.hpp
@@ -52,19 +52,19 @@ public:
   {
     return this->get_initial_condition(
         pde.get_dimensions(),
-        pde.has_exact_time() ? pde.exact_time(0.0) : static_cast<P>(1.0),
+//        pde.has_exact_time() ? pde.exact_time(0.0) : static_cast<P>(1.0),
         pde.num_terms(), pde.get_terms(), transformer, cli_opts);
   }
 
   fk::vector<P> get_initial_condition(
-      std::vector<dimension<P>> &dims, P const mult, int const num_terms,
+      std::vector<dimension<P>> &dims, int const num_terms,
       std::vector<std::vector<term<P>>> &terms,
       basis::wavelet_transform<P, resource::host> const &transformer,
       options const &cli_opts);
 
   void get_initial_condition(
       options const &cli_opts, std::vector<dimension<P>> const &dims,
-      std::vector<vector_func<P>> const &v_functions, P const mult,
+      std::vector<vector_func<P>> const &v_functions,
       basis::wavelet_transform<P, resource::host> const &transformer,
       fk::vector<P, mem_type::view> result);
 
