### 2024-06-27 c4dc068
remove copy, fk::vector initializes
