### 2024-06-27 9de762c
fix warning and template deduction
