### 2024-06-27 abb43ce
new interpolation framework
