+++ src/asgard_interpolation1d.cpp
@@ -1,0 +1,612 @@
+#include "asgard_interpolation1d.hpp"
+
+namespace asgard
+{
+#ifdef KRON_MODE_GLOBAL_BLOCK
+
+template<typename P>
+struct linear
+{
+  static P constexpr s3 = 1.73205080756887729; // sqrt(3.0)
+  // projection basis
+  static P plag0(P) { return 1; }
+  static P plag1(P x) { return 2 * s3 * x - s3; }
+  static P pwav0L(P x) { return s3 * (1 - 4 * x); }
+  static P pwav0R(P x) { return s3 * (-3 + 4 * x); }
+  static P pwav1L(P x) { return -1 + 6 * x; }
+  static P pwav1R(P x) { return -5 + 6 * x; }
+  // interpolation basis
+  static P ibas0(P x) { return -3 * x + 2; }
+  static P ibas1(P x) { return 3 * x - 1; }
+  static P iwav0L(P x) { return -6 * x + 2; }
+  static P iwav0R(P) { return 0; }
+  static P iwav1L(P) { return 0; }
+  static P iwav1R(P x) { return 6 * x - 4; }
+};
+
+/*
+ * The 0 level basis (Legendre polynomials or simple Lagrane ones)
+ * has support over the whole domain and it just linear functions.
+ * But all wavelets (projection or interpolation) have either a kinck
+ * or a dicontinuity at the middle of the domain.
+ * The inteprolation wavelets are also zero on half of the domain.
+ * Therefore, we need to consider intergation in two sub-domains,
+ * labeled L and R (left-right).
+ *
+ * Linear times linear functions give quadratic, 2 points
+ * of Gauss-Legendre is enough to integrate up to cubic basis.
+ *
+ * The other trick is when the two basis have partial overlap,
+ * then integration is done over the domain of the shorter support,
+ * taking that to be canonical (0, 1), which is transformed
+ * into the higher support with xl + scale * x_canonical,
+ * where xl is the left-point and scale is the ratio of
+ * the support of the two functions (scale > 1).
+ *
+ * At scale == 1, the domains overlap and no transformation is needed.
+ *
+ * The scaling is done outside of this class.
+ * Naturally, all integrals and and basis functions have to be
+ * scaled to the appropriate size. The integrals are scaled
+ * by the support of the smaller functions and the projection
+ * basis is scaled by sqrt(2.0)^level.
+ *
+ * Domain transformation and how to compute xl and scale:
+ *
+ * Assume that we have cells r and c (row/column but the argument
+ * is symmetic and can be applied in reverse).
+ * Also assume that size-r < size-c and r is inside c.
+ * We start with a canonical variable over r, call that \xi \in (0, 1),
+ * and we want to translate that to a canonical variable over cell c.
+ * We call the second variable z \in (0, 1).
+ * The reason for the translation is because we only have formulas
+ * for the basis functions over the canonical cell [0, 1].
+ *
+ * Over cell r, we integrate using canonical points and basis,
+ * but the canonical quadrature points translate to real
+ * points using:
+ *   x = rl + dr * \xi
+ * where rl is the left (real) end of the cell, dr is the sell size,
[…]
