+++ src/asgard_interpolation_tests.cpp
@@ -1,0 +1,341 @@
+#include "tests_general.hpp"
+
+#include "asgard_interpolation.hpp"
+#include "asgard_interptest_common.hpp"
+
+using namespace asgard;
+
+#ifdef KRON_MODE_GLOBAL_BLOCK
+void make_cellsd2p5(int *c)
+{
+  std::array<int, 10> i = {0, 0, 0, 1, 0, 2, 0, 3, 1, 0};
+  std::copy(i.begin(), i.end(), c);
+}
+
+template<typename precision>
+std::array<precision, 40> make_nodesd2p5()
+{
+  return {
+      // (0, 0)
+      precision{1} / 3, precision{1} / 3, precision{1} / 3, precision{2} / 3,
+      precision{2} / 3, precision{1} / 3, precision{2} / 3, precision{2} / 3,
+      // (0, 1)
+      precision{1} / 3, precision{1} / 6, precision{1} / 3, precision{5} / 6,
+      precision{2} / 3, precision{1} / 6, precision{2} / 3, precision{5} / 6,
+      // (0, 2)
+      precision{1} / 3, precision{1} / 12, precision{1} / 3, precision{5} / 12,
+      precision{2} / 3, precision{1} / 12, precision{2} / 3, precision{5} / 12,
+      // (0, 3)
+      precision{1} / 3, precision{7} / 12, precision{1} / 3, precision{11} / 12,
+      precision{2} / 3, precision{7} / 12, precision{2} / 3, precision{11} / 12,
+      // (1, 0)
+      precision{1} / 6, precision{1} / 3, precision{1} / 6, precision{2} / 3,
+      precision{5} / 6, precision{1} / 3, precision{5} / 6, precision{2} / 3,
+  };
+}
+
+/////////////////////////////////////////////////////////////////////
+//  Testing the loaded interpolation nodes
+/////////////////////////////////////////////////////////////////////
+TEMPLATE_TEST_CASE("md interpolation nodes", "[linear]", test_precs)
+{
+  constexpr TestType tol = (std::is_same_v<TestType, double>) ? 1.E-15 : 1.E-7;
+
+  //constexpr int order = 1;
+  connect_1d conn(2, connect_1d::hierarchy::volume);
+  kronmult::block_global_workspace<TestType> workspace;
+  interpolation<TestType> interp(2, &conn, &workspace);
+
+  vector2d<int> cells(2, 5);
+  make_cellsd2p5(cells[0]);
+
+  vector2d<TestType> nodes = interp.get_nodes(cells);
+  auto gold = make_nodesd2p5<TestType>();
+
+  REQUIRE(nodes.num_strips() == 20);
+
+  for (int i = 0; i < nodes.num_strips(); i++)
+    for (int j = 0; j < 2; j++)
+      REQUIRE(std::abs(nodes[i][j] - gold[i * 2 + j]) < tol);
+}
+
+/////////////////////////////////////////////////////////////////////
+//  Testing reconstruction of function values
+/////////////////////////////////////////////////////////////////////
+template<typename precision>
+void project_inver_md(int num_dimensions, int num_levels,
+                      std::vector<std::function<precision(precision)>> fcalls)
+{
+  constexpr precision tol = (std::is_same_v<precision, double>) ? 1.E-12 : 1.E-5;
+
+  constexpr int pterms = 2;
+
+  std::vector<vector_func<precision>> funcs;
+  for (int d = 0; d < num_dimensions; d++)
+    funcs.push_back([&, d](fk::vector<precision> const &x, precision const)
+                        -> fk::vector<precision>
+                    {
+                      fk::vector<precision> fx(x.size());
+                      for (int i = 0; i < fx.size(); i++)
+                        fx[i] = fcalls[d](x[i]);
+                      return fx;
+                    });
+
[…]
