### 2024-07-05 2d1aaa5
Update src/device/asgard_glkronmult_bcpu.cpp
Co-authored-by: Steven Hahn <hahnse@ornl.gov>
