### 2024-07-11 c22717b
interpolation for the pde class
