### 2024-07-11 eba29cf
fix some math
