### 2024-07-11 dd92d6a
Update src/main.cpp
Co-authored-by: Steven Hahn <hahnse@ornl.gov>
