### 2024-07-18 #6766 @joglekara
Done, nice work .

One comment - It would be nice to have a python plotter too :)
