### 2024-07-18 6660980
extracted the main logic into a separate file
