+++ examples/continuity_2d.cpp
@@ -1,0 +1,258 @@
+#include "asgard.hpp"
+
+// if ASGarD is compiled with double precision, this defaults to double
+// if only single precision is avaiable, this will be float
+using precision = asgard::default_precision;
+
+// this problem is identical to PDE_continuity2d
+// the only difference is that it is set as a stand-alone example
+// and it is not incorporated into the core library
+class example_continuity2d : public asgard::PDE<precision>
+{
+public:
+  // ratio of circumference to diameter of a circle
+  static constexpr precision PI = 3.141592653589793;
+
+  // short-hand notation of the asgard vector
+  using vector = asgard::fk::vector<precision>;
+  // short-hand notation for the asgard 1d dimension
+  using dimension = asgard::dimension<precision>;
+
+  // short-hand notations
+  // a collection of PDE operators (term_set) is broken up into separable
+  //   multidimensional terms (term_md)
+  // each multidimensional terms has num-dimensions 1d components (term1d)
+  // each one dimensional term consists of one or more partial terms
+  using partial_term_1d = asgard::partial_term<precision>;
+  using term_1d = asgard::term<precision>;
+  using term_md = std::vector<asgard::term<precision>>;
+  // term_set is actually std::vector<term_md>
+  using term_set = asgard::term_set<precision>;
+
+  // short-hand notation
+  using source     = asgard::source<precision>;
+  using source_set = std::vector<source>;
+
+  example_continuity2d(asgard::parser const &cli_input)
+  {
+    // these fields check correctness of the specification
+    int constexpr num_dimensions = 2;
+    int constexpr num_sources    = 3;
+    int constexpr num_terms      = 2;
+
+    // the Poisson solver is not used here
+    bool constexpr do_poisson_solve  = false;
+
+    // flagging terms as time_independent (using the false value)
+    // improves speed by keeping some constant matrices across time-steps
+    bool constexpr time_independent = false;
+
+    // define the two dimensional domain and provide the initial conditions
+    // the level and degree can be modified via the command line arguments
+    dimension dim0(-1.0,                // domain min
+                   1.0,                 // domain max
+                   2,                   // levels
+                   2,                   // degree
+                   initial_condition_x, // initial condition
+                   nullptr,             // Cartesian coordinates
+                   "x");                // reference name
+
+    dimension dim1(-2.0,                // domain min
+                   2.0,                 // domain max
+                   2,                   // levels
+                   2,                   // degree
+                   initial_condition_y, // initial condition
+                   nullptr,             // Cartesian coordinates
+                   "y");                // reference name
+
+    // building up the divergence operator from simple components
+    partial_term_1d par_diff(
+        asgard::coefficient_type::div, // uses derivative
+        op_coeff,                      // -1.0
+        nullptr,                       // no r.h.s. mass component
+        asgard::flux_type::downwind,
+        asgard::boundary_condition::periodic,
+        asgard::boundary_condition::periodic);
+
+    partial_term_1d par_mass(
+        asgard::coefficient_type::mass, // mass matrix, identity
+        nullptr,                        // no coefficient (Cartesian domain)
+        nullptr,                        // no r.h.s. mass component
+        asgard::flux_type::central,
+        asgard::boundary_condition::periodic,
+        asgard::boundary_condition::periodic);
+
+    term_1d diff_x(time_independent, "v_x.d_dx", {par_diff});
[…]
