+++ src/pde.hpp
@@ -54,8 +54,29 @@ namespace asgard
 //
 // ---------------------------------------------------------------------------
 
-template<typename P>
+template<typename type_to_test, typename control_type = void>
+struct has_precision_mode : std::false_type
+{};
 
+template<typename type_to_test>
+struct has_precision_mode<type_to_test, std::void_t<typename type_to_test::precision_mode>> : std::true_type
+{};
+
+template<typename pde_class>
+auto make_custom_pde(parser const &cli_input)
+{
+  static_assert(has_precision_mode<pde_class>::value,
+                "incorrect precision_mode, the pde_class must inherit from PDE<float> or PDE<double>");
+
+  using precision = typename pde_class::precision_mode;
+
+  static_assert(std::is_base_of<PDE<precision>, pde_class>::value,
+                "the requested PDE class must inherit from the asgard::PDE base-class");
+
+  return std::unique_ptr<PDE<precision>>(std::make_unique<pde_class>(cli_input));
+}
+
+template<typename P>
 std::unique_ptr<PDE<P>> make_PDE(parser const &cli_input)
 {
   switch (cli_input.get_selected_pde())
