### 2024-07-20 b36e910
max level sets automatically
