+++ python/pyasgard_test.py
@@ -1,0 +1,239 @@
+
+import unittest
+import numpy as np
+import os
+
+import asgard
+
+def basis_linear0(x):
+    return np.ones(x.shape)
+
+def basis_linear1(x):
+    return 2.0 * np.sqrt(3.0) * x - np.sqrt(3.0)
+
+def basis_linear2(x):
+    return ((x < 0.5) * np.sqrt(3.0) * (1.0 - 4.0 * x)
+            + (x > 0.5) * np.sqrt(3.0) * (-3.0 + 4.0 * x))
+
+def basis_linear3(x):
+    return (x < 0.5) * (-1.0 + 6.0 * x) + (x > 0.5) * (-5.0 + 6.0 * x)
+
+def basis_linear4(x):
+    return ((x < 0.5) * np.sqrt(2.0) * ((x < 0.25) * np.sqrt(3.0) * (1.0 - 8.0 * x)
+                                        + (x > 0.25) * np.sqrt(3.0) * (-3.0 + 8.0 * x)))
+
+def basis_linear5(x):
+    return ((x < 0.5) * np.sqrt(2.0) * ((x < 0.25) * (-1.0 + 12.0 * x)
+                                        + (x > 0.25) * (-5.0 + 12.0 * x)))
+
+def basis_linear6(x):
+    return ((x > 0.5) * np.sqrt(2.0) * ((x < 0.75) * np.sqrt(3.0) * (5.0 - 8.0 * x)
+                                        + (x > 0.75) * np.sqrt(3.0) * (-7.0 + 8.0 * x)))
+
+def basis_linear7(x):
+    return ((x > 0.5) * np.sqrt(2.0) * ((x < 0.75) * (-7.0 + 12.0 * x)
+                                        + (x > 0.75) * (-11.0 + 12.0 * x)))
+
+def continuity_1_exact(x, t):
+    return np.cos(2.0 * np.pi * x) * np.sin(t)
+
+def continuity_2_exact(x, y, t):
+    return np.cos(np.pi * x) * np.sin(2.0 * np.pi * y) * np.sin(2.0 * t)
+
+def continuity_3_exact(x, y, z, t):
+    return continuity_2_exact(x, y, t) * np.cos(2.0 * np.pi * z / 3.0)
+
+class asgard_reconstruction_tests(unittest.TestCase):
+    def almost_equal(self, x, y, message):
+        verbose = True
+        digits  = 13
+        np.testing.assert_almost_equal(x, y, digits, message, verbose)
+
+    def onedim_match_basis(self, ibasis, basis):
+        assert ibasis < 8, "1d test is set for 4 cells or 8 basis functions"
+        num_cells = 4
+        dmin = np.array((0.0,))
+        dmax = np.array((1.0,))
+        cells = np.array((0, 0, 1, 0, 2, 0, 2, 1), np.int32)
+        state = np.zeros((2 * num_cells,), np.float64)
+        state[ibasis] = 1.0
+        libasgard = asgard.libasgard
+        pntr = libasgard.asgard_make_dreconstruct_solution(
+                1, num_cells, np.ctypeslib.as_ctypes(cells),
+                2, np.ctypeslib.as_ctypes(state))
+
+        libasgard.asgard_reconstruct_solution_setbounds(
+            pntr, np.ctypeslib.as_ctypes(dmin), np.ctypeslib.as_ctypes(dmax))
+
+        x = np.linspace(0.0001, 0.9999, 8)
+        y = np.zeros(x.shape)
+
+        yref = basis(x)
+        libasgard.asgard_reconstruct_solution(
+            pntr, np.ctypeslib.as_ctypes(x), x.size, np.ctypeslib.as_ctypes(y))
+
[…]
