+++ src/asgard_reconstruct_tests.cpp
@@ -1,0 +1,10 @@
+#include "asgard_reconstruct.hpp"
+
+#include "tests_general.hpp"
+
+TEST_CASE("reconstruct a solution", "[reconstruct]")
+{
+  // The code is actually tested in Python when we reconstruct
+  // solutions to PDEs with known analytic solutions.
+  REQUIRE(true);
+}
