+++ src/asgard_reconstruct.cpp
@@ -1,0 +1,340 @@
+#include "asgard_reconstruct.hpp"
+
+#include "asgard_wavelet_basis.hpp"
+
+namespace asgard
+{
+
+template<typename precision>
+reconstruct_solution::reconstruct_solution(
+    int dims, int64_t num_cells, int const asg_cells[],
+    int degree, precision const solution[])
+    : pterms_(degree), block_size_(fm::ipow(pterms_, dims)), domain_scale(1.0)
+{
+  vector2d<int> raw_cells(dims, num_cells);
+  for (int64_t i = 0; i < num_cells; i++)
+    asg2tsg_convert(dims, asg_cells + 2 * dims * i, raw_cells[i]);
+
+  std::fill(inv_slope.begin(), inv_slope.end(), 1.0);
+  std::fill(shift.begin(), shift.end(), 0.0);
+
+  // compute a map that gives the sorted order of the indexes
+  std::vector<int64_t> imap(num_cells);
+  std::iota(imap.begin(), imap.end(), 0);
+
+  std::sort(imap.begin(), imap.end(),
+            [&](int64_t a, int64_t b) -> bool {
+              for (int d = 0; d < dims; d++)
+              {
+                if (raw_cells[a][d] < raw_cells[b][d])
+                  return true;
+                if (raw_cells[a][d] > raw_cells[b][d])
+                  return false;
+              }
+              return false; // equal should be false, as in < operator
+            });
+
+  // copy the solution and indexes in the correct order
+  coeff_.resize(num_cells * block_size_);
+  std::vector<int> sorted_cells(dims * num_cells);
+
+  auto is = sorted_cells.begin();
+  auto ic = coeff_.begin();
+  for (int64_t i = 0; i < num_cells; i++)
+  {
+    is = std::copy_n(raw_cells[imap[i]], dims, is);
+    ic = std::copy_n(solution + block_size_ * imap[i], block_size_, ic);
+  }
+
+  cells_ = indexset(dims, std::move(sorted_cells));
+
+  // analyze the graph and prepare cache data
+  build_tree();
+}
+
+void reconstruct_solution::set_domain_bounds(double const amin[], double const amax[])
+{
+  domain_scale = 1.0;
+  for (int d = 0; d < cells_.num_dimensions(); d++)
+  {
+    shift[d]    = amin[d];
+    double size = amax[d] - amin[d];
+
+    inv_slope[d]  = 1.0 / size;
+    domain_scale *= size;
+  }
+  domain_scale = 1.0 / std::sqrt(domain_scale);
+}
+
+void reconstruct_solution::reconstruct(double const x[], int num_x, double y[]) const
+{
+  int const num_dimensions = cells_.num_dimensions();
+  span2d<double const> x2d(num_dimensions, num_x, x);
+#pragma omp parallel
+{
+  std::array<double, max_num_dimensions> xn;
+#pragma omp for
+  for (int i = 0; i < num_x; i++)
+  {
+    for (int d = 0; d < num_dimensions; d++)
+      xn[d] = inv_slope[d] * (x2d[i][d] - shift[d]);
+    y[i] = domain_scale * walk_tree(xn.data());
+  }
+}
+}
+
+vector2d<int> reconstruct_solution::compute_dag_down() const
+{
+  int constexpr max_1d_kids = 2; // change for a different hierarchy
+
+  int num_dimensions = cells_.num_dimensions();
+  int64_t num_cells  = cells_.num_indexes();
+
+  vector2d<int> kids(num_dimensions * max_1d_kids, num_cells);
+
+#pragma omp parallel
+{
+  std::array<int, max_num_dimensions> kid;
+
+#pragma omp for
[…]

[…]
