+++ src/program_options.hpp
@@ -270,7 +270,11 @@ public:
         kmode(kmode_in), gmres_tolerance(gmres_tolerance_in),
         gmres_inner_iterations(gmres_inner_iterations_in),
         gmres_outer_iterations(gmres_outer_iterations_in),
-        max_adapt_levels(max_adapt_levels_in), restart_file(restart_file_in){};
+        max_adapt_levels(max_adapt_levels_in), restart_file(restart_file_in)
+  {
+    for (auto l : starting_levels_in)
+      max_level = std::max(max_level, l);
+  }
 
   explicit parser(
       std::string const &pde_choice_in, fk::vector<int> starting_levels_in,
