### 2024-07-23 a36c6aa
better check for base-of
