### 2024-07-24 85079fb
fixing install rpaths
