+++ examples/continuity_2d.cpp
@@ -10,9 +10,6 @@ using precision = asgard::default_precision;
 class example_continuity2d : public asgard::PDE<precision>
 {
 public:
-  // ratio of circumference to diameter of a circle
-  static constexpr precision PI = 3.141592653589793;
-
   // short-hand notation of the asgard vector
   using vector = asgard::fk::vector<precision>;
   // short-hand notation for the asgard 1d dimension
@@ -30,8 +27,8 @@ public:
   using term_set = asgard::term_set<precision>;
 
   // short-hand notation
-  using source     = asgard::source<precision>;
-  using source_set = std::vector<source>;
+  using source_md  = asgard::source<precision>;
+  using source_set = std::vector<source_md>;
 
   example_continuity2d(asgard::parser const &cli_input)
   {
@@ -66,50 +63,50 @@ public:
                    "y");                // reference name
 
     // building up the divergence operator from simple components
-    partial_term_1d par_diff(
+    partial_term_1d par_derivative(
         asgard::coefficient_type::div, // uses derivative
-        op_coeff,                      // -1.0
-        nullptr,                       // no r.h.s. mass component
+        op_coeff,                      // -1.0 (appears on the r.h.s)
+        nullptr,                       // no l.h.s. coefficient
         asgard::flux_type::downwind,
         asgard::boundary_condition::periodic,
         asgard::boundary_condition::periodic);
 
     partial_term_1d par_mass(
         asgard::coefficient_type::mass, // mass matrix, identity
-        nullptr,                        // no coefficient (Cartesian domain)
-        nullptr,                        // no r.h.s. mass component
+        nullptr,                        // no r.h.s. coefficient
+        nullptr,                        // no l.h.s. coefficient
         asgard::flux_type::central,
         asgard::boundary_condition::periodic,
         asgard::boundary_condition::periodic);
 
-    term_1d diff_x(time_independent, "v_x.d_dx", {par_diff});
+    term_1d d_x(time_independent, "d_x", {par_derivative});
     term_1d mass_y(time_independent, "mass_y", {par_mass});
 
     term_1d mass_x(time_independent, "mass_x", {par_mass});
-    term_1d diff_y(time_independent, "v_y.d_dy", {par_diff});
+    term_1d d_y(time_independent, "d_y", {par_derivative});
 
     term_set terms = {
-        term_md{diff_x, mass_y},
-        term_md{mass_x, diff_y}
+        term_md{d_x, mass_y},
+        term_md{mass_x, d_y}
     };
-
-    // in order to manufacture a specific analytic solution
-    // we use artificial source terms that more-or-less cancel the derivatives
-    // see the comments before the definition of diff_time
-    // each separable source consists of num-dimensions spacial funcitons
-    // and a time function
-    source s0({initial_condition_dx, initial_condition_y},  exact_sol_time);
-    source s1({initial_condition_x,  initial_condition_dy}, exact_sol_time);
-    source s2({initial_condition_x,  initial_condition_y},  exact_solution_dt);
-
-    source_set sources = {s0, s1, s2};
 
     // the PDE has analytic solution
     // the computed solution will be verified against the analytic one
     // the solution is the product of the the 3 functions
     static bool constexpr has_analytic_solution = true;
     std::vector<asgard::vector_func<precision>> exact_solution = {
-        initial_condition_x, initial_condition_y, exact_solution_time};
+        initial_condition_x, initial_condition_y, exact_time_vector};
+
+    // in order to manufacture a specific analytic solution
+    // we use artificial source terms that balance the derivatives
+    // see the comments before the definition of diff_time
[…]
