+++ src/pde/pde_riemann_1x2v.hpp
@@ -249,7 +249,7 @@ private:
   }
 
   inline static const partial_term<P> e1_pterm_x = partial_term<P>(
-      coefficient_type::div, e1_g1, nullptr, flux_type::downwind,
+      coefficient_type::div, e1_g1, nullptr, flux_type::upwind,
       boundary_condition::periodic, boundary_condition::periodic);
 
   inline static const partial_term<P> e1_pterm_v = partial_term<P>(
@@ -286,7 +286,7 @@ private:
   }
 
   inline static const partial_term<P> e2_pterm_x = partial_term<P>(
-      coefficient_type::div, e2_g1, nullptr, flux_type::upwind,
+      coefficient_type::div, e2_g1, nullptr, flux_type::downwind,
       boundary_condition::periodic, boundary_condition::periodic);
 
   inline static const partial_term<P> e2_pterm_v = partial_term<P>(
@@ -332,7 +332,7 @@ private:
   }
 
   inline static const partial_term<P> nu_v_pterm = partial_term<P>(
-      coefficient_type::div, nu_v_func, nullptr, flux_type::downwind,
+      coefficient_type::div, nu_v_func, nullptr, flux_type::upwind,
       boundary_condition::dirichlet, boundary_condition::dirichlet);
 
   inline static term<P> const nu_v_term =
