### 2024-07-26 #6766 @jedbrown
This entry is missing trailing commas after each line, which causes the build :point_up: to fail.
```bibtex
@article{huang2024,
title = {Adaptive Sparse Grid Discontinuous Galerkin Method: Review and Software Implementation}
journal = {Communications on Applied Mathematics and Computation}
```
Please also strip the `https://doi.org/` part out of the `doi` fields (notice how it's replicated in the output).
