+++ src/asgard_pde_system.hpp
@@ -1,213 +1,0 @@
-#pragma once
-
-#include "asgard_discretization.hpp"
-
-/*!
- * \defgroup AsgardPDESystem Asgard PDE system
- *
- * \par PDE System
- * Asgard can solve a system of partial differential equations (PDE)
- * described by multiple dimensions, fields and equations.
- * The data associated with the system is held in an object of the
- * asgard::pde_system class.
- *
- * \par Components
- * The PDE system is comprised by several components provided by the user.
- * - command line input
- * - a list (std::vector) of asgard::dimension_description objects
- * - a list (std::vector) of asgard::field_description objects
- */
-
-namespace asgard
-{
-/*!
- * \internal
- * \ingroup AsgardPDESystem
- * \brief Converts the user-provided descriptions into field objects that are used internally in the pde_system.
- *
- * Used in the initialization (constructor) of the pde_system objects.
- * \endinternal
- */
-template<typename precision>
-std::vector<field<precision>> make_fields_vector(
-    std::vector<field_description<precision>> const &descriptions)
-{
-  std::vector<field<precision>> result;
-  result.reserve(descriptions.size());
-  for (auto const &d : descriptions)
-    result.push_back(field<precision>(d));
-  return result;
-}
-
-/*!
- * \ingroup AsgardPDESystem
- * \brief Object of this class will contain the data describing a PDE system.
- *
- * \tparam precision is either \b float or \b double which indicates the use of 32-bit or 64-bit precision
- * \tparam rescr only supports asgard::resource::host, eventually will support asgard::resource::device for GPU support
- *
- * \par
- * The pde_system is object is the main container for the data associated with
- * PDE system. Time-stepping and solvers operate on the pde_system.
- *
- * Internally, the object will also store all relevant information about the
- * sparse grid discretization (e.g., the hash-tables) and will also hold the
- * current PDE state (i.e., the solution at the current time).
- *
- * \internal
- * \par
- * The constructor takes the user-provided descriptions objects,
- * but will convert those into the internal asgard::dimension and asgard::field
- * objects. The pde_system also contains the table associated with the
- * discretization. \b note : wll also contain the operators. \endinternal
- */
-template<typename precision, resource resrc = asgard::resource::host>
-class pde_system
-{
-public:
-  /*!
-   * \brief Constructor that initializes the internal data-structures.
-   *
-   * \param cli_input contains the command line inputs that can overwrite the Asgard defaults for discretization order and density
-   * \param dimensions contains the descriptions for all dimensions associated with the pde-system
-   * \param flieds_terms contains the descriptions for the fields used by the PDE system
-   *
-   * \throws runtime_error if the fields or dimensions have entries with repeated names
-   *         or if the fields are associated with non-existing dimensions
-   *
-   * \par
-   * Each filed must be associated with dimensions from the provided list.
-   * The fields and dimensions are related by the "name" parameters.
-   */
-  pde_system(parser const &cli_input,
-             std::vector<dimension_description<precision>> const &dimensions,
-             std::vector<field_description<precision>> const &field_terms)
-      : cli(cli_input), dims(cli, dimensions),
-        fields(make_fields_vector(field_terms)),
-        transformer(cli_input, dimensions.front().degree, true)
-  {
-    static_assert(std::is_same_v<precision, float> or
-                      std::is_same_v<precision, double>,
-                  "ASGARD supports only float and double as template "
-                  "parameters for precision.");
-
-    expect(dimensions.size() > 0);
-
-    for (auto &f : field_terms)
-      f.verify_dimensions(dims);
-
-    std::vector<std::string> field_names(fields.size());
[…]
