+++ src/asgard_tools.hpp
@@ -43,8 +43,7 @@
 #endif
 
 // simple profiling object
-// this is NOT thread safe for now - only one thread should be calling class
-// funcs at a time, if we need this, just need to wrap map access with locks
+// this is NOT thread safe
 namespace asgard::tools
 {
 #ifndef NDEBUG
