### 2024-07-29 1295b97
remove redundant method
