### 2024-07-29 d2e1208
simpler (optional) for-loop construct
