+++ src/asgard_tools_tests.cpp
@@ -122,3 +122,40 @@ TEMPLATE_TEST_CASE("test timer", "[timing test]", time_event_tag,
 
   SECTION("count") { REQUIRE(calls == static_cast<int>(times.size())); }
 }
+
+TEST_CASE("for-indexof testing", "[indexing testing]")
+{
+  std::vector<double> x(10);
+  std::vector<int64_t> r;
+  r.reserve(x.size());
+
+  for (auto i : indexof(x))
+  {
+    static_assert(std::is_same_v<decltype(i), int64_t>);
+    r.push_back(i);
+  }
+
+  REQUIRE(r.size() == x.size());
+  for (int64_t i = 0; i < 10; i++)
+    REQUIRE(r[i] == i);
+
+  std::vector<int> ir;
+  ir.reserve(8);
+
+  for (auto i : indexof<int>(1, 6))
+  {
+    static_assert(std::is_same_v<decltype(i), int>);
+    ir.push_back(i);
+  }
+
+  for (int i = 1; i < 6; i++)
+    REQUIRE(ir[i - 1] == i);
+
+  size_t s = 0;
+  for (auto i : indexof<size_t>(x.size()))
+  {
+    static_assert(std::is_same_v<decltype(i), size_t>);
+    s += i;
+  }
+  REQUIRE(s == 45);
+}
