+++ src/build_info.hpp.in
@@ -1,5 +1,10 @@
 #pragma once
 
+#define ASGARD_VERSION_MAJOR @asgard_VERSION_MAJOR@
+#define ASGARD_VERSION_MINOR @asgard_VERSION_MINOR@
+#define ASGARD_VERSION_PATCH @asgard_VERSION_PATCH@
+#define ASGARD_VERSION "@asgard_VERSION@"
+
 #define GIT_BRANCH "@GIT_BRANCH@"
 #define GIT_COMMIT_HASH "@GIT_COMMIT_HASH@"
 #define GIT_COMMIT_SUMMARY "@GIT_COMMIT_SUMMARY@"
