+++ src/sparse.hpp
@@ -1,5 +1,4 @@
 #pragma once
-#include "build_info.hpp"
 
 #ifdef ASGARD_USE_CUDA
 #include <cuda_runtime.h>
