+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{juno2018,
-   title={Discontinuous Galerkin algorithms for fully kinetic plasmas},
+   title={Discontinuous {Galerkin} algorithms for fully kinetic plasmas},
    volume={353},
    ISSN={0021-9991},
    url={http://dx.doi.org/10.1016/j.jcp.2017.10.009},
@@ -47,7 +47,7 @@ eprint = {https://doi.org/10.1063/1.3167820}
 }
 
 @misc{e2020,
-      title={Algorithms for Solving High Dimensional PDEs: From Nonlinear Monte Carlo to Machine Learning}, 
+      title={Algorithms for Solving High Dimensional PDEs: From Nonlinear {Monte} {Carlo} to Machine Learning},
       author={Weinan E and Jiequn Han and Arnulf Jentzen},
       year={2020},
       eprint={2008.13333},
@@ -72,7 +72,7 @@ abstract = {Sparse grids allow one to employ grid-based discretization methods i
 }
 
 @article{dazevedo2020,
-title = {Discontinuous Galerkin sparse grids methods for time domain Maxwell’s equations},
+title = {Discontinuous {Galerkin} sparse grids methods for time domain {Maxwell}'s equations},
 journal = {Computer Physics Communications},
 volume = {256},
 pages = {107412},
@@ -86,7 +86,7 @@ abstract = {In this paper, we present a time-explicit sparse grid discontinuous 
 }
 
 @misc{schnake2024,
-      title={Sparse-grid Discontinuous Galerkin Methods for the Vlasov-Poisson-Lenard-Bernstein Model}, 
+      title={Sparse-grid Discontinuous {Galerkin} Methods for the {Vlasov-Poisson-Lenard-Bernstein} Model},
       author={Stefan Schnake and Coleman Kendrick and Eirik Endeve and Miroslav Stoyanov and Steven Hahn and Cory D Hauck and David L Green and Phil Snyder and John Canik},
       year={2024},
       eprint={2402.06493},
@@ -95,7 +95,7 @@ abstract = {In this paper, we present a time-explicit sparse grid discontinuous 
 }
 
 @article{huang2024,
-title = {Adaptive Sparse Grid Discontinuous Galerkin Method: Review and Software Implementation},
+title = {Adaptive Sparse Grid Discontinuous {Galerkin} Method: Review and Software Implementation},
 journal = {Communications on Applied Mathematics and Computation},
 volume = {6},
 pages = {501-532},
@@ -107,7 +107,7 @@ abstract = { This paper reviews the adaptive sparse grid discontinuous Galerkin 
 }
 
 @misc{atanasov2017,
-      title={Sparse Grid Discretizations based on a Discontinuous Galerkin Method},
+      title={Sparse Grid Discretizations based on a Discontinuous {Galerkin} Method},
       author={Alexander B. Atanasov and Erik Schnetter},
       year={2017},
       eprint={1710.09356},
