### 2024-07-30 dcfc0e4
The publisher (Open Journals) acknowledges the DOE notice.
