### 2024-07-30 ad0ec6c
improved initialization
