### 2024-07-30 c111cd2
some more initialization
