### 2024-07-30 c1faed7
matlab plotting without building against matlab
