+++ examples/CMakeLists.txt
@@ -10,5 +10,6 @@ add_executable(example_continuity_2d  continuity_2d.cpp)
 target_link_libraries(example_continuity_2d asgard::asgard)
 
 if (asgard_PYTHON_FOUND)
-    configure_file("${CMAKE_CURRENT_SOURCE_DIR}/example_continuity_2d.py"  "${CMAKE_CURRENT_BINARY_DIR}/continuity_2d.py")
+    configure_file("${CMAKE_CURRENT_SOURCE_DIR}/example_continuity_2d.py"  "${CMAKE_CURRENT_BINARY_DIR}/example_continuity_2d.py")
+    configure_file("${CMAKE_CURRENT_SOURCE_DIR}/example_continuity_2d.m"   "${CMAKE_CURRENT_BINARY_DIR}/example_continuity_2d.m")
 endif()

[…]
