### 2024-07-31 4c46c1c
more detailed install docs
