+++ doxygen/installation.md
@@ -1,0 +1,142 @@
+# Installation
+
+[TOC]
+
+### Document Sections
+* Requirements
+* Installation
+* Basic usage
+
+
+### Requirements
+
+Minimum requirements run ASGarD
+* a C/C++ compiler with support for C++-17
+* [CMake](https://cmake.org/) build system
+* [Basic Linear Algebra Subroutine (BLAS)](https://en.wikipedia.org/wiki/Basic_Linear_Algebra_Subprograms) and [Linear Algebra PACKage](http://www.netlib.org/lapack/)
+    * many optimized BLAS and LAPACK implementations exist, e.g., OpenBLAS, MKL, Blis/Flame
+* GIT which works with CMake to find additional dependencies for testing
+
+Recommended but optional
+* [OpenMP](https://en.wikipedia.org/wiki/OpenMP) for CPU multi-threading
+    * supported by GCC and most recent versions of Clang
+* If you have Nvidia GPU ASGarD can take advantage of the [linear algebra libraries](https://developer.nvidia.com/cublas) and custom [CUDA kernels](https://developer.nvidia.com/cuda-zone)
+* [HDF5](https://en.wikipedia.org/wiki/Hierarchical_Data_Format) and [HighFive](https://bluebrain.github.io/HighFive/) libraries to output the solution
+* Python bindings using [h5py](https://www.h5py.org/) and [numpy](https://numpy.org/) for easier visualization and HDF5 post-processing
+
+Other CMake options
+* dynamic/shared libraries are used by default, static build is possible `-DBUILD_SHARED_LIBS=OFF`
+    * Python bindings require shared libraries
+* tests with CMake's ctest are enabled by default, disable with `-DASGARD_BUILD_TESTS=OFF`
+* ASGarD builds with both single (float) and double precision pick one for faster compile time
+
+ASGarD has the ability to automatically download and install OpenBLAS and HDF5.
+However, it is recommended to use system provided libraries, available in most Linux distributions and HPC systems.
+
+Limited support
+* [Message Passing Interface (MPI)](https://en.wikipedia.org/wiki/Message_Passing_Interface)
+* [Doxygen](http://www.doxygen.org/) documentation
+
+Note that ASGarD is under heavy development and some of the requirements may change.
+This document will be kept up-to-date with the changes.
+
+
+### Installation
+
+CMake uses out-of-source build, clone the repo and build in a subfolder
+```
+  git clone https://github.com/project-asgard/asgard.git
+  cd asgard
+  mkdir build
+  cd build
+  cmake \
+    -D CMAKE_BUILD_TYPE=Release \
+    -D CMAKE_INSTALL_PREFIX=../install \
+    -D CMAKE_CXX_FLAGS="-march=native -mtune=native" \
+    -D ASGARD_USE_OPENMP=ON \
+    -D ASGARD_USE_PYTHON=ON \
+    -D ASGARD_IO_HIGHFIVE=ON \
+    -D ASGARD_PRECISIONS=double \
+    ..
+  cmake --build . -j
+  ctest
+  cmake install .
+```
+
+On a OSX system, users have reported instabilities with homebrew provided HDF5.
+Also, OpenMP has limited benefits, due to what appears to be kernel scheduling.
+The BLAS/LAPACK acceleration need a flag to be set to the most recent mode.
+```
+  cmake \
+    -D CMAKE_BUILD_TYPE=Release \
+    -D CMAKE_INSTALL_PREFIX=../install \
[…]
