+++ README.md
@@ -47,72 +47,6 @@ Oak Ridge National Laboratory.
 
 *  C++17
 *  cmake 3.19
-*  blas
+*  BLAS/LAPACK
 
-# Optional depedencies
-
-* cuda
-* mpi
-* HighFive/hdf5
-* MATLAB
-* ScaLAPACK
-
-# Quickstart
-
-Download and build
-```
-git clone https://github.com/project-asgard/asgard.git
-cd asgard
-mkdir build && cd build
-cmake ../
-make
-ctest
-./asgard
-```
-
-For best performance (especially on accelerators) please pass `-DCMAKE_BUILD_TYPE=Release` to disable asserts when building the code.
-
-For specific platform build instructions, [see this wiki page.](https://github.com/project-asgard/asgard/wiki/platforms)
-
-# Example Execution
-
-Examples demonstrating ASGarD's capabilities are available [in the wiki](https://github.com/project-asgard/asgard/wiki/Example-Execution)
-
-To see a list of available PDEs, run `./asgard --available_pdes`. The listed PDEs can be selected using the `-p` argument to asgard.
-
-To see the list of all runtime options, run `./asgard --help`.
-
-## HighFive/HDF5 IO
-
-To enable HDF5 wavelet and realspace output, pass the `-DASGARD_IO_HIGHFIVE=ON`.
-
-Pass `-DASGARD_BUILD_HDF5=TRUE` and ASGarD will download and build HDF5 and HighFive libraries. If HIGHFIVE and/or HDF5 are already
-available, pass `-DASGARD_HIGHFIVE_PATH=<path to HighFive>` and/or `-DASGARD_HDF5_PATH=<path to HDF5>`.
-
-## MATLAB Interface
-
-Asgard can interface with MATLAB to plot, share data, and run scripts. To enable this, compile with `-DASGARD_USE_MATLAB=ON`.
-
-If using only for plotting purposes, then Asgard can be run without any additional arguments. However, for
-sharing data with MATLAB, Asgard will need to connect to a shared session. This can be done in MATLAB by
-running
- ```
- matlab.engine.shareEngine
- ```
-
-Asgard should automatically connect with this shared session, but the engine name can also be specified
-directly by giving the result of
-```
-matlab.engine.engineName
-```
-to Asgard with the `--matlab_name` option.
-
-Note: do not start MATLAB in the DG-SparseGrid repo when trying to use this for plotting.
-
-## ScaLAPACK Integration
-
-The ASGARD\_USE\_SCALAPACK build option adds the "scalapack" solver option to ASGarD. In the implicit timestep method, this option
-distributes the matrix solve across multiple processes and nodes and enables larger problems that are too large for a single computer.
-
-This build option is only available when ASGARD\_USE\_MPI=ON. Unfortunately, setting ASGARD\_USE\_SCALAPACK alone is not sufficient
-and doesn't change the ASGARD\_USE\_MPI setting.
+See the detailed [installation](Doxygen/Installation.md) instructions.
