### 2024-07-31 988e7b3
updated the meaning of degree
