+++ src/time_advance_tests.cpp
@@ -108,8 +108,7 @@ void time_advance_test(parser const &parse,
     auto const gold = read_vector_from_txt_file<P>(file_path);
 
     // each rank generates partial answer
-    auto const dof =
-        static_cast<int>(std::pow(parse.get_degree(), pde->num_dims()));
+    int64_t const dof  = fm::ipow(parse.get_degree() + 1, pde->num_dims());
     auto const subgrid = adaptive_grid.get_subgrid(get_rank());
     REQUIRE((subgrid.col_stop + 1) * dof - 1 <= gold.size());
     auto const my_gold = fk::vector<P, mem_type::const_view>(
@@ -205,9 +204,9 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", test_precs)
   std::string const pde_choice = "diffusion_2";
   int const num_dims           = 2;
 
-  SECTION("diffusion2, explicit, sparse grid, level 2, degree 2")
+  SECTION("diffusion2, explicit, sparse grid, level 2, degree 1")
   {
-    int const degree = 2;
+    int const degree = 1;
     int const level  = 2;
 
     auto constexpr tol_factor = get_tolerance<TestType>(100);
@@ -222,9 +221,9 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", test_precs)
     time_advance_test(parse, gold_base, tol_factor);
   }
 
-  SECTION("diffusion2, explicit, sparse grid, level 3, degree 3")
+  SECTION("diffusion2, explicit, sparse grid, level 3, degree 2")
   {
-    int const degree          = 3;
+    int const degree          = 2;
     int const level           = 3;
     auto constexpr tol_factor = get_tolerance<TestType>(100);
 
@@ -238,9 +237,9 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", test_precs)
     time_advance_test(parse, gold_base, tol_factor);
   }
 
-  SECTION("diffusion2, explicit, sparse grid, level 4, degree 4")
+  SECTION("diffusion2, explicit, sparse grid, level 4, degree 3")
   {
-    int const degree          = 4;
+    int const degree          = 3;
     int const level           = 4;
     auto constexpr tol_factor = get_tolerance<TestType>(1000000);
     auto const gold_base      = time_advance_base_dir / "diffusion2_sg_l4_d4_t";
@@ -253,9 +252,9 @@ TEMPLATE_TEST_CASE("time advance - diffusion 2", "[time_advance]", test_precs)
     time_advance_test(parse, gold_base, tol_factor);
   }
 
-  SECTION("diffusion2, explicit/non-uniform level, sparse grid, degree 2")
+  SECTION("diffusion2, explicit/non-uniform level, sparse grid, degree 1")
   {
-    int const degree          = 2;
+    int const degree          = 1;
     auto constexpr tol_factor = get_tolerance<TestType>(100);
 
     fk::vector<int> const levels{4, 5};
@@ -284,7 +283,7 @@ TEST_CASE("adaptive time advance")
   {
     auto const tol_factor        = get_tolerance<default_precision>(1000);
     std::string const pde_choice = "diffusion_2";
-    int const degree             = 4;
+    int const degree             = 3;
     fk::vector<int> const levels{3, 3};
     auto const gold_base =
         time_advance_base_dir / "diffusion2_ad_implicit_sg_l3_d4_t";
@@ -322,7 +321,7 @@ TEST_CASE("adaptive time advance")
   {
     auto const tol_factor        = get_tolerance<default_precision>(1000);
[…]
