### 2024-07-31 57b664f
Don't use gold as the starting value
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
