### 2024-08-01 9e51c30
new prog_opts class with better usage
