+++ src/boundary_conditions_tests.cpp
@@ -16,16 +16,9 @@ void test_boundary_condition_vector(PDE<P> &pde,
                                     std::filesystem::path const &gold_filename,
                                     P const tol_factor)
 {
-  /* setup stuff */
-  dimension<P> const &d = pde.get_dimensions()[0];
-  int const level       = d.get_level();
-  int const degree      = d.get_degree();
+  elements::table const table(pde);
 
-  auto const opts =
-      make_options({"-l", std::to_string(level), "-d", std::to_string(degree)});
-  elements::table const table(opts, pde);
-
-  basis::wavelet_transform<P, resource::host> const transformer(opts, pde);
+  basis::wavelet_transform<P, resource::host> const transformer(pde);
   generate_dimension_mass_mat<P>(pde, transformer);
   generate_all_coefficients<P>(pde, transformer);
 
@@ -56,20 +49,14 @@ void test_compute_boundary_condition(PDE<P> &pde,
                                      std::string gold_filename_prefix,
                                      P const tol_factor)
 {
-  // FIXME assume uniform degree
-  int const level  = pde.get_dimensions()[0].get_level();
-  int const degree = pde.get_dimensions()[0].get_degree();
-
-  auto const opts =
-      make_options({"-l", std::to_string(level), "-d", std::to_string(degree)});
-  basis::wavelet_transform<P, resource::host> const transformer(opts, pde);
+  basis::wavelet_transform<P, resource::host> const transformer(pde);
   generate_all_coefficients<P>(pde, transformer);
 
   term_set<P> const &terms_vec_vec = pde.get_terms();
 
   std::vector<dimension<P>> const &dimensions = pde.get_dimensions();
 
-  elements::table const table(opts, pde);
+  elements::table const table(pde);
 
   /* this timestep value must be consistent with the value used in the gold data
      generation scripts in matlab */
@@ -144,18 +131,11 @@ TEMPLATE_TEST_CASE("problem separability", "[boundary_condition]", test_precs)
      first and scale by multiplicative factors to at time + t */
   SECTION("time separability")
   {
-    /* setup stuff */
-    int const level  = 5;
-    int const degree = 4;
-    auto const pde   = make_PDE<TestType>(PDE_opts::diffusion_1, level, degree);
+    auto const pde = make_PDE<TestType>("-p diffusion_1 -l 5 -d 4");
 
-    auto const opts = make_options(
-        {"-l", std::to_string(level), "-d", std::to_string(degree)});
+    elements::table const table(*pde);
 
-    elements::table const table(opts, *pde);
-
-    basis::wavelet_transform<TestType, resource::host> const transformer(opts,
-                                                                         *pde);
+    basis::wavelet_transform<TestType, resource::host> const transformer(*pde);
     generate_dimension_mass_mat<TestType>(*pde, transformer);
     generate_all_coefficients<TestType>(*pde, transformer);
 
@@ -191,16 +171,11 @@ TEMPLATE_TEST_CASE("problem separability", "[boundary_condition]", test_precs)
   SECTION("element table split")
   {
     /* setup stuff */
-    int const level  = 5;
-    int const degree = 5;
-    auto const pde   = make_PDE<TestType>(PDE_opts::diffusion_1, level, degree);
-    auto const opts  = make_options(
-        {"-l", std::to_string(level), "-d", std::to_string(degree)});
+    auto const pde = make_PDE<TestType>("-p diffusion_1 -l 5 -d 5");
 
[…]
