+++ src/asgard.cpp
@@ -4,77 +4,28 @@ namespace asgard
 {
 
 template<typename precision>
-void simulate(parser const &cli_input, std::unique_ptr<PDE<precision>> &pde)
+void simulate(std::unique_ptr<PDE<precision>> &pde)
 {
-  options const opts(cli_input);
-
-  if (cli_input.show_libinfo())
-  {
-    print_info();
-    return;
-  }
+  rassert(!!pde, "invalid pde object");
+  auto const &options = pde->options();
 
   node_out() << "Branch: " << GIT_BRANCH << '\n';
   node_out() << "Commit Summary: " << GIT_COMMIT_HASH
-                     << GIT_COMMIT_SUMMARY << '\n';
+                 << GIT_COMMIT_SUMMARY << '\n';
   node_out() << "This executable was built on " << BUILD_TIME << '\n';
 
-  // -- generate pde
-  node_out() << "generating: pde..." << '\n';
-  if (not pde)
+#ifdef ASGARD_IO_HIGHFIVE
+  if (not options.restart_file.empty())
   {
-    if (cli_input.get_pde_string() == "custom")
-    {
-      node_out() << "when using a 'custom' pde the user must provide the object\n";
-      throw std::runtime_error("requested to work with a custom pde"
-                               " but no pde is provided");
-    }
-    pde = make_PDE<precision>(cli_input);
+    node_out() << "--- restarting from a file ---\n";
+    node_out() << "  filename: " << options.restart_file << '\n';
   }
-
-  // do this only once to avoid confusion
-  // if we ever do go to p-adaptivity (variable degree) we can change it then
-  auto const degree = pde->get_dimensions()[0].get_degree();
-
-  node_out() << "ASGarD problem configuration:" << '\n';
-  node_out() << "  selected PDE: " << cli_input.get_pde_string()
-             << '\n';
-  switch (degree)
-  {
-  case 0:
-    node_out() << "  degree: constant (0) \n";
-    break;
-  case 1:
-    node_out() << "  degree: linear (1) \n";
-    break;
-  case 2:
-    node_out() << "  degree: quadratic (2) \n";
-    break;
-  case 3:
-    node_out() << "  degree: cubic (3) \n";
-    break;
-  default:
-    node_out() << "  degree: " << degree << '\n';
-  };
-  node_out() << "  N steps: " << opts.num_time_steps << '\n';
-  node_out() << "  write freq: " << opts.wavelet_output_freq << '\n';
-  node_out() << "  realspace freq: " << opts.realspace_output_freq
-             << '\n';
-  node_out() << "  implicit: " << opts.use_implicit_stepping << '\n';
-  node_out() << "  full grid: " << opts.use_full_grid << '\n';
-  node_out() << "  CFL number: " << cli_input.get_cfl() << '\n';
-  node_out() << "  Poisson solve: " << opts.do_poisson_solve << '\n';
-  node_out() << "  starting levels: ";
-  node_out() << std::accumulate(
-                    pde->get_dimensions().begin(),
-                    pde->get_dimensions().end(), std::string(),
-                    [](std::string const &accum,
-                       dimension<precision> const &dim) {
-                      return accum + std::to_string(dim.get_level()) +
-                             " ";
-                    })
-             << '\n';
-  node_out() << "  max adaptivity levels: " << opts.max_level << '\n';
+  else if (get_local_rank() == 0)
+      options.print();
+#else
+  if (get_local_rank() == 0)
+    options.print();
+#endif
 
   node_out() << "--- begin setup ---" << '\n';
 
@@ -82,7 +33,9 @@ void simulate(parser const &cli_input, std::unique_ptr<PDE<precision>> &pde)
[…]
