### 2024-08-05 0767cdf
implementing input files
