+++ src/main.cpp
@@ -14,8 +14,18 @@ int main(int argc, char **argv)
     return 0;
   }
 
+  // custom projects may implement their own inputs
+  //   in addition to ASGarD standard ones
+  // however, this the main executable and all
+  //   unknonw cli commands are to be treated as errors
+  // first we generate warnigns for the user
+  bool constexpr ignore_unknown = false;
   // -- parse cli
-  asgard::prog_opts const cli_input(argc, argv);
+  asgard::prog_opts const cli_input(argc, argv, ignore_unknown);
+
+  // if there were unknown options, throw an error
+  if (not cli_input.externals.empty())
+    throw std::runtime_error("encountered unrecognized command line option");
 
   // main call to asgard, does all the work
   asgard::simulate<precision>(cli_input);
