+++ src/program_options_tests.cpp
@@ -16,242 +16,353 @@ TEST_CASE("new program options", "[single options]")
 {
   SECTION("no opts")
   {
-    prog_opts prog({""});
+    prog_opts prog(vecstrview({""}));
     REQUIRE_FALSE((prog.show_help and prog.show_pde_help));
   }
   SECTION("no help")
   {
-    prog_opts prog({"", "--help"});
+    prog_opts prog(vecstrview({"", "--help"}));
     REQUIRE(prog.show_help);
-    REQUIRE(prog_opts({"", "-?"}).show_help);
-    REQUIRE(prog_opts({"", "-h"}).show_help);
-    REQUIRE(prog_opts({"", "-help"}).show_help);
-    REQUIRE(prog_opts({"", "help"}).show_help);
-    prog_opts prog2({"", "-p?"});
+    REQUIRE(prog_opts(vecstrview({"", "-?"})).show_help);
+    REQUIRE(prog_opts(vecstrview({"", "-h"})).show_help);
+    REQUIRE(prog_opts(vecstrview({"", "-help"})).show_help);
+    REQUIRE(prog_opts(vecstrview({"", "help"})).show_help);
+    prog_opts prog2(vecstrview({"", "-p?"}));
     REQUIRE(prog2.show_pde_help);
-    REQUIRE(prog_opts({"", "-pde?"}).show_pde_help);
+    REQUIRE(prog_opts(vecstrview({"", "-pde?"})).show_pde_help);
   }
   SECTION("-step-method")
   {
-    REQUIRE_THROWS_WITH(prog_opts({"exe", "-step-method"}), "-step-method must be followed by a value, see exe -help");
-    REQUIRE_THROWS_WITH(prog_opts({"exe", "-s", "dummy"}), "invalid value for -s, see exe -help");
-    prog_opts prog({"", "-s", "expl"});
+    REQUIRE_THROWS_WITH(prog_opts(vecstrview({"exe", "-step-method"})),
+                        "-step-method must be followed by a value, see exe -help");
+    REQUIRE_THROWS_WITH(prog_opts(vecstrview({"exe", "-s", "dummy"})),
+                        "invalid value for -s, see exe -help");
+    prog_opts prog(vecstrview({"", "-s", "expl"}));
     REQUIRE(prog.step_method);
     REQUIRE(*prog.step_method == time_advance::method::exp);
-    prog = prog_opts({"", "-s", "impl"});
+    prog = prog_opts(vecstrview({"", "-s", "impl"}));
     REQUIRE(prog.step_method);
     REQUIRE(*prog.step_method == time_advance::method::imp);
-    REQUIRE(prog_opts({"", "-s", "imex"}).step_method.value() == time_advance::method::imex);
+    REQUIRE(prog_opts(vecstrview({"", "-s", "imex"})).step_method.value()
+            == time_advance::method::imex);
   }
   SECTION("-adapt-norm")
   {
-    REQUIRE_THROWS_WITH(prog_opts({"exe", "-adapt-norm"}),
+    REQUIRE_THROWS_WITH(prog_opts(vecstrview({"exe", "-adapt-norm"})),
                         "-adapt-norm must be followed by a value, see exe -help");
-    REQUIRE_THROWS_WITH(prog_opts({"exe", "-an", "dummy"}), "invalid value for -an, see exe -help");
-    prog_opts prog({"", "-an", "l2"});
+    REQUIRE_THROWS_WITH(prog_opts(vecstrview({"exe", "-an", "dummy"})),
+                        "invalid value for -an, see exe -help");
+    prog_opts prog(vecstrview({"", "-an", "l2"}));
[…]
