+++ testing/sandbox.cpp
@@ -6,6 +6,8 @@ using prec = asgard::default_precision;
 
 int main(int argc, char **argv)
 {
+  ignore(argc);
+  ignore(argv);
   // keep this file clean for each PR
   // allows someone to easily come here, dump code and start playing
   // this is good for prototyping and quick-testing features/behavior
