### 2024-08-06 5f52b1e
if not testing, skip test-install
