### 2024-08-06 84e4b90
move the install path of the shell env script
