+++ asgard-env.sh
@@ -2,7 +2,7 @@
 
 # file to source that will load the ASGarD environment
 
-export PATH=$PATH:"@CMAKE_INSTALL_PREFIX@"/bin/
+export PATH=$PATH:"@CMAKE_INSTALL_PREFIX@/bin/"
 
 if [[ "@ASGARD_USE_PYTHON@" == "ON" ]]; then
     export PYTHONPATH=$PYTHONPATH:"@_asgard_python_path@"
