### 2024-08-06 fb15d7f
rerrange cmake options and clean up deprecated ones
