+++ CMakeLists.txt
@@ -89,34 +89,31 @@ endif()
 # set up possible commandline input variable defaults (override with -D)
 include(CMakeDependentOption)
 option (BUILD_SHARED_LIBS "Build the ASGarD library with dynamic or static linking" ON)
+
 option (ASGARD_BUILD_TESTS "Build tests for asgard" ON)
-option (ASGARD_BUILD_OPENBLAS "Download and build our own OpenBLAS" OFF)
+option (ASGARD_USE_OPENMP "Optional OpenMP support for asgard" OFF)
+option (ASGARD_USE_CUDA "Optional CUDA support for asgard" OFF)
+cmake_dependent_option (ASGARD_USE_GPU_MEM_LIMIT "Allow the ability to limit the GPU memory used by kronmult (can hurt performance)" OFF "ASGARD_USE_CUDA" OFF)
+option (ASGARD_USE_MPI "Optional distributed computing support for asgard" OFF)
+cmake_dependent_option (ASGARD_USE_SCALAPACK "(deprecated) Use optional scalapack linear algebra library" OFF "ASGARD_USE_MPI" OFF)
+
+option (ASGARD_USE_PYTHON "Optional Python tool for post-processing, plotting and quick prototyping" OFF)
 option (ASGARD_IO_HIGHFIVE "Use the HighFive HDF5 header library for I/O" OFF)
+
+option (ASGARD_BUILD_OPENBLAS "Download and build our own OpenBLAS" OFF)
 option (ASGARD_BUILD_HDF5 "Download and build our own HDF5/HighFive" OFF)
 option (ASGARD_HIGHFIVE_PATH "Optional location of existing HighFive library" "")
 option (ASGARD_HDF5_PATH "Optional location of lib/ containing HDF5" "")
-option (ASGARD_BUILD_PROFILE_DEPS "Download and build our own tools (e.g.  graphviz)" OFF)
-option (ASGARD_PROFILE_GPROF "enable profiling support for using gprof" "")
-option (ASGARD_PROFILE_XRAY "enable profiling support for using LLVM XRAY" "")
-option (ASGARD_PROFILE_GPERF_CPU "enable cpu profiling support using gperftools" "")
-option (ASGARD_PROFILE_GPERF_MEM "enable memory profiling support using gperftools" "")
-option (ASGARD_PROFILE_PERF "enable profiling support for using linux perf" "")
-option (ASGARD_PROFILE_VALGRIND "enable profiling support for using valgrind" "")
-option (ASGARD_GRAPHVIZ_PATH "optional location of bin/ containing dot executable" "")
-option (ASGARD_USE_CUDA "Optional CUDA support for asgard" OFF)
-cmake_dependent_option (ASGARD_USE_GPU_MEM_LIMIT "Allow the ability to limit the GPU memory used by kronmult (can hurt performance)" OFF "ASGARD_USE_CUDA" OFF)
-option (ASGARD_USE_OPENMP "Optional OpenMP support for asgard" OFF)
-option (ASGARD_USE_MPI "Optional distributed computing support for asgard" OFF)
-cmake_dependent_option (ASGARD_USE_SCALAPACK "Use optional scalapack linear algebra library" OFF "ASGARD_USE_MPI" OFF)
-option (ASGARD_USE_MATLAB "Optional MATLAB plotting support for asgard" OFF)
+
+option (ASGARD_USE_MATLAB "(deprecated) Optional MATLAB plotting support for asgard" OFF)
 option (ASGARD_USE_PCH "Enable precompiled header files." OFF)
[…]
