### 2024-08-06 6db4a37
cleaned some of the STL clutter in the headers
