+++ src/distribution.hpp
@@ -8,10 +8,6 @@
 #include "cblacs_grid.hpp"
 #endif
 
-#include <list>
-#include <map>
-#include <vector>
-
 namespace asgard
 {
 // simple struct for representing a range within the element grid
