+++ src/permutations.hpp
@@ -1,8 +1,6 @@
 #pragma once
 #include "asgard_matrix.hpp"
 #include "asgard_vector.hpp"
-#include <functional>
-#include <vector>
 
 // -----------------------------------------------------------------------------
 // permutations
