+++ src/quadrature.hpp
@@ -2,7 +2,6 @@
 
 #include "asgard_matrix.hpp"
 #include "asgard_vector.hpp"
-#include <array>
 
 namespace asgard
 {
