### 2024-08-06 a3628c6
templates the reconstruct way, degree 0 and 1
