### 2024-08-07 7d58d23
Update missing DOIs in references
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
