+++ paper/paper.bib
@@ -17,7 +17,9 @@
   author={Bandrauk, Andr{\'e} D and Delfour, Michel C and Le Bris, Claude},
   volume={41},
   year={2007},
-  publisher={American Mathematical Soc.}
+  publisher={American Mathematical Soc.},
+  doi = {10.1090/crmp/041},
+  URL = {https://doi.org/10.1090/crmp/041}
 }
 
 @article{dorf2013,
@@ -46,13 +48,18 @@ URL = {https://doi.org/10.1063/1.3167820},
 eprint = {https://doi.org/10.1063/1.3167820}
 }
 
-@misc{e2020,
-      title={Algorithms for Solving High Dimensional PDEs: From Nonlinear {Monte} {Carlo} to Machine Learning},
-      author={Weinan E and Jiequn Han and Arnulf Jentzen},
-      year={2020},
-      eprint={2008.13333},
-      archivePrefix={arXiv},
-      primaryClass={math.NA}
+@article{e2020,
+doi = {10.1088/1361-6544/ac337f},
+url = {https://dx.doi.org/10.1088/1361-6544/ac337f},
+year = {2021},
+month = {dec},
+publisher = {IOP Publishing},
+volume = {35},
+number = {1},
+pages = {278},
+author = {Weinan E and Jiequn Han and Arnulf Jentzen},
+title = {Algorithms for solving high dimensional PDEs: from nonlinear {Monte} {Carlo} to machine learning},
+journal = {Nonlinearity}
 }
 
 @article{pfluger2010,
@@ -85,13 +92,17 @@ keywords = {Maxwell’s equations, Time domain, Discontinuous Galerkin method, S
 abstract = {In this paper, we present a time-explicit sparse grid discontinuous Galerkin method for solving the three-dimensional time-domain Maxwell equations. The conservation properties and convergence rates are established for different choices of numerical fluxes. The convergence rates are proved theoretically and then verified by several numerical examples. Even though our scheme does not preserve the divergence, but by implying the higher order polynomial, one can observe the same convergence rate as numerical solution. Several numerical tests are presented to validate these conclusions.}
 }
 
-@misc{schnake2024,
-      title={Sparse-grid Discontinuous {Galerkin} Methods for the {Vlasov-Poisson-Lenard-Bernstein} Model},
-      author={Stefan Schnake and Coleman Kendrick and Eirik Endeve and Miroslav Stoyanov and Steven Hahn and Cory D Hauck and David L Green and Phil Snyder and John Canik},
-      year={2024},
-      eprint={2402.06493},
-      archivePrefix={arXiv},
-      primaryClass={math.NA}
+@article{schnake2024,
+title = {Sparse-grid discontinuous {Galerkin} methods for the {Vlasov–Poisson–Lenard–Bernstein} model},
+journal = {Journal of Computational Physics},
+volume = {510},
+pages = {113053},
+year = {2024},
+issn = {0021-9991},
+doi = {https://doi.org/10.1016/j.jcp.2024.113053},
+url = {https://www.sciencedirect.com/science/article/pii/S0021999124003024},
+author = {Stefan Schnake and Coleman Kendrick and Eirik Endeve and Miroslav Stoyanov and Steven Hahn and Cory D. Hauck and David L. Green and Phil Snyder and John Canik},
+keywords = {Kinetic equation, Discontinuous Galerkin, Implicit-explicit, Sparse grids, Vlasov–Poisson, Lenard–Bernstein},
 }
 
 @article{huang2024,
@@ -112,6 +123,8 @@ abstract = { This paper reviews the adaptive sparse grid discontinuous Galerkin 
       year={2017},
       eprint={1710.09356},
[…]
