### 2024-08-07 58a4137
remove doi prefix
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
