+++ paper/paper.bib
@@ -99,7 +99,7 @@ volume = {510},
 pages = {113053},
 year = {2024},
 issn = {0021-9991},
-doi = {https://doi.org/10.1016/j.jcp.2024.113053},
+doi = {10.1016/j.jcp.2024.113053},
 url = {https://www.sciencedirect.com/science/article/pii/S0021999124003024},
 author = {Stefan Schnake and Coleman Kendrick and Eirik Endeve and Miroslav Stoyanov and Steven Hahn and Cory D. Hauck and David L. Green and Phil Snyder and John Canik},
 keywords = {Kinetic equation, Discontinuous Galerkin, Implicit-explicit, Sparse grids, Vlasov–Poisson, Lenard–Bernstein},
