### 2024-08-07 ad2c76a
plotting works for all degrees
