+++ src/asgard_reconstruct.cpp
@@ -2,6 +2,8 @@
 
 #include "asgard_wavelet_basis.hpp"
 
+#include "basis.hpp"
+
 namespace asgard
 {
 
@@ -50,6 +52,35 @@ reconstruct_solution::reconstruct_solution(
 
   // analyze the graph and prepare cache data
   build_tree();
+
+  // prepare the wavelet basis, if using general degree
+  if (degree >= 3)
+  {
+    // using coefficient matrices for the left/right wavelets
+    // and scale factors for the legendre basis
+    wavelets.resize(2 * pterms_ * pterms_ + pterms_);
+    wleft  = wavelets.data() + pterms_;
+    wright = wleft + pterms_ * pterms_;
+
+    // first pterms_ entries are the scale factors of the
+    // orthogonal Legendre polynomials on the interval (0, 1)
+    wavelets[0] = 1.0;
+    for (int i = 1; i < pterms_; i++)
+      wavelets[i] = std::sqrt(2.0 * i + 1.0);
+
+    // using ASGarD to build the matrices, but may be differen precision
+    using prec = default_precision;
+    std::array<fk::matrix<prec>, 6> matx = generate_multi_wavelets<prec>(degree);
+    fk::matrix<prec> const &c = matx[4]; // only need one matrix
+
+    // reversing the order, since we will be running from degree 0 to pterms_-1
+    auto w = wleft;
+    for (auto i : indexof<int>(pterms_))
+      w = std::copy_n(c.data(0, pterms_ - i - 1), pterms_, w);
+
+    for (auto i : indexof<int>(pterms_))
+      w = std::copy_n(c.data(pterms_, pterms_ - i - 1), pterms_, w);
+  }
 }
 
 void reconstruct_solution::set_domain_bounds(double const amin[], double const amax[])
@@ -94,8 +125,11 @@ void reconstruct_solution::reconstruct(double const x[], int num_x, double y[]) 
   case 1:
     reconstruct<1>(x, num_x, y);
     break;
+  case 2:
+    reconstruct<2>(x, num_x, y);
+    break;
   default:
-    throw std::runtime_error("incorrect degree");
+    reconstruct<-1>(x, num_x, y);
   }
 }
 
@@ -230,72 +264,214 @@ void reconstruct_solution::build_tree()
   std::copy_if(tree[0], tree[0] + tree.total_size(), indx.begin(), [](int t)->bool{ return (t > -1); });
 }
 
+template<int degree>
 std::optional<double>
-reconstruct_solution::basis_value0(int const p[], double const x[], double const c[]) const
+reconstruct_solution::basis_value(int const p[], double const x[],
+                                   double const c[], vector2d<double> &scratch) const
 {
+  // degree 0, 1, and 2 are hard-coded as analytic formulat
+  // degree -1 means that we are using general degree and should use pterms_
+  static_assert(degree >= -1 and degree <= 2,
+                "unsupported degree, use 0, 1, 2 for analytic expressions, "
+                "and degree -1 for the general case");
+
   // first translat the x values to normalized values for the given cell
   std::array<double, max_num_dimensions> xn;
-  std::array<double, max_num_dimensions> w;
+  double w = 1.0;
   int const &num_dimensions = cells_.num_dimensions();
-  for (int d = 0; d < num_dimensions; d++)
+  for (auto d : indexof<int>(num_dimensions))
   {
     int p2l2;
-    fm::intlog2_pow2pows2(p[d], p2l2, w[d]);
+    double wd;
+    fm::intlog2_pow2pows2(p[d], p2l2, wd);
+    w *= wd;
[…]
