### 2024-08-08 f5d059e
cleaner wavelet construction
