+++ src/coefficients.cpp
@@ -2,7 +2,8 @@
 
 #include "asgard_matrix.hpp"
 #include "asgard_vector.hpp"
-#include "basis.hpp"
+
+#include "asgard_basis.hpp"
 #include "fast_math.hpp"
 #include "matlab_utilities.hpp"
 #include "pde.hpp"
