+++ src/asgard_wavelet_basis.hpp
@@ -1,10 +1,14 @@
 #pragma once
 
-namespace asgard
-{
+#include "asgard_indexset.hpp"
 
+#include "quadrature.hpp"
+
+namespace asgard::basis
+{
+// hardcoded linear basis functions
 template<typename P>
-struct linear_basis
+struct linear
 {
   static P constexpr s3 = 1.73205080756887729; // sqrt(3.0)
   // projection basis
@@ -22,15 +26,15 @@ struct linear_basis
   static P iwav1L(P) { return 0; }
   static P iwav1R(P x) { return 6 * x - 4; }
 };
-
+// hardcoded quadratic basis functions
 template<typename P>
-struct quadratic_basis
+struct quadratic
 {
-  static P constexpr s3 = linear_basis<P>::s3; // sqrt(3.0)
+  static P constexpr s3 = linear<P>::s3; // sqrt(3.0)
   static P constexpr s5 = 2.2360679774997897;  // sqrt(5.0)
 
-  static constexpr auto pleg0 = linear_basis<P>::pleg0;
-  static constexpr auto pleg1 = linear_basis<P>::pleg1;
+  static constexpr auto pleg0 = linear<P>::pleg0;
+  static constexpr auto pleg1 = linear<P>::pleg1;
   static P pleg2(P x) { return s5 * (6.0 * x * x - 6.0 * x + 1.0); }
 
   static P pwav0L(P x) { return -(1.0 - 12.0 * x + 24.0 * x * x) * s5; }
@@ -42,4 +46,316 @@ struct quadratic_basis
   static P pwav2R(P x) { return 5.0 - 6.0 * x; }
 };
 
-} // namespace asgard
+// algorithms for general basis
+// this is a cleaner version of the generate_multi_wavelets method
+// in addition to the reduction of calls to new/delete the goal is
+// to have less code and easier to read compared to C++-2-MATLAB-isms
+// the inner product is still computed with a quadrature as the quadrature
+// appears to be numerically more stable then the analytic integration
+// of the individual monomials
+// the monomial approach simply leads to numbers close in absolute value
+// but with opposite signs
+
+// these methods do a lot of polynomial manipulations
+// here, a polynomial is defined by the monomial coefficients
+// p(x) = sum_i c_i x^i -> p(x) is represented as a vector (c_0, c_1 ... c_n)
+// addition/subtraction/scalar-mult applies the action to the vector/s
+
+//! Indicates whether to integrate left, right sub-interval or all of (-1, 1)
+enum class integ_range
+{
+  //! integrate over (-1,  0)
+  left,
+  //! integrate over ( 0, +1)
+  right,
+  //! integrate over (-1, +1)
+  full
+};
+
+/*!
+ * \brief Constructs the Legendre orthogonal polynomials
+ *
+ * The Legendre polynomials are represented as a coefficients for the
+ * monomial basis functions. The actual values of the polynomials
+ * can be computed from:
+ * \code
+ *   auto legendre = basis::legendre_poly(n);
+ *
+ *   // i-th Legendre polynomial of degree n at point x
+ *   // is computed as follow
+ *   P Li = 0.0;
+ *   P mono = 1.0;
+ *   for (int j = 0; j <= n; j++) {
+ *     Li += legendre[i][j] * mono;
+ *     mono *= x;
+ *   }
+ * \endcode
+ *
+ * The polynomials are rescaled to have unit norm over the specified range,
+ * e.g., (-1, 0), (0, 1) or (-1, 1).
+ */
+template<typename P, integ_range range = integ_range::full>
+vector2d<P> legendre_poly(int const degree)
+{
+  // the level 0 wavelets, in this case, the Legendre polynomials
[…]
