+++ src/asgard_basis.cpp
@@ -1,4 +1,4 @@
-#include "basis.hpp"
+#include "asgard_basis.hpp"
 #include "asgard_matrix.hpp"
 #include "asgard_vector.hpp"
 #include "distribution.hpp"
@@ -13,7 +13,7 @@ namespace asgard
 // used (outside of this routine) to construct the forward multi-wavelet
 // transform
 template<typename P>
-std::array<fk::matrix<P>, 6> generate_multi_wavelets(int const degree)
+std::array<fk::matrix<P>, 4> generate_multi_wavelets(int const degree)
 {
   expect(degree >= 0);
 
@@ -23,254 +23,77 @@ std::array<fk::matrix<P>, 6> generate_multi_wavelets(int const degree)
   // g0,g1,h0, and h1 are two-scale coefficients
   // The returned phi_co is the wavelet basis
   // scalet_coefficients are the scaling function basis
+  //   -- the scalet coefficients form the legendre basis
+  //      from a monomial basis
+
+  // hard-cording degree 0, 1, 2 (mostly for less rounding)
+  if (degree <= 2)
+  {
+    constexpr P s2 = 1.41421356237309505;
+
+    switch (degree)
+    {
+    case 0: {
+      P const is2 = 1 / s2;
+      fk::matrix<P> h0 = {{is2,},};
+      fk::matrix<P> h1 = {{is2,},};
+      fk::matrix<P> g0 = {{-is2,},};
+      fk::matrix<P> g1 = {{is2,},};
+      return {h0, h1, g0, g1};
+    }
+    case 1: {
+      P const is2  = 1 / s2;
+      P const is22 = 1 / (2 * s2);
+      P const is6  = std::sqrt(P{6}) / 4;
+      fk::matrix<P> h0 = {{is2, 0}, {-is6, is22}};
+      fk::matrix<P> h1 = {{is2, 0}, {is6, is22}};
+      fk::matrix<P> g0 = {{0, -is2}, {is22, is6}};
+      fk::matrix<P> g1 = {{0, is2}, {-is22, is6}};
+      return {h0, h1, g0, g1};
+    }
+    case 2: {
+      P const is2  = 1 / s2;
+      P const is22 = 1 / (2 * s2);
+      P const is24 = 1 / (4 * s2);
+      P const is6  = std::sqrt(P{6}) / 4;
+      P const is30 = 15 / (P{4} * std::sqrt(P{30}));
+      fk::matrix<P> h0 = {{is2, 0, 0}, {-is6, is22, 0}, {0, -is30, is24}};
+      fk::matrix<P> h1 = {{is2, 0, 0}, {is6, is22, 0}, {0, is30, is24}};
+      fk::matrix<P> g0 = {{0, 0, -is2}, {0, is24, is30}, {-is22, -is6, 0}};
+      fk::matrix<P> g1 = {{0, 0, is2}, {0, -is24, is30}, {is22, -is6, 0}};
+      return {h0, h1, g0, g1};
+    }
+    default:
+      break;
+    };
+  }
+
   fk::matrix<P> g0(pdof, pdof);
   fk::matrix<P> g1(pdof, pdof);
   fk::matrix<P> h0(pdof, pdof);
   fk::matrix<P> h1(pdof, pdof);
-  fk::matrix<P> scalet_coefficients(pdof, pdof);
 
-  // Set up parameters for quadrature - used for inner product evaluation
-  // In this routine there are three intervals for the quadrature
-  // Negative one to zero --zero to one -- and negative one to one
-  // since this is done on the "reference element" -1 to 1
-  // the intervals are fixed to be (-1,0), (0,1), and (-1,1)
-  P const neg1 = -1.0;
-  P const zero = 0.0;
[…]
