### 2024-08-09 #6766 @quantumsteve
> Thanks for merging the PR. Please check refs for the above including your JCP paper (https://doi.org/10.1016/j.jcp.2024.113053), then you may proceed to
>
> * Make a tagged release of your software, and list the version tag of the archived version here.
> * Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * Please list the DOI of the archived version here.
>
> I can then move forward with recommending acceptance of the submission.

https://doi.org/10.5281/zenodo.13287107
