### 2024-08-12 5b58daa
consistent names for the examples
