+++ examples/inputs_1d.py
@@ -13,7 +13,12 @@ import asgard
 # 3. The solution and the exact solution are plotted together
 
 if __name__ == '__main__':
-    if not os.path.isfile('example_inputs_1d'):
+    # build folder and install folder names can be different
+    if os.path.isfile('inputs_1d'):
+        exefilename = 'inputs_1d'
+    elif os.path.isfile('example_inputs_1d'):
+        exefilename = 'example_inputs_1d'
+    else:
         print("You must first build this project using CMake, e.g.,")
         print("  mkdir build")
         print("  cd build")
@@ -29,8 +34,8 @@ if __name__ == '__main__':
     outfile2 = 'waves2.h5'
 
     # providing the input file and also forcing the degree to 1
-    os.system("./example_inputs_1d -if inputs_1d_1.txt -d 2 -of %s" % outfile1)
-    os.system("./example_inputs_1d -if inputs_1d_2.txt -d 2 -of %s" % outfile2)
+    os.system("./{} -if inputs_1d_1.txt -d 2 -of {}".format(exefilename, outfile1))
+    os.system("./{} -if inputs_1d_2.txt -d 2 -of {}".format(exefilename, outfile2))
 
     if not os.path.isfile(outfile1) or not os.path.isfile(outfile2):
         print("ERROR: example_inputs_1d did not generate an output file")
