+++ testing/automated/clang-format-check.sh
@@ -100,8 +100,8 @@ elements.hpp
 #elements_tests.cpp
 fast_math.hpp
 #fast_math_tests.cpp
-io.hpp
-io_tests.cpp
+asgard_io.hpp
+asgard_io_tests.cpp
 #lib_dispatch.cpp
 lib_dispatch.hpp
 #lib_dispatch_tests.cpp
