### 2024-08-12 dcf1049
dictretization class
