+++ src/pde/pde_riemann_1x3v.hpp
@@ -142,31 +142,14 @@ private:
     return f;
   }
 
-  inline static moment<P> const moment0 = // (f,1)_v
-      moment<P>(std::vector<md_func_type<P>>(
-          {{moment0_f1, moment0_f1, moment0_f1, moment0_f1, moment0_f1}}));
-  inline static moment<P> const moment1v1 = // (f,v_x)_v
-      moment<P>(std::vector<md_func_type<P>>(
-          {{moment0_f1, moment1_f1, moment0_f1, moment0_f1, moment0_f1}}));
-  inline static moment<P> const moment1v2 = // (f,v_y)_v
-      moment<P>(std::vector<md_func_type<P>>(
-          {{moment0_f1, moment0_f1, moment1_f1, moment0_f1, moment0_f1}}));
-  inline static moment<P> const moment1v3 = // (f,v_z)_v
-      moment<P>(std::vector<md_func_type<P>>(
-          {{moment0_f1, moment0_f1, moment0_f1, moment1_f1, moment0_f1}}));
-  inline static moment<P> const moment2v1 = // (f,v_x^2)_v
-      moment<P>(std::vector<md_func_type<P>>(
-          {{moment0_f1, moment2_f1, moment0_f1, moment0_f1, moment0_f1}}));
-  inline static moment<P> const moment2v2 = // (f,v_y^2)_v
-      moment<P>(std::vector<md_func_type<P>>(
-          {{moment0_f1, moment0_f1, moment2_f1, moment0_f1, moment0_f1}}));
-  inline static moment<P> const moment2v3 = // (f,v_z^2)_v
-      moment<P>(std::vector<md_func_type<P>>(
-          {{moment0_f1, moment0_f1, moment0_f1, moment2_f1, moment0_f1}}));
-
-  inline static std::vector<moment<P>> const moments_ = {
-      moment0,   moment1v1, moment1v2, moment1v3,
-      moment2v1, moment2v2, moment2v3};
+  inline static moment_funcs<P> const moments_ = {
+      {{moment0_f1, moment0_f1, moment0_f1, moment0_f1, moment0_f1}},
+      {{moment0_f1, moment1_f1, moment0_f1, moment0_f1, moment0_f1}},
+      {{moment0_f1, moment0_f1, moment1_f1, moment0_f1, moment0_f1}},
+      {{moment0_f1, moment0_f1, moment0_f1, moment1_f1, moment0_f1}},
+      {{moment0_f1, moment2_f1, moment0_f1, moment0_f1, moment0_f1}},
+      {{moment0_f1, moment0_f1, moment2_f1, moment0_f1, moment0_f1}},
+      {{moment0_f1, moment0_f1, moment0_f1, moment2_f1, moment0_f1}}};
 
   /* Construct (n, u, theta) */
   static P n(P const &x, P const t = 0)
