+++ src/asgard.hpp
@@ -1,29 +1,18 @@
 #pragma once
 
-#include "coefficients.hpp"
-
-#ifdef ASGARD_IO_HIGHFIVE
-#include "asgard_io.hpp"
-#endif
-
-#ifdef ASGARD_USE_MPI
-#include <mpi.h>
-#endif
-
-#ifdef ASGARD_USE_MATLAB
-#include "matlab_plot.hpp"
-#endif
-
-#include "time_advance.hpp"
+#include "asgard_time_advance.hpp"
 
 namespace asgard
 {
-
+/*!
+ * \brief Simulates one of the builtin PDEs
+ *
+ * The options.pde_choice must contain a valid PDE_opts that is not
+ * PDE_opts::custom.
+ * This method is equivalent to calling the asgard executable.
+ */
 template<typename precision>
-void simulate(std::unique_ptr<PDE<precision>> &pde);
-
-template<typename precision>
-void simulate(prog_opts const &options)
+void simulate_builtin(prog_opts const &options)
 {
   if (options.show_help)
   {
@@ -47,10 +36,77 @@ void simulate(prog_opts const &options)
   rassert(options.pde_choice and options.pde_choice.value() != PDE_opts::custom,
           "must provide a valid PDE choice");
 
-  auto pde = make_PDE<precision>(options);
-  simulate(pde);
+  discretization_manager discretization(make_PDE<precision>(options),
+                                        verbosity_level::high);
+
+  advance_time(discretization);
+
+  discretization.save_final_snapshot();
+
+  // collects the state across MPI ranks but ignores the result
+  discretization.current_mpistate();
+
+  node_out() << tools::timer.report() << '\n';
 }
 
+/*!
+ * \brief Creates a discretization for the given pde and options
+ *
+ * Creates a new PDE discretization, unlike the one-shot method asgard::simulate(),
+ * the discretization allows for finer control of the time steeping and
+ * output frequency.
+ *
+ * \tparam pde_class is a user provided PDE specification that is derived from
+ *                   the base asgard::PDE class
+ *
+ * \param options is a set of options, either provided from the command line,
+ *                a file or manually set by the user
+ *
+ * \param verbosity indicates how much output ASGarD should provide,
+ *                  verbosity_level::high is useful for development and debugging
+ *                  while quiet is suitable for long simulations
+ *
+ * \returns an instance of asgard::discretization_manager
+ */
+template<typename pde_class>
+auto discretize(prog_opts const &options, verbosity_level verbosity = verbosity_level::quiet)
+{
+  static_assert(std::is_base_of_v<PDE<float>, pde_class> or std::is_base_of_v<PDE<double>, pde_class>,
+                "the requested PDE class must inherit from the asgard::PDE base-class");
+
+  return discretization_manager(make_custom_pde<pde_class>(options), verbosity);
+}
+
+/*!
+ * \brief One shot method, simulate the pde with the given options
+ *
+ * Creates a new PDE, discretizes it and advances in time according to the
+ * options specified by the prog_opts instance and the PDE specification.
+ *
+ * \tparam pde_class is a user provided PDE specification that is derived from
+ *                   the base asgard::PDE class
+ *
+ * \param options is a set of options, either provided from the command line,
+ *                a file or manually set by the user
+ *
+ * \param verbosity indicates how much output ASGarD should provide,
+ *                  verbosity_level::high is useful for development and debugging
+ *                  while quiet is suitable for long simulations
+ *
+ * Note: options.pde_choice must be either empty (does not contain a value)
+ *       or set to PDE_opts::custom
+ */
+template<typename pde_class>
+void simulate(prog_opts const &options, verbosity_level verbosity = verbosity_level::quiet)
+{
[…]
