+++ src/main.cpp
@@ -17,7 +17,7 @@ int main(int argc, char **argv)
   // custom projects may implement their own inputs
   //   in addition to ASGarD standard ones
   // however, this the main executable and all
-  //   unknonw cli commands are to be treated as errors
+  //   unknown cli commands are to be treated as errors
   // first we generate warnigns for the user
   bool constexpr ignore_unknown = false;
   // -- parse cli
@@ -25,10 +25,10 @@ int main(int argc, char **argv)
 
   // if there were unknown options, throw an error
   if (not cli_input.externals.empty())
-    throw std::runtime_error("encountered unrecognized command line option");
+    throw std::runtime_error("encountered unrecognized command line option(s)");
 
   // main call to asgard, does all the work
-  asgard::simulate<precision>(cli_input);
+  asgard::simulate_builtin<precision>(cli_input);
 
   asgard::finalize_distribution();
 
