+++ src/pde/pde_vlasov_lb_full_f.hpp
@@ -129,7 +129,6 @@ private:
       {{moment0_f1, moment1_f1, moment0_f1}},
       {{moment0_f1, moment2_f1, moment0_f1}}};
 
-
   /* Construct (n, u, theta) */
   static P n(P const &x, P const t = 0)
   {
