+++ src/matlab_utilities.hpp
@@ -1,13 +1,5 @@
 #pragma once
-
-#include "asgard_matrix.hpp"
-#include "asgard_vector.hpp"
 #include "fast_math.hpp"
-#include <algorithm>
-#include <filesystem>
-#include <functional>
-#include <string>
-#include <vector>
 
 namespace asgard
 {
