+++ src/asgard_dimension.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "asgard_matrix.hpp"
-#include "asgard_vector.hpp"
 #include "fast_math.hpp"
 #include "matlab_utilities.hpp"
 #include "program_options.hpp"
