+++ src/asgard_kronmult_matrix_tests.cpp
@@ -1,3 +1,5 @@
+#include "tests_general.hpp"
+
 #include "asgard_kronmult_tests.hpp"
 
 template<typename T>
