+++ src/asgard_adapt.cpp
@@ -1,4 +1,4 @@
-#include "adapt.hpp"
+#include "asgard_adapt.hpp"
 #include "distribution.hpp"
 #include "transformations.hpp"
 
