+++ src/asgard_batch.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "adapt.hpp"
+#include "asgard_adapt.hpp"
 
 namespace asgard
 {
