+++ src/asgard_distribution.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "asgard_mpi.hpp"
-#include "elements.hpp"
+#include "asgard_elements.hpp"
 
 #ifdef ASGARD_USE_SCALAPACK
 #include "cblacs_grid.hpp"
