+++ src/asgard_discretization.hpp
@@ -2,8 +2,8 @@
 #include "asgard_reconstruct.hpp"
 #include "asgard_boundary_conditions.hpp"
 #include "asgard_coefficients.hpp"
-#include "moment.hpp"
-#include "solver.hpp"
+#include "asgard_moment.hpp"
+#include "asgard_solver.hpp"
 
 #ifdef ASGARD_IO_HIGHFIVE
 #include "asgard_io.hpp"
