+++ src/asgard_elements.cpp
@@ -2,8 +2,8 @@
 
 #include "asgard_matrix.hpp"
 #include "asgard_vector.hpp"
-#include "fast_math.hpp"
-#include "matlab_utilities.hpp"
+#include "asgard_fast_math.hpp"
+#include "asgard_matlab_utilities.hpp"
 #include "program_options.hpp"
 
 namespace asgard::elements
