+++ src/asgard_dimension.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "fast_math.hpp"
-#include "matlab_utilities.hpp"
+#include "asgard_fast_math.hpp"
+#include "asgard_matlab_utilities.hpp"
 #include "program_options.hpp"
 
 namespace asgard
