+++ src/quadrature.cpp
@@ -1,5 +1,5 @@
 #include "quadrature.hpp"
-#include "matlab_utilities.hpp"
+#include "asgard_matlab_utilities.hpp"
 
 namespace asgard
 {
