+++ src/permutations.cpp
@@ -1,6 +1,6 @@
 #include "permutations.hpp"
 
-#include "matlab_utilities.hpp"
+#include "asgard_matlab_utilities.hpp"
 
 namespace asgard::permutations
 {
