+++ src/asgard_lib_dispatch.cpp
@@ -1,4 +1,4 @@
-#include "lib_dispatch.hpp"
+#include "asgard_lib_dispatch.hpp"
 #include "asgard_distribution.hpp" // seems needed for MPI
 #include "sparse.hpp"
 
