+++ src/asgard_matlab_utilities.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "fast_math.hpp"
+#include "asgard_fast_math.hpp"
 
 namespace asgard
 {
