+++ src/asgard_elements.cpp
@@ -4,7 +4,7 @@
 #include "asgard_vector.hpp"
 #include "asgard_fast_math.hpp"
 #include "asgard_matlab_utilities.hpp"
-#include "program_options.hpp"
+#include "asgard_program_options.hpp"
 
 namespace asgard::elements
 {
