+++ src/asgard_dimension.hpp
@@ -2,7 +2,7 @@
 
 #include "asgard_fast_math.hpp"
 #include "asgard_matlab_utilities.hpp"
-#include "program_options.hpp"
+#include "asgard_program_options.hpp"
 
 namespace asgard
 {
