+++ src/asgard_lib_dispatch.cpp
@@ -1,6 +1,6 @@
 #include "asgard_lib_dispatch.hpp"
 #include "asgard_distribution.hpp" // seems needed for MPI
-#include "sparse.hpp"
+#include "asgard_sparse.hpp"
 
 // ==========================================================================
 // external declarations for calling blas routines linked with -lblas
