+++ src/asgard_permutations.cpp
@@ -1,4 +1,4 @@
-#include "permutations.hpp"
+#include "asgard_permutations.hpp"
 
 #include "asgard_matlab_utilities.hpp"
 
