+++ src/asgard_adapt.cpp
@@ -1,8 +1,6 @@
 #include "asgard_adapt.hpp"
 #include "asgard_distribution.hpp"
-#include "transformations.hpp"
-
-#include <unordered_set>
+#include "asgard_transformations.hpp"
 
 namespace asgard::adapt
 {
