+++ src/asgard_scalapack_matrix_info.cpp
@@ -1,5 +1,5 @@
-#include "scalapack_matrix_info.hpp"
-#include "cblacs_grid.hpp"
+#include "asgard_scalapack_matrix_info.hpp"
+#include "asgard_cblacs_grid.hpp"
 #include <stdexcept>
 
 extern "C"
