+++ src/asgard_distribution_tests.cpp
@@ -1,8 +1,8 @@
 #include "tests_general.hpp"
 
 #ifdef ASGARD_USE_SCALAPACK
-#include "cblacs_grid.hpp"
-#include "scalapack_vector_info.hpp"
+#include "asgard_cblacs_grid.hpp"
+#include "asgard_scalapack_vector_info.hpp"
 #endif
 
 struct distribution_test_init
