+++ src/asgard_pde.hpp
@@ -1,27 +1,27 @@
 #pragma once
-#include "pde/pde_advection1.hpp"
-#include "pde/pde_collisional_landau.hpp"
-#include "pde/pde_collisional_landau_1x2v.hpp"
-#include "pde/pde_collisional_landau_1x3v.hpp"
-#include "pde/pde_continuity1.hpp"
-#include "pde/pde_continuity2.hpp"
-#include "pde/pde_continuity3.hpp"
-#include "pde/pde_continuity6.hpp"
-#include "pde/pde_diffusion1.hpp"
-#include "pde/pde_diffusion2.hpp"
-#include "pde/pde_fokkerplanck1_4p3.hpp"
-#include "pde/pde_fokkerplanck1_4p4.hpp"
-#include "pde/pde_fokkerplanck1_4p5.hpp"
-#include "pde/pde_fokkerplanck1_pitch_C.hpp"
-#include "pde/pde_fokkerplanck1_pitch_E.hpp"
-#include "pde/pde_fokkerplanck2_complete.hpp"
-#include "pde/pde_relaxation_1x1v.hpp"
-#include "pde/pde_relaxation_1x2v.hpp"
-#include "pde/pde_relaxation_1x3v.hpp"
-#include "pde/pde_riemann_1x2v.hpp"
-#include "pde/pde_riemann_1x3v.hpp"
-#include "pde/pde_two_stream.hpp"
-#include "pde/pde_vlasov_lb_full_f.hpp"
+#include "pde/asgard_pde_advection1.hpp"
+#include "pde/asgard_pde_collisional_landau.hpp"
+#include "pde/asgard_pde_collisional_landau_1x2v.hpp"
+#include "pde/asgard_pde_collisional_landau_1x3v.hpp"
+#include "pde/asgard_pde_continuity1.hpp"
+#include "pde/asgard_pde_continuity2.hpp"
+#include "pde/asgard_pde_continuity3.hpp"
+#include "pde/asgard_pde_continuity6.hpp"
+#include "pde/asgard_pde_diffusion1.hpp"
+#include "pde/asgard_pde_diffusion2.hpp"
+#include "pde/asgard_pde_fokkerplanck1_4p3.hpp"
+#include "pde/asgard_pde_fokkerplanck1_4p4.hpp"
+#include "pde/asgard_pde_fokkerplanck1_4p5.hpp"
+#include "pde/asgard_pde_fokkerplanck1_pitch_C.hpp"
+#include "pde/asgard_pde_fokkerplanck1_pitch_E.hpp"
+#include "pde/asgard_pde_fokkerplanck2_complete.hpp"
+#include "pde/asgard_pde_relaxation_1x1v.hpp"
+#include "pde/asgard_pde_relaxation_1x2v.hpp"
+#include "pde/asgard_pde_relaxation_1x3v.hpp"
+#include "pde/asgard_pde_riemann_1x2v.hpp"
+#include "pde/asgard_pde_riemann_1x3v.hpp"
+#include "pde/asgard_pde_two_stream.hpp"
+#include "pde/asgard_pde_vlasov_lb_full_f.hpp"
 
 namespace asgard
 {
