+++ testing/automated/clang-format-check.sh
@@ -23,33 +23,33 @@ device/asgard_spkronmult_cpu_general.hpp
 device/asgard_spkronmult_cycle1.hpp
 device/asgard_spkronmult_cycle2.hpp
 device/asgard_spkronmult_cyclex.hpp
-#pde/pde_advection1.hpp
-#pde/pde_base.hpp
-pde/pde_collisional_landau_1x2v.hpp
-#pde/pde_collisional_landau_1x3v.hpp
-pde/pde_collisional_landau.hpp
-pde/pde_continuity1.hpp
-#pde/pde_continuity2.hpp
-#pde/pde_continuity3.hpp
-#pde/pde_continuity6.hpp
-#pde/pde_diffusion1.hpp
-pde/pde_diffusion2.hpp
-pde/pde_fokkerplanck1_4p3.hpp
-pde/pde_fokkerplanck1_4p4.hpp
-pde/pde_fokkerplanck1_4p5.hpp
-#pde/pde_fokkerplanck1_pitch_C.hpp
-pde/pde_fokkerplanck1_pitch_E.hpp
-pde/pde_fokkerplanck2_complete.hpp
-pde/pde_relaxation_1x1v.hpp
-pde/pde_relaxation_1x2v.hpp
-#pde/pde_relaxation_1x3v.hpp
-pde/pde_riemann_1x2v.hpp
-#pde/pde_riemann_1x3v.hpp
-pde/pde_two_stream.hpp
-pde/pde_vlasov_lb_full_f.hpp
-#adapt.cpp
-adapt.hpp
-adapt_tests.cpp
+#pde/asgard_pde_advection1.hpp
+#pde/asgard_pde_base.hpp
+pde/asgard_pde_collisional_landau_1x2v.hpp
+#pde/asgard_pde_collisional_landau_1x3v.hpp
+pde/asgard_pde_collisional_landau.hpp
+pde/asgard_pde_continuity1.hpp
+#pde/asgard_pde_continuity2.hpp
+#pde/asgard_pde_continuity3.hpp
+#pde/asgard_pde_continuity6.hpp
+#pde/asgard_pde_diffusion1.hpp
+pde/asgard_pde_diffusion2.hpp
+pde/asgard_pde_fokkerplanck1_4p3.hpp
+pde/asgard_pde_fokkerplanck1_4p4.hpp
+pde/asgard_pde_fokkerplanck1_4p5.hpp
+#pde/asgard_pde_fokkerplanck1_pitch_C.hpp
+pde/asgard_pde_fokkerplanck1_pitch_E.hpp
+pde/asgard_pde_fokkerplanck2_complete.hpp
+pde/asgard_pde_relaxation_1x1v.hpp
+pde/asgard_pde_relaxation_1x2v.hpp
+#pde/asgard_pde_relaxation_1x3v.hpp
+pde/asgard_pde_riemann_1x2v.hpp
+#pde/asgard_pde_riemann_1x3v.hpp
+pde/asgard_pde_two_stream.hpp
+pde/asgard_pde_vlasov_lb_full_f.hpp
+#asgard_adapt.cpp
+asgard_adapt.hpp
+asgard_adapt_tests.cpp
 asgard_grid_1d.hpp
 asgard_indexset.cpp
 asgard_indexset.hpp
@@ -81,69 +81,69 @@ asgard_tools.cpp
 asgard_tools.hpp
 asgard_tools_tests.cpp
 asgard_vector.hpp
-#basis.cpp
+#asgard_basis.cpp
 asgard_basis.hpp
-#basis_tests.cpp
-batch.cpp
-#batch.hpp
-#batch_tests.cpp
-boundary_conditions.cpp
-boundary_conditions.hpp
[…]
