+++ src/asgard_matrix.hpp
@@ -1,5 +1,4 @@
 #pragma once
-#include "asgard_resources.hpp"
 #include "asgard_lib_dispatch.hpp"
 
 namespace asgard
