+++ src/asgard_program_options.hpp
@@ -1,18 +1,64 @@
 #pragma once
 #include "asgard_tools.hpp"
 
+/*!
+ * \file asgard_program_options.hpp
+ * \brief Defines common enums and the options manager class prog_opts
+ * \author The ASGarD Team
+ * \ingroup asgard_common_options
+ */
+
 namespace asgard
 {
-// implemented solvers for implicit stepping
+/*!
+ * \defgroup asgard_common_options ASGarD Common Options
+ *
+ * Common options shared by most or even all PDEs and discretization methods.
+ * The tools provided here allow for reading the options from either the
+ * command line or an input file and also specify PDE specific options.
+ */
+
+/*!
+ * \ingroup asgard_common_options
+ * \brief Allows reducing the amount of cout-noise
+ *
+ * The high noise is usually desired for large simulations as the cout stream
+ * will become a log for the various aspects of the problem.
+ * This is very useful for debugging, catching early problems and keeping
+ * an eye on a long simulation.
+ *
+ * However, high noise is bad for testing and potentially some large
+ * applications, e.g., high verbosity may drown important messages from other
+ * sub-systems.
+ */
+enum class verbosity_level
+{
+  //! do not generate cout output, except on errors and important warnings
+  quiet,
+  //! provide a detailed log of the various aspects of the simulation
+  high
+};
+
+/*!
+ * \ingroup asgard_common_options
+ * \brief the available solvers for implicit time stepping
+ */
 enum class solve_opts
 {
+  //! direct solve using LAPACK, slow but stable
   direct,
+  //! popular iterative solver, can be sensitive to the tolerance and restart frequency
   gmres,
+  //! alternative to gmres, cheaper when taking many steps between restarts
   bicgstab,
+  //! distributed direct solver, MPI-powered alternative to LAPACK
   scalapack
 };
 
-// the choices for supported PDE types
+/*!
+ * \ingroup asgard_common_options
+ * \brief list of builtin PDE specifications, refer to the specs in the src/pde folder
+ */
 enum class PDE_opts
 {
   custom = 0, // user provided pde
@@ -45,7 +91,13 @@ enum class PDE_opts
   collisional_landau_1x3v
 };
 
-// Not very informative, maybe rename these of handle in a different wat
+/*!
+ * \internal
+ * \ingroup asgard_common_options
+ * \brief some PDE options allow for variants
+ *
+ * \endinternal
+ */
 enum class PDE_case_opts
 {
   case0,
@@ -54,10 +106,10 @@ enum class PDE_case_opts
   case3,
   case4,
   case_count
-  // FIXME will need to add the user supplied PDE cases choice
 };
 
 /*!
+ * \ingroup asgard_common_options
  * \brief Indicates whether we should be using sparse or dense kronmult.
  *
  * Used by the local kronmult, global kronmult (block or non-block case)
@@ -72,6 +124,7 @@ enum class kronmult_mode
 };
 
 /*!
+ * \ingroup asgard_common_options
  * \brief Type of discretization grid
  */
 enum class grid_type
@@ -86,26 +139,44 @@ enum class grid_type
 
 namespace time_advance
 {
-//! types of time time advance methods, declared here to be used in the program options
+/*!
+ * \ingroup asgard_common_options
+ * types of time time advance methods, declared here to be used in the program options
+ */
 enum class method
 {
+  //! implicit solve, backward Euler
   imp,
+  //! (default) explicit Runge–Kutta
   exp, // explicit is reserved keyword
+  //! implicit-explicit scheme for nonlinear Vlasov-Poisson problems
   imex
 };
 } // namespace time_advance
 
+/*!
[…]
