+++ src/asgard_reconstruct.hpp
@@ -2,17 +2,46 @@
 
 #include "asgard_interpolation.hpp"
 
+/*!
+ * \file asgard_reconstruct.hpp
+ * \brief Provides the class for reconstructing solution at arbitrary points in the domain.
+ * \author The ASGarD Team
+ * \ingroup asgard_postprocess
+ */
+
 namespace asgard
 {
 
 /*!
- * \brief Reconstruct the solution at arbitrary points in the domain from wavelet data
+ * \defgroup asgard_postprocess ASGarD Post Processing Tools
  *
- * The class is inteded to be used from within Python for plotting
+ * The internal data-structures of ASGarD store the solution state in hierarchical
+ * format, i.e., vector of coefficients associated with sparse grid points.
+ * Interpretation of the data in raw format is challenging at best but converting
+ * the data into a more standard dense-grid format is also not feasible in
+ * high dimensions.
+ *
+ * The asgard::reconstruct_solution can find the value of the solution at
+ * arbitrary points of the domain but computing the sum of coefficients times
+ * the values of the basis functions. The class also comes with bindings for
+ * Python and MATLAB. See the included examples.
+ *
+ * \internal
+ * The C bindings allow Python to hold onto a reconstruct_solution object via a
+ * void pointer. On every call, reinterpret_cast is used on the pointer to match
+ * the correct type and a call to the corresponding member method is made.
+ * \endinternal
+ */
+
+/*!
+ * \ingroup asgard_postprocess
+ * \brief Reconstruct the solution at arbitrary points in the domain from hierarchical data
+ *
+ * The class is intended to be used from within Python for plotting
  * and post-processing of the data. The API uses raw-arrays to facilitate
  * interfacing with the Python ctypes module.
  *
- * The reconstruction algorihms are taken from Tasmanian.
+ * The reconstruction algorithms are taken from Tasmanian.
  */
 class reconstruct_solution
 {
@@ -55,7 +84,18 @@ public:
    */
   void reconstruct(double const x[], int num_x, double y[]) const;
 
+  //! Returns the number of dimensions
+  int num_dimensions() const { return cells_.num_dimensions(); }
+  //! Returns the number of SG cells
+  int64_t num_cells() const { return cells_.num_indexes(); }
+
+  /*!
+   * \brief Returns the centers of the sparse grid cells
+   */
+  void cell_centers(double x[]) const;
+
 protected:
+#ifndef __ASGARD_DOXYGEN_SKIP_INTERNAL
   //! template dispatch based on the degree
   template<int degree>
   void reconstruct(double const x[], int num_x, double y[]) const;
@@ -68,21 +108,22 @@ protected:
    * The process is expensive and becomes less viable after level 10.
    *
    * The alternative is to find analytic expressions for the wavelets and avoid
-   * going to "real" or "non-wavelet" space. This leads to limitations on
-   * the polynomial degree, i.e., wavelets have to be hard-coded (for now).
-   *
-   * Furthermore, since the basis is local, many of the functions can be
+   * going to "real" or "non-wavelet" space
+   * and because the basis is local, many of the functions can be
    * skipped in the reconstruction. The sparse grids cells natually form
-   * a directed-asyclic-graph (DAG) and we can use the fact that the support
+   * a directed-acyclic-graph (DAG) and we can use the fact that the support
    * of the children is included in the support of the parent, so if the parent
    * is not supported over a given point in the domain, all children are
-   * also non-supported and can be skipepd.
+   * also non-supported and can be skipepd in the evaluation.
    *
-   * On the other hand, since the DAG allows for multiple parents, it is likely
-   * to double-count the supported functions. This method repalces the DAG with
[…]
