+++ src/asgard_distribution.hpp
@@ -233,6 +233,10 @@ void reduce_results(fk::vector<P, src_mem, resrc> const &source,
                     fk::vector<P, dst_mem, resrc> &dest,
                     distribution_plan const &plan, int const my_rank);
 
+template<typename P>
+void reduce_results(std::vector<P> const &source, std::vector<P> &dest,
+                    distribution_plan const &plan, int const my_rank);
+
 // generate a message list for each rank for exchange_results function;
 // conceptually an internal component function, exposed for testing
 std::vector<std::vector<message>> const
@@ -242,6 +246,10 @@ generate_messages(distribution_plan const &plan);
 template<typename P, mem_type src_mem, mem_type dst_mem, resource resrc>
 void exchange_results(fk::vector<P, src_mem, resrc> const &source,
                       fk::vector<P, dst_mem, resrc> &dest,
+                      int const segment_size, distribution_plan const &plan,
+                      int const my_rank);
+template<typename P>
+void exchange_results(std::vector<P> const &source, std::vector<P> &dest,
                       int const segment_size, distribution_plan const &plan,
                       int const my_rank);
 
