+++ src/asgard_time_advance.cpp
@@ -8,6 +8,42 @@
 namespace asgard::time_advance
 {
 template<typename P>
+fk::vector<P>
+rungekutta3(discretization_manager<P> const &dist, std::vector<P> const &current)
+{
+  P const dt = dist.dt();
+
+  // 3 fight-hand-sides and the intermediate step
+  // the assumption is that the time-stepping scheme does not change much
+  // thus it makes sense to make these static and avoid repeated allocation
+  static std::vector<P> k1, k2, k3, s1;
+
+  k1.resize(current.size());
+  k2.resize(current.size());
+  k3.resize(current.size());
+  s1.resize(current.size());
+
+  dist.ode_rhs(imex_flag::unspecified, dist.time(), current, k1);
+
+  for (auto i : indexof(s1))
+    s1[i] = current[i] + 0.5 * dt * k1[i];
+
+  dist.ode_rhs(imex_flag::unspecified, dist.time() + 0.5 * dt, s1, k2);
+
+  for (auto i : indexof(s1))
+    s1[i] = current[i] - dt * k1[i] + 2 * dt * k2[i];
+
+  dist.ode_rhs(imex_flag::unspecified, dist.time() + dt, s1, k3);
+
+  fk::vector<P> r(current.size());
+  for (auto i : indexof(r))
+    r[i] = current[i] + dt * (k1[i] + 4 * k2[i] + k3[i]) / P{6};
+
+  return r;
+}
+
+#ifdef ASGARD_USE_SCALAPACK
+template<typename P>
 static fk::vector<P>
 get_sources(PDE<P> const &pde, adapt::distributed_grid<P> const &grid,
             basis::wavelet_transform<P, resource::host> const &transformer,
@@ -28,144 +64,6 @@ get_sources(PDE<P> const &pde, adapt::distributed_grid<P> const &grid,
   }
   return sources;
 }
-
-// this function executes an explicit time step using the current solution
-// vector x. on exit, the next solution vector is stored in x.
-template<typename P>
-fk::vector<P>
-explicit_advance(PDE<P> const &pde, kron_operators<P> &operator_matrices,
-                 adapt::distributed_grid<P> const &adaptive_grid,
-                 basis::wavelet_transform<P, resource::host> const &transformer,
-                 std::array<boundary_conditions::unscaled_bc_parts<P>, 2> const
-                     &unscaled_parts,
-                 fk::vector<P> const &x_orig, P const time)
-{
-  auto const &plan     = adaptive_grid.get_distrib_plan();
-  auto const &grid     = adaptive_grid.get_subgrid(get_rank());
-  auto const elem_size = element_segment_size(pde);
-  auto const dt        = pde.get_dt();
-  auto const col_size  = elem_size * static_cast<int64_t>(grid.ncols());
-  expect(x_orig.size() == col_size);
-  auto const row_size = elem_size * static_cast<int64_t>(grid.nrows());
-  expect(col_size < INT_MAX);
-  expect(row_size < INT_MAX);
-
-  operator_matrices.make(imex_flag::unspecified, pde, adaptive_grid);
-
-  // time advance working vectors
-  // input vector for apply_A
-  fk::vector<P> x(x_orig);
-  // a buffer for reducing across subgrid row
-  fk::vector<P> reduced_fx(row_size);
-
-  expect(time >= 0);
-  expect(dt > 0);
-
-  // see
-  // https://en.wikipedia.org/wiki/Runge%E2%80%93Kutta_methods#Explicit_Runge%E2%80%93Kutta_methods
-  P const a21 = 0.5;
-  P const a31 = -1.0;
[…]
