+++ src/asgard_matlab_plot.cpp
@@ -1,839 +1,0 @@
-#include "asgard_matlab_plot.hpp"
-
-#include "asgard_dimension.hpp"
-#include "quadrature.hpp"
-#include <cmath>
-#include <type_traits>
-namespace asgard::ml
-{
-matlab_plot &matlab_plot::get_instance()
-{
-  static matlab_plot instance;
-  return instance;
-}
-
-void matlab_plot::connect_async()
-{
-  if (!is_open())
-  {
-    matlab::engine::FutureResult<std::unique_ptr<matlab::engine::MATLABEngine>>
-        ml_connection;
-    ml_connection = matlab::engine::connectMATLABAsync();
-
-    this->matlab_inst_ = ml_connection.get();
-  }
-}
-
-void matlab_plot::connect(std::string const &name)
-{
-  // Connects to a shared MATLAB session with given name, or first available
-  // if no name is given. Creates a session if none exist.
-  if (!is_open())
-  {
-    if (name.empty() || name.compare("none") == 0)
-    {
-      // add message since this could take awhile if a new MATLAB session
-      // needs to be started
-      std::cout << "  connecting with MATLAB...\n";
-      auto const &sessions = matlab::engine::findMATLAB();
-      if (sessions.empty())
-      {
-        // if there were no shared sessions, set flag to wait for any plots at
-        // the end
-        started_ = true;
-      }
-      this->matlab_inst_ = matlab::engine::connectMATLAB();
-    }
-    else
-    {
-      bool get_name            = false;
-      std::string session_name = name;
-      std::cout << "  connecting with MATLAB session '" << session_name
-                << "'\n";
-      // Check if the named session exists
-      if (!find_session(name, false, session_name))
-      {
-        std::cerr << "  Specified MATLAB session '" << name
-                  << "' does not exist" << '\n';
-        // Try to find next session instead of exiting
-        get_name = true;
-      }
-      // If not found, try to connect to first running session
-      if (get_name)
-      {
-        bool found = find_session(name, true, session_name);
-        if (found)
-        {
-          std::cout << "  Found next running MATLAB session '" << session_name
-                    << "'\n";
-        }
-        else
-        {
-          std::cerr << "  Could not find another running MATLAB session.. "
-                       "creating a new one\n";
-          // Make a new shared session in this case
-          this->matlab_inst_ = matlab::engine::connectMATLAB();
-          started_           = true;
-        }
-      }
-      // convert to utf16 for matlab
-      ml_string const name_conv = matlab::engine::convertUTF8StringToUTF16String(session_name);
-      this->matlab_inst_ = matlab::engine::connectMATLAB(name_conv);
-    }
-  }
-  // Add the matlab scripts directory to the matlab path
-  this->add_param(std::string(ASGARD_SCRIPTS_DIR) + "matlab/");
-  this->call("addpath");
-}
-
-void matlab_plot::start(std::vector<ml_string> const &args)
-{
-  if (!is_open())
-  {
-    this->matlab_inst_ = matlab::engine::startMATLAB(args);
-    started_           = true;
-    // Add the matlab scripts directory to the matlab path
-    this->add_param(std::string(ASGARD_SCRIPTS_DIR) + "matlab/");
-    this->call("addpath");
-  }
-}
-
-void matlab_plot::close()
-{
-  if (is_open())
-  {
-    if (started_)
-    {
[…]
