+++ src/asgard_adapt_tests.cpp
@@ -1,223 +1,0 @@
-
-#include "tests_general.hpp"
-
-static auto constexpr adapt_threshold = 1e-4;
-
-static auto const adapt_base_dir = gold_base_dir / "adapt";
-
-using namespace asgard;
-
-struct distribution_test_init
-{
-  void set_my_rank(const int rank) { my_rank = rank; }
-  void set_num_ranks(const int size) { num_ranks = size; }
-  int get_my_rank() const { return my_rank; }
-  int get_num_ranks() const { return num_ranks; }
-
-private:
-  int my_rank;
-  int num_ranks;
-};
-static distribution_test_init distrib_test_info;
-
-int main(int argc, char *argv[])
-{
-  auto const [rank, total_ranks] = initialize_distribution();
-  distrib_test_info.set_my_rank(rank);
-  distrib_test_info.set_num_ranks(total_ranks);
-
-  int result = Catch::Session().run(argc, argv);
-
-  finalize_distribution();
-
-  return result;
-}
-
-template<typename P>
-void test_adapt(prog_opts const &opts, std::filesystem::path gold_base)
-{
-  auto const prefix = gold_base.filename().string();
-  gold_base.remove_filename();
-  auto const fval_orig_path    = gold_base / (prefix + "orig.dat");
-  auto const fval_refine_path  = gold_base / (prefix + "refine.dat");
-  auto const fval_coarse_path  = gold_base / (prefix + "coarse.dat");
-  auto const table_refine_path = gold_base / (prefix + "refine_table.dat");
-  auto const table_coarse_path = gold_base / (prefix + "coarse_table.dat");
-
-  auto const fval_orig   = read_vector_from_txt_file<P>(fval_orig_path);
-  auto const gold_coarse = read_vector_from_txt_file<P>(fval_coarse_path);
-  auto const gold_refine = [fval_refine_path]() {
-    auto gold = read_vector_from_txt_file<P>(fval_refine_path);
-    for (auto i = 0; i < gold.size(); ++i)
-    {
-      // matlab stores new refined coefficients as 1e-15 (0 deletes from sparse
-      // vect)
-      if (std::abs(gold(i)) < 1e-14)
-      {
-        gold(i) = 0.0;
-      }
-    }
-    return gold;
-  }();
-
-  auto const test_tables = [](elements::table const &test,
-                              fk::matrix<int> const &gold) {
-    REQUIRE(test.size() == gold.nrows());
-    for (int64_t i = 0; i < test.size(); ++i)
-    {
-      auto const &test_coords = test.get_coords(i);
-      fk::vector<int> const gold_coords =
-          gold.extract_submatrix(i, 0, 1, gold.ncols());
-      REQUIRE(test_coords == gold_coords);
-    }
-  };
-
-  auto const gold_coarse_table =
-      read_matrix_from_txt_file<int>(table_coarse_path);
-  auto const gold_refine_table =
-      fk::matrix<int>(read_matrix_from_txt_file<int>(table_refine_path));
-
-  auto const pde = make_PDE<P>(opts);
-  // using second opts after reset from the pde
-  auto const &opts2 = pde->options();
-
-  adapt::distributed_grid<P> refine_grid(*pde);
-  auto const my_subgrid   = refine_grid.get_subgrid(get_rank());
-  auto const segment_size = element_segment_size(*pde);
-  auto const my_fval_orig =
-      fval_orig.extract(my_subgrid.col_start * segment_size,
[…]
