+++ src/asgard_fast_math_tests.cpp
@@ -1,1057 +1,0 @@
-#ifdef ASGARD_USE_SCALAPACK
-#include "asgard_cblacs_grid.hpp"
-#endif
-
-#include "tests_general.hpp"
-
-using namespace asgard;
-
-int main(int argc, char *argv[])
-{
-  initialize_distribution();
-
-  int result = Catch::Session().run(argc, argv);
-
-  finalize_distribution();
-
-  return result;
-}
-
-TEMPLATE_TEST_CASE("floating point norms", "[fast_math]", test_precs)
-{
-  /* vector 2-norm tests */
-  fk::vector<TestType, mem_type::owner, resource::host> v = {
-      2,  3,  4, 5, 6, 6, 5, 4, 3, 2, 10, 10, 10,
-      10, 10, 7, 3, 5, 9, 9, 4, 5, 4, 5,  7};
-
-  TestType const gold_l2_norm = 4.0 * std::sqrt(66.0);
-
-  SECTION("fk::vector: owner, host")
-  {
-    TestType norm = fm::nrm2(v);
-
-    relaxed_fp_comparison(gold_l2_norm, norm);
-  }
-
-  SECTION("fk::vector: const view, host")
-  {
-    fk::vector<TestType, mem_type::const_view, resource::host> const
-        v_view_host(v);
-
-    TestType norm = fm::nrm2(v_view_host);
-
-    relaxed_fp_comparison(gold_l2_norm, norm);
-  }
-
-  SECTION("fk::vector: view, host")
-  {
-    fk::vector<TestType, mem_type::view, resource::host> const v_view_host(v);
-
-    TestType norm = fm::nrm2(v_view_host);
-
-    relaxed_fp_comparison(gold_l2_norm, norm);
-  }
-
-  TestType const gold_inf_norm = v[10];
-
-  SECTION("fk::vector: owner, host")
-  {
-    TestType norm = fm::nrminf(v);
-
-    relaxed_fp_comparison(gold_inf_norm, norm);
-  }
-
-  SECTION("fk::vector: const view, host")
-  {
-    fk::vector<TestType, mem_type::const_view, resource::host> const
-        v_view_host(v);
-
-    TestType norm = fm::nrminf(v_view_host);
-
-    relaxed_fp_comparison(gold_inf_norm, norm);
-  }
-
-  SECTION("fk::vector: view, host")
-  {
-    fk::vector<TestType, mem_type::view, resource::host> const v_view_host(v);
-
-    TestType norm = fm::nrminf(v_view_host);
-
-    relaxed_fp_comparison(gold_inf_norm, norm);
-  }
-
-  /* matrix Frobenius norm tests */
-  fk::matrix<TestType, mem_type::owner, resource::host> m = {
-      {2, 3, 4, 5, 6},
-      {6, 5, 4, 3, 2},
-      {10, 10, 10, 10, 10},
-      {7, 3, 5, 9, 9},
-      {4, 5, 4, 5, 7}};
-
-  TestType gold_norm_submatrix = 3.0 * std::sqrt(26.0);
-
-  SECTION("fk::matrix: owner, host")
-  {
-    TestType norm = fm::frobenius(m);
-
-    relaxed_fp_comparison(gold_l2_norm, norm);
-  }
-
-  SECTION("fk::matrix: view, host")
-  {
-    fk::matrix<TestType, mem_type::view, resource::host> m_view_host(m);
-
-    TestType norm = fm::frobenius(m_view_host);
-
-    relaxed_fp_comparison(gold_l2_norm, norm);
-  }
-
-  SECTION("fk::matrix submatrix: view, host")
-  {
-    fk::matrix<TestType, mem_type::view, resource::host> const m_view(m, 2, 3,
[…]
