### 2024-08-16 13dd0f3
added simple quickplot capability with documentation
