+++ python/asgard.py
@@ -61,15 +61,17 @@ class pde_snapshot:
             self.num_dimensions = fdata['ndims'][()]
             self.degree         = fdata['degree'][()]
 
-            self.state = fdata['soln'][()]
+            self.state = fdata['state'][()]
             self.cells = fdata['elements'][()]
             self.time  = fdata['time'][()]
 
             self.num_cells = int(len(self.cells) / (2 * self.num_dimensions))
 
+            self.dimension_names = [None for i in range(self.num_dimensions)]
             self.dimension_min = np.zeros((self.num_dimensions,))
             self.dimension_max = np.zeros((self.num_dimensions,))
             for i in range(self.num_dimensions):
+                self.dimension_names[i] = fdata['dim{0}_name'.format(i)][()].decode("utf-8")
                 self.dimension_min[i] = fdata['dim{0}_min'.format(i)][()]
                 self.dimension_max[i] = fdata['dim{0}_max'.format(i)][()]
 
@@ -79,7 +81,7 @@ class pde_snapshot:
 
         if self.verbose:
             if self.state.dtype == np.float64:
-                print('using double precisoin data (double) - 64-bit')
+                print('using double precision data (double) - 64-bit')
             else:
                 print('using single precision data (float) - 32-bit')
             print('dimensions:  {0}'.format(self.num_dimensions))
@@ -211,6 +213,16 @@ class pde_snapshot:
         return self.evaluate(pgrid).reshape(XX.shape), XX, YY
 
     def evaluate(self, points):
+        '''
+        points must be a 2d numpy array with shape (num_x, num_dimensions)
+        where the num_x is the number of points for the evaluate command
+
+        each row must describe a point inside the domain, but no checking will
+        be performed
+
+        returns a vector with the size num_x holding the corresponding values
+        of the field
+        '''
         assert len(points.shape) == 2, 'points must be a 2d numpy array with num-columns equal to self.num_dimensions'
         assert points.shape[1] == self.num_dimensions, 'points.shape[1] (%d) must be equal to self.num_dimensions (%d)' % (points.shape[1], self.num_dimensions)
 
@@ -222,6 +234,13 @@ class pde_snapshot:
         return presult
 
     def cell_centers(self):
+        '''
+        returns the geometric centers of the sparse grid cells
+
+        do not attempt to plot at these points as they are set right on the
+        the discontinuities of the basis functions and the evaluate()
+        algorithm may be unstable due to rounding error
+        '''
         presult = np.empty((self.num_cells, self.num_dimensions), np.float64)
 
         libasgard.asgard_reconstruct_cell_centers(self.recsol,
@@ -242,16 +261,27 @@ class pde_snapshot:
 if __name__ == "__main__":
     if len(sys.argv) < 2:
         libasgard.asgard_print_version_help()
+    elif sys.argv[1] in ("-s", "-stat", "-stats", "-summary"):
+        if len(sys.argv) < 3:
+            print("stats summary option requires a filename")
+        else:
+            shot = pde_snapshot(sys.argv[2])
+            print("\n", shot, "\n")
     elif not _matplotlib_found_:
         print("could not 'import matplotlib'")
+        print("to print only stats summary use")
+        print("  python3 -m asgard -s %s" % sys.argv[2])
         libasgard.asgard_print_version_help()
     else:
         shot = pde_snapshot(sys.argv[1])
-        print(shot)
+        print("\n", shot, "\n")
+
+        asgplot.title(shot.title, fontsize = 'large')
 
         if shot.num_dimensions == 1:
[…]
