+++ src/asgard_block_matrix_tests.cpp
@@ -1,0 +1,13 @@
+#include "tests_general.hpp"
+
+#include "asgard_small_mats.hpp"
+
+using namespace asgard;
+
+TEST_CASE("block matrix tests", "[block tests]")
+{
+  SECTION("scal")
+  {
+    REQUIRE(true);
+  }
+}
