+++ src/asgard_kronmult_matrix_tests.cpp
@@ -384,221 +384,10 @@ TEMPLATE_TEST_CASE("testing kronmult gpu 6d", "[gpu_dense 6d]", test_precs)
   test_kronmult_dense<TestType, asgard::resource::host>(6, n, 2, 1);
 }
 #endif
+
+#else
+TEMPLATE_TEST_CASE("nothing to test", "[cpu_sparse 1d]", test_precs)
+{
+  REQUIRE(true);
+}
 #endif // KRON_MODE_GLOBAL
-
-#ifdef KRON_MODE_GLOBAL
-
-TEMPLATE_TEST_CASE("testing simple 1d", "[global kron]", test_precs)
-{
-  std::minstd_rand park_miller(42);
-  std::uniform_real_distribution<TestType> unif(-1.0, 1.0);
-
-  std::vector<int> nindex = {10, 20, 44};
-  std::vector<int> levels = {4, 5, 6};
-
-  for (size_t tcase = 0; tcase < nindex.size(); tcase++)
-  {
-    // very simple test
-    asgard::connect_1d conn(levels[tcase], asgard::connect_1d::hierarchy::volume);
-
-    asgard::vector2d<int> ilist(1, nindex[tcase]);
-    std::iota(ilist[0], ilist[0] + nindex[tcase], 0);
-
-    // 1d, 1 term, random operator
-    std::vector<std::vector<TestType>> vals(1);
-    vals[0] = std::vector<TestType>(conn.num_connections());
-    for (auto &v : vals[0])
-      v = unif(park_miller);
-
-    // random vector
-    std::vector<TestType> x(ilist.total_size());
-    for (auto &v : x)
-      v = unif(park_miller);
-
-    std::vector<TestType> y_ref(x.size(), TestType{0});
-
-    int const num = static_cast<int>(y_ref.size());
-    for (int i = 0; i < num; i++)
-    {
-      for (int j = 0; j < num; j++)
-      {
-        int const op_index = conn.get_offset(i, j);
-        if (op_index > -1) // connected
-          y_ref[i] += x[j] * vals[0][op_index];
-      }
-    }
-
-    asgard::kronmult::permutes perms(1);
-    asgard::dimension_sort dsort(ilist);
-
-    std::vector<TestType> y(x.size(), TestType{0});
-    std::vector<TestType> w1(y.size(), TestType{0});
-    std::vector<TestType> w2(y.size(), TestType{0});
-    asgard::kronmult::global_cpu(perms, ilist, dsort, conn, {0}, vals,
-                                 TestType{1}, x.data(), y.data(),
-                                 w1.data(), w1.data());
-
-    test_almost_equal(y, y_ref);
-  }
-}
-
-int int_log2(int x)
-{
-  int l = 0;
-  while (x > 0)
-  {
-    x /= 2;
-    l += 1;
-  }
-  return l;
-}
-
-template<typename precision>
-void test_global_kron(int num_dimensions, int level)
-{
-  int constexpr max_num_dimensions = asgard::max_num_dimensions;
-
-  std::minstd_rand park_miller(42);
-  std::uniform_real_distribution<precision> unif(-1.0, 1.0);
-
-  auto indexes = asgard::permutations::generate_lower_index_set(
-      num_dimensions,
-      [&](std::array<int, max_num_dimensions> const &index) -> bool {
-        int L = 0;
-        for (int i = 0; i < num_dimensions; i++)
-          L += int_log2(index[i]);
-        return (L <= level);
-      });
-
[…]

[…]
