### 2024-08-20 d7a8dfe
fixes for two-stream and rieman problems
