### 2024-08-22 #6766 @danielskatz
@quantumsteve - This all looks good with two small issues:
1. The second sentence in the paper isn't a sentence.
2. It would be nice if the first reference said arXiv in it as the way it was published.

Can you fix these? Then we can move to publishing it.
