+++ paper/paper.bib
@@ -124,6 +124,7 @@ abstract = { This paper reviews the adaptive sparse grid discontinuous Galerkin 
       eprint={1710.09356},
       archivePrefix={arXiv},
       primaryClass={cs.NA},
+      journal={arXiv},
       doi = {10.48550/arXiv.1710.09356},
       URL = {https://doi.org/10.48550/arXiv.1710.09356}
 }
