+++ paper/paper.bib
@@ -117,13 +117,13 @@ author = {Huang, Juntao and Guo, Wei and Cheng, Yingda},
 abstract = { This paper reviews the adaptive sparse grid discontinuous Galerkin (aSG-DG) method for computing high dimensional partial differential equations (PDEs) and its software implementation. The C++ software package called AdaM-DG, implementing the aSG-DG method, is available on GitHub at https://github.com/JuntaoHuang/adaptive-multiresolution-DG. The package is capable of treating a large class of high dimensional linear and nonlinear PDEs. We review the essential components of the algorithm and the functionality of the software, including the multiwavelets used, assembling of bilinear operators, fast matrix-vector product for data with hierarchical structures. We further demonstrate the performance of the package by reporting the numerical error and the CPU cost for several benchmark tests, including linear transport equations, wave equations, and Hamilton-Jacobi (HJ) equations.}
 }
 
-@misc{atanasov2017,
+@article{atanasov2017,
       title={Sparse Grid Discretizations based on a Discontinuous {Galerkin} Method},
       author={Alexander B. Atanasov and Erik Schnetter},
       year={2017},
-      eprint={1710.09356},
-      archivePrefix={arXiv},
-      primaryClass={cs.NA},
+#      eprint={1710.09356},
+#      archivePrefix={arXiv},
+#      primaryClass={cs.NA},
       journal={arXiv},
       doi = {10.48550/arXiv.1710.09356},
       URL = {https://doi.org/10.48550/arXiv.1710.09356}
