+++ paper/paper.bib
@@ -121,9 +121,6 @@ abstract = { This paper reviews the adaptive sparse grid discontinuous Galerkin 
       title={Sparse Grid Discretizations based on a Discontinuous {Galerkin} Method},
       author={Alexander B. Atanasov and Erik Schnetter},
       year={2017},
-#      eprint={1710.09356},
-#      archivePrefix={arXiv},
-#      primaryClass={cs.NA},
       journal={arXiv},
       doi = {10.48550/arXiv.1710.09356},
       URL = {https://doi.org/10.48550/arXiv.1710.09356}
