### 2024-08-22 #6766 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
