### 2024-08-22 #6766 @danielskatz
Congratulations to @quantumsteve (Steven Hahn) and co-authors on your publication!!

JOSS depends on volunteers and wouldn't be successful without you
